refactor: split bulkModals.ts into per-modal files (#124)
BulkConfirmModal, BulkProgressModal, and BulkSummaryModal shared no
state and no types — only the word "Bulk" in their names — and each
changes for a different reason (confirmation copy, progress UX,
summary layout). Co-locating them invited collisions whenever any one
of them grew features.

- src/bulkConfirmModal.ts — BulkConfirmModal + ConfirmModalInfo +
  LARGE_BATCH_THRESHOLD (used only by this modal).
- src/bulkProgressModal.ts — BulkProgressModal.
- src/bulkSummaryModal.ts — BulkSummaryModal.
- bulkOrchestrator imports updated; no behavior change.

Closes #124
