+++ main.js
@@ -1,12 +1,12 @@
-var{defineProperty:e0,getOwnPropertyNames:o1,getOwnPropertyDescriptor:r1}=Object,s1=Object.prototype.hasOwnProperty;function e1(w){return this[w]}var w$=(w)=>{var N=(uN??=new WeakMap).get(w),$;if(N)return N;if(N=e0({},"__esModule",{value:!0}),w&&typeof w==="object"||typeof w==="function"){for(var V of o1(w))if(!s1.call(N,V))e0(N,V,{get:e1.bind(w,V),enumerable:!($=r1(w,V))||$.enumerable})}return uN.set(w,N),N},uN;var N$=(w)=>w;function $$(w,N){this[w]=N$.bind(null,N)}var V$=(w,N)=>{for(var $ in N)e0(w,$,{get:N[$],enumerable:!0,configurable:!0,set:$$.bind(N,$)})};var L4={};V$(L4,{migrateSettings:()=>a1,default:()=>cN});module.exports=w$(L4);var s0=require("obsidian");var o0=require("obsidian");function x(w,N,$,V,K){if(V==="m")throw TypeError("Private method is not writable");if(V==="a"&&!K)throw TypeError("Private accessor was defined without a setter");if(typeof N==="function"?w!==N||!K:!N.has(w))throw TypeError("Cannot write private member to an object whose class did not declare it");return V==="a"?K.call(w,$):K?K.value=$:N.set(w,$),$}function z(w,N,$,V){if($==="a"&&!V)throw TypeError("Private accessor was defined without a getter");if(typeof N==="function"?w!==N||!V:!N.has(w))throw TypeError("Cannot read private member from an object whose class did not declare it");return $==="m"?V:$==="a"?V.call(w):V?V.value:N.get(w)}var wN=function(){let{crypto:w}=globalThis;if(w?.randomUUID)return wN=w.randomUUID.bind(w),w.randomUUID();let N=new Uint8Array(1),$=w?()=>w.getRandomValues(N)[0]:()=>Math.random()*255&255;return"10000000-1000-4000-8000-100000000000".replace(/[018]/g,(V)=>(+V^$()&15>>+V/4).toString(16))};function l(w){return typeof w==="object"&&w!==null&&(("name"in w)&&w.name==="AbortError"||("message"in w)&&String(w.message).includes("FetchRequestCanceledException"))}var Ew=(w)=>{if(w instanceof Error)return w;if(typeof w==="object"&&w!==null){try{if(Object.prototype.toString.call(w)==="[object Error]"){let N=Error(w.message,w.cause?{cause:w.cause}:{});if(w.stack)N.stack=w.stack;if(w.cause&&!N.cause)N.cause=w.cause;if(w.name)N.name=w.name;return N}}catch{}try{return Error(JSON.stringify(w))}catch{}}return Error(w)};class Q extends Error{}class L extends Q{constructor(w,N,$,V,K){super(`${L.makeMessage(w,N,$)}`);this.status=w,this.headers=V,this.requestID=V?.get("request-id"),this.error=N,this.type=K??null}static makeMessage(w,N,$){let V=N?.message?typeof N.message==="string"?N.message:JSON.stringify(N.message):N?JSON.stringify(N):$;if(w&&V)return`${w} ${V}`;if(w)return`${w} status code (no body)`;if(V)return V;return"(no status code or body)"}static generate(w,N,$,V){if(!w||!V)return new Ww({message:$,cause:Ew(N)});let K=N,W=K?.error?.type;if(w===400)return new Dw(w,K,$,V,W);if(w===401)return new gw(w,K,$,V,W);if(w===403)return new Bw(w,K,$,V,W);if(w===404)return new Iw(w,K,$,V,W);if(w===409)return new jw(w,K,$,V,W);if(w===422)return new yw(w,K,$,V,W);if(w===429)return new kw(w,K,$,V,W);if(w>=500)return new hw(w,K,$,V,W);return new L(w,K,$,V,W)}}class B extends L{constructor({message:w}={}){super(void 0,void 0,w||"Request was aborted.",void 0)}}class Ww extends L{constructor({message:w,cause:N}){super(void 0,void 0,w||"Connection error.",void 0);if(N)this.cause=N}}class Pw extends Ww{constructor({message:w}={}){super({message:w??"Request timed out."})}}class Dw extends L{}class gw extends L{}class Bw extends L{}class Iw extends L{}class jw extends L{}class yw extends L{}class kw extends L{}class hw extends L{}var W$=/^[a-z][a-z0-9+.-]*:/i,iN=(w)=>{return W$.test(w)},NN=(w)=>(NN=Array.isArray,NN(w)),$N=NN;function S0(w){if(typeof w!=="object")return{};return w??{}}function VN(w){if(!w)return!0;for(let N in w)return!1;return!0}function pN(w,N){return Object.prototype.hasOwnProperty.call(w,N)}var lN=(w,N)=>{if(typeof N!=="number"||!Number.isInteger(N))throw new Q(`${w} must be an integer`);if(N<0)throw new Q(`${w} must be a positive integer`);return N};var f0=(w)=>{try{return JSON.parse(w)}catch(N){return}};var nN=(w)=>new Promise((N)=>setTimeout(N,w));var s="0.89.0";var rN=()=>{return typeof window<"u"&&typeof window.document<"u"&&typeof navigator<"u"};function X$(){if(typeof Deno<"u"&&Deno.build!=null)return"deno";if(typeof EdgeRuntime<"u")return"edge";if(Object.prototype.toString.call(typeof globalThis.process<"u"?globalThis.process:0)==="[object process]")return"node";return"unknown"}var z$=()=>{let w=X$();if(w==="deno")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":aN(Deno.build.os),"X-Stainless-Arch":tN(Deno.build.arch),"X-Stainless-Runtime":"deno","X-Stainless-Runtime-Version":typeof Deno.version==="string"?Deno.version:Deno.version?.deno??"unknown"};if(typeof EdgeRuntime<"u")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":`other:${EdgeRuntime}`,"X-Stainless-Runtime":"edge","X-Stainless-Runtime-Version":globalThis.process.version};if(w==="node")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":aN(globalThis.process.platform??"unknown"),"X-Stainless-Arch":tN(globalThis.process.arch??"unknown"),"X-Stainless-Runtime":"node","X-Stainless-Runtime-Version":globalThis.process.version??"unknown"};let N=Z$();if(N)return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":`browser:${N.browser}`,"X-Stainless-Runtime-Version":N.version};return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":"unknown","X-Stainless-Runtime-Version":"unknown"}};function Z$(){if(typeof navigator>"u"||!navigator)return null;let w=[{key:"edge",pattern:/Edge(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/MSIE(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/Trident(?:.*rv\:(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"chrome",pattern:/Chrome(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"firefox",pattern:/Firefox(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"safari",pattern:/(?:Version\W+(\d+)\.(\d+)(?:\.(\d+))?)?(?:\W+Mobile\S*)?\W+Safari/}];for(let{key:N,pattern:$}of w){let V=$.exec(navigator.userAgent);if(V){let K=V[1]||0,W=V[2]||0,X=V[3]||0;return{browser:N,version:`${K}.${W}.${X}`}}}return null}var tN=(w)=>{if(w==="x32")return"x32";if(w==="x86_64"||w==="x64")return"x64";if(w==="arm")return"arm";if(w==="aarch64"||w==="arm64")return"arm64";if(w)return`other:${w}`;return"unknown"},aN=(w)=>{if(w=w.toLowerCase(),w.includes("ios"))return"iOS";if(w==="android")return"Android";if(w==="darwin")return"MacOS";if(w==="win32")return"Windows";if(w==="freebsd")return"FreeBSD";if(w==="openbsd")return"OpenBSD";if(w==="linux")return"Linux";if(w)return`Other:${w}`;return"Unknown"},oN,sN=()=>{return oN??(oN=z$())};function eN(){if(typeof fetch<"u")return fetch;throw Error("`fetch` is not defined as a global; Either pass `fetch` to the client, `new Anthropic({ fetch })` or polyfill the global, `globalThis.fetch = fetch`")}function KN(...w){let N=globalThis.ReadableStream;if(typeof N>"u")throw Error("`ReadableStream` is not defined as a global; You will need to polyfill it, `globalThis.ReadableStream = ReadableStream`");return new N(...w)}function A0(w){let N=Symbol.asyncIterator in w?w[Symbol.asyncIterator]():w[Symbol.iterator]();return KN({start(){},async pull($){let{done:V,value:K}=await N.next();if(V)$.close();else $.enqueue(K)},async cancel(){await N.return?.()}})}function Rw(w){if(w[Symbol.asyncIterator])return w;let N=w.getReader();return{async next(){try{let $=await N.read();if($?.done)N.releaseLock();return $}catch($){throw N.releaseLock(),$}},async return(){let $=N.cancel();return N.releaseLock(),await $,{done:!0,value:void 0}},[Symbol.asyncIterator](){return this}}}async function w1(w){if(w===null||typeof w!=="object")return;if(w[Symbol.asyncIterator]){await w[Symbol.asyncIterator]().return?.();return}let N=w.getReader(),$=N.cancel();N.releaseLock(),await $}var N1=({headers:w,body:N})=>{return{bodyHeaders:{"content-type":"application/json"},body:JSON.stringify(N)}};function $1(w){return Object.entries(w).filter(([N,$])=>typeof $<"u").map(([N,$])=>{if(typeof $==="string"||typeof $==="number"||typeof $==="boolean")return`${encodeURIComponent(N)}=${encodeURIComponent($)}`;if($===null)return`${encodeURIComponent(N)}=`;throw new Q(`Cannot stringify type ${typeof $}; Expected string, number, boolean, or null. If you need to pass nested query parameters, you can manually encode them, e.g. { query: { 'foo[key1]': value1, 'foo[key2]': value2 } }, and please open a GitHub issue requesting better support for your use case.`)}).join("&")}function W1(w){let N=0;for(let K of w)N+=K.length;let $=new Uint8Array(N),V=0;for(let K of w)$.set(K,V),V+=K.length;return $}var V1;function _w(w){let N;return(V1??(N=new globalThis.TextEncoder,V1=N.encode.bind(N)))(w)}var K1;function WN(w){let N;return(K1??(N=new globalThis.TextDecoder,K1=N.decode.bind(N)))(w)}var R,_;class e{constructor(){R.set(this,void 0),_.set(this,void 0),x(this,R,new Uint8Array,"f"),x(this,_,null,"f")}decode(w){if(w==null)return[];let N=w instanceof ArrayBuffer?new Uint8Array(w):typeof w==="string"?_w(w):w;x(this,R,W1([z(this,R,"f"),N]),"f");let $=[],V;while((V=x$(z(this,R,"f"),z(this,_,"f")))!=null){if(V.carriage&&z(this,_,"f")==null){x(this,_,V.index,"f");continue}if(z(this,_,"f")!=null&&(V.index!==z(this,_,"f")+1||V.carriage)){$.push(WN(z(this,R,"f").subarray(0,z(this,_,"f")-1))),x(this,R,z(this,R,"f").subarray(z(this,_,"f")),"f"),x(this,_,null,"f");continue}let K=z(this,_,"f")!==null?V.preceding-1:V.preceding,W=WN(z(this,R,"f").subarray(0,K));$.push(W),x(this,R,z(this,R,"f").subarray(V.index),"f"),x(this,_,null,"f")}return $}flush(){if(!z(this,R,"f").length)return[];return this.decode(`
+var{defineProperty:s0,getOwnPropertyNames:s$,getOwnPropertyDescriptor:e$}=Object,w1=Object.prototype.hasOwnProperty;function N1(w){return this[w]}var $1=(w)=>{var N=(mN??=new WeakMap).get(w),$;if(N)return N;if(N=s0({},"__esModule",{value:!0}),w&&typeof w==="object"||typeof w==="function"){for(var V of s$(w))if(!w1.call(N,V))s0(N,V,{get:N1.bind(w,V),enumerable:!($=e$(w,V))||$.enumerable})}return mN.set(w,N),N},mN;var V1=(w)=>w;function K1(w,N){this[w]=V1.bind(null,N)}var W1=(w,N)=>{for(var $ in N)s0(w,$,{get:N[$],enumerable:!0,configurable:!0,set:K1.bind(N,$)})};var P2={};W1(P2,{migrateSettings:()=>r$,default:()=>dN});module.exports=$1(P2);var r0=require("obsidian");var a0=require("obsidian");var uN=require("obsidian"),X1=100;class e0 extends uN.Modal{resolver;resolved=!1;info;constructor(w,N){super(w);this.info=N}openAndAwait(){return new Promise((w)=>{this.resolver=w,this.open()})}onOpen(){let{contentEl:w}=this,{folderPath:N,total:$,willChange:V,willSkip:K,settings:W}=this.info;w.createEl("h2",{text:"Generate metadata for folder"}),w.createEl("p",{text:`Folder: ${N}`}),w.createEl("p",{text:`${$} notes — ${V} will change, ${K} will skip`});let X=W.truncateContent?W.truncateMethod:"disabled";if(w.createEl("p",{text:`Model: ${W.anthropicModel} · Update: ${W.updateMethod} · Truncation: ${X}`}),$>X1){let C=w.createEl("p",{text:`⚠ Large batch — this will make up to ${V} API calls`});C.style.color="var(--text-warning)",C.style.fontWeight="bold"}let Z=w.createDiv({cls:"modal-button-container"});Z.createEl("button",{text:"Cancel"}).addEventListener("click",()=>{this.resolve(!1),this.close()}),Z.createEl("button",{text:`Generate (${V})`,cls:"mod-cta"}).addEventListener("click",()=>{this.resolve(!0),this.close()})}onClose(){this.resolve(!1),this.contentEl.empty()}resolve(w){if(this.resolved)return;this.resolved=!0,this.resolver?.(w)}}function x(w,N,$,V,K){if(V==="m")throw TypeError("Private method is not writable");if(V==="a"&&!K)throw TypeError("Private accessor was defined without a setter");if(typeof N==="function"?w!==N||!K:!N.has(w))throw TypeError("Cannot write private member to an object whose class did not declare it");return V==="a"?K.call(w,$):K?K.value=$:N.set(w,$),$}function z(w,N,$,V){if($==="a"&&!V)throw TypeError("Private accessor was defined without a getter");if(typeof N==="function"?w!==N||!V:!N.has(w))throw TypeError("Cannot read private member from an object whose class did not declare it");return $==="m"?V:$==="a"?V.call(w):V?V.value:N.get(w)}var wN=function(){let{crypto:w}=globalThis;if(w?.randomUUID)return wN=w.randomUUID.bind(w),w.randomUUID();let N=new Uint8Array(1),$=w?()=>w.getRandomValues(N)[0]:()=>Math.random()*255&255;return"10000000-1000-4000-8000-100000000000".replace(/[018]/g,(V)=>(+V^$()&15>>+V/4).toString(16))};function l(w){return typeof w==="object"&&w!==null&&(("name"in w)&&w.name==="AbortError"||("message"in w)&&String(w.message).includes("FetchRequestCanceledException"))}var Hw=(w)=>{if(w instanceof Error)return w;if(typeof w==="object"&&w!==null){try{if(Object.prototype.toString.call(w)==="[object Error]"){let N=Error(w.message,w.cause?{cause:w.cause}:{});if(w.stack)N.stack=w.stack;if(w.cause&&!N.cause)N.cause=w.cause;if(w.name)N.name=w.name;return N}}catch{}try{return Error(JSON.stringify(w))}catch{}}return Error(w)};class Q extends Error{}class L extends Q{constructor(w,N,$,V,K){super(`${L.makeMessage(w,N,$)}`);this.status=w,this.headers=V,this.requestID=V?.get("request-id"),this.error=N,this.type=K??null}static makeMessage(w,N,$){let V=N?.message?typeof N.message==="string"?N.message:JSON.stringify(N.message):N?JSON.stringify(N):$;if(w&&V)return`${w} ${V}`;if(w)return`${w} status code (no body)`;if(V)return V;return"(no status code or body)"}static generate(w,N,$,V){if(!w||!V)return new Ww({message:$,cause:Hw(N)});let K=N,W=K?.error?.type;if(w===400)return new Dw(w,K,$,V,W);if(w===401)return new gw(w,K,$,V,W);if(w===403)return new Bw(w,K,$,V,W);if(w===404)return new Iw(w,K,$,V,W);if(w===409)return new yw(w,K,$,V,W);if(w===422)return new jw(w,K,$,V,W);if(w===429)return new kw(w,K,$,V,W);if(w>=500)return new hw(w,K,$,V,W);return new L(w,K,$,V,W)}}class B extends L{constructor({message:w}={}){super(void 0,void 0,w||"Request was aborted.",void 0)}}class Ww extends L{constructor({message:w,cause:N}){super(void 0,void 0,w||"Connection error.",void 0);if(N)this.cause=N}}class Pw extends Ww{constructor({message:w}={}){super({message:w??"Request timed out."})}}class Dw extends L{}class gw extends L{}class Bw extends L{}class Iw extends L{}class yw extends L{}class jw extends L{}class kw extends L{}class hw extends L{}var Z1=/^[a-z][a-z0-9+.-]*:/i,iN=(w)=>{return Z1.test(w)},NN=(w)=>(NN=Array.isArray,NN(w)),$N=NN;function F0(w){if(typeof w!=="object")return{};return w??{}}function VN(w){if(!w)return!0;for(let N in w)return!1;return!0}function pN(w,N){return Object.prototype.hasOwnProperty.call(w,N)}var lN=(w,N)=>{if(typeof N!=="number"||!Number.isInteger(N))throw new Q(`${w} must be an integer`);if(N<0)throw new Q(`${w} must be a positive integer`);return N};var S0=(w)=>{try{return JSON.parse(w)}catch(N){return}};var nN=(w)=>new Promise((N)=>setTimeout(N,w));var s="0.89.0";var rN=()=>{return typeof window<"u"&&typeof window.document<"u"&&typeof navigator<"u"};function J1(){if(typeof Deno<"u"&&Deno.build!=null)return"deno";if(typeof EdgeRuntime<"u")return"edge";if(Object.prototype.toString.call(typeof globalThis.process<"u"?globalThis.process:0)==="[object process]")return"node";return"unknown"}var Y1=()=>{let w=J1();if(w==="deno")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":aN(Deno.build.os),"X-Stainless-Arch":tN(Deno.build.arch),"X-Stainless-Runtime":"deno","X-Stainless-Runtime-Version":typeof Deno.version==="string"?Deno.version:Deno.version?.deno??"unknown"};if(typeof EdgeRuntime<"u")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":`other:${EdgeRuntime}`,"X-Stainless-Runtime":"edge","X-Stainless-Runtime-Version":globalThis.process.version};if(w==="node")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":aN(globalThis.process.platform??"unknown"),"X-Stainless-Arch":tN(globalThis.process.arch??"unknown"),"X-Stainless-Runtime":"node","X-Stainless-Runtime-Version":globalThis.process.version??"unknown"};let N=x1();if(N)return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":`browser:${N.browser}`,"X-Stainless-Runtime-Version":N.version};return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":"unknown","X-Stainless-Runtime-Version":"unknown"}};function x1(){if(typeof navigator>"u"||!navigator)return null;let w=[{key:"edge",pattern:/Edge(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/MSIE(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/Trident(?:.*rv\:(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"chrome",pattern:/Chrome(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"firefox",pattern:/Firefox(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"safari",pattern:/(?:Version\W+(\d+)\.(\d+)(?:\.(\d+))?)?(?:\W+Mobile\S*)?\W+Safari/}];for(let{key:N,pattern:$}of w){let V=$.exec(navigator.userAgent);if(V){let K=V[1]||0,W=V[2]||0,X=V[3]||0;return{browser:N,version:`${K}.${W}.${X}`}}}return null}var tN=(w)=>{if(w==="x32")return"x32";if(w==="x86_64"||w==="x64")return"x64";if(w==="arm")return"arm";if(w==="aarch64"||w==="arm64")return"arm64";if(w)return`other:${w}`;return"unknown"},aN=(w)=>{if(w=w.toLowerCase(),w.includes("ios"))return"iOS";if(w==="android")return"Android";if(w==="darwin")return"MacOS";if(w==="win32")return"Windows";if(w==="freebsd")return"FreeBSD";if(w==="openbsd")return"OpenBSD";if(w==="linux")return"Linux";if(w)return`Other:${w}`;return"Unknown"},oN,sN=()=>{return oN??(oN=Y1())};function eN(){if(typeof fetch<"u")return fetch;throw Error("`fetch` is not defined as a global; Either pass `fetch` to the client, `new Anthropic({ fetch })` or polyfill the global, `globalThis.fetch = fetch`")}function KN(...w){let N=globalThis.ReadableStream;if(typeof N>"u")throw Error("`ReadableStream` is not defined as a global; You will need to polyfill it, `globalThis.ReadableStream = ReadableStream`");return new N(...w)}function A0(w){let N=Symbol.asyncIterator in w?w[Symbol.asyncIterator]():w[Symbol.iterator]();return KN({start(){},async pull($){let{done:V,value:K}=await N.next();if(V)$.close();else $.enqueue(K)},async cancel(){await N.return?.()}})}function Rw(w){if(w[Symbol.asyncIterator])return w;let N=w.getReader();return{async next(){try{let $=await N.read();if($?.done)N.releaseLock();return $}catch($){throw N.releaseLock(),$}},async return(){let $=N.cancel();return N.releaseLock(),await $,{done:!0,value:void 0}},[Symbol.asyncIterator](){return this}}}async function w$(w){if(w===null||typeof w!=="object")return;if(w[Symbol.asyncIterator]){await w[Symbol.asyncIterator]().return?.();return}let N=w.getReader(),$=N.cancel();N.releaseLock(),await $}var N$=({headers:w,body:N})=>{return{bodyHeaders:{"content-type":"application/json"},body:JSON.stringify(N)}};function $$(w){return Object.entries(w).filter(([N,$])=>typeof $<"u").map(([N,$])=>{if(typeof $==="string"||typeof $==="number"||typeof $==="boolean")return`${encodeURIComponent(N)}=${encodeURIComponent($)}`;if($===null)return`${encodeURIComponent(N)}=`;throw new Q(`Cannot stringify type ${typeof $}; Expected string, number, boolean, or null. If you need to pass nested query parameters, you can manually encode them, e.g. { query: { 'foo[key1]': value1, 'foo[key2]': value2 } }, and please open a GitHub issue requesting better support for your use case.`)}).join("&")}function W$(w){let N=0;for(let K of w)N+=K.length;let $=new Uint8Array(N),V=0;for(let K of w)$.set(K,V),V+=K.length;return $}var V$;function _w(w){let N;return(V$??(N=new globalThis.TextEncoder,V$=N.encode.bind(N)))(w)}var K$;function WN(w){let N;return(K$??(N=new globalThis.TextDecoder,K$=N.decode.bind(N)))(w)}var R,_;class e{constructor(){R.set(this,void 0),_.set(this,void 0),x(this,R,new Uint8Array,"f"),x(this,_,null,"f")}decode(w){if(w==null)return[];let N=w instanceof ArrayBuffer?new Uint8Array(w):typeof w==="string"?_w(w):w;x(this,R,W$([z(this,R,"f"),N]),"f");let $=[],V;while((V=G1(z(this,R,"f"),z(this,_,"f")))!=null){if(V.carriage&&z(this,_,"f")==null){x(this,_,V.index,"f");continue}if(z(this,_,"f")!=null&&(V.index!==z(this,_,"f")+1||V.carriage)){$.push(WN(z(this,R,"f").subarray(0,z(this,_,"f")-1))),x(this,R,z(this,R,"f").subarray(z(this,_,"f")),"f"),x(this,_,null,"f");continue}let K=z(this,_,"f")!==null?V.preceding-1:V.preceding,W=WN(z(this,R,"f").subarray(0,K));$.push(W),x(this,R,z(this,R,"f").subarray(V.index),"f"),x(this,_,null,"f")}return $}flush(){if(!z(this,R,"f").length)return[];return this.decode(`
 `)}}R=new WeakMap,_=new WeakMap;e.NEWLINE_CHARS=new Set([`
-`,"\r"]);e.NEWLINE_REGEXP=/\r\n|[\n\r]/g;function x$(w,N){for(let K=N??0;K<w.length;K++){if(w[K]===10)return{preceding:K,index:K+1,carriage:!1};if(w[K]===13)return{preceding:K,index:K+1,carriage:!0}}return null}function X1(w){for(let V=0;V<w.length-1;V++){if(w[V]===10&&w[V+1]===10)return V+2;if(w[V]===13&&w[V+1]===13)return V+2;if(w[V]===13&&w[V+1]===10&&V+3<w.length&&w[V+2]===13&&w[V+3]===10)return V+4}return-1}var H0={off:0,error:200,warn:300,info:400,debug:500},XN=(w,N,$)=>{if(!w)return;if(pN(H0,w))return w;D($).warn(`${N} was set to ${JSON.stringify(w)}, expected one of ${JSON.stringify(Object.keys(H0))}`);return};function bw(){}function q0(w,N,$){if(!N||H0[w]>H0[$])return bw;else return N[w].bind(N)}var M$={error:bw,warn:bw,info:bw,debug:bw},z1=new WeakMap;function D(w){let N=w.logger,$=w.logLevel??"off";if(!N)return M$;let V=z1.get(N);if(V&&V[0]===$)return V[1];let K={error:q0("error",N,$),warn:q0("warn",N,$),info:q0("info",N,$),debug:q0("debug",N,$)};return z1.set(N,[$,K]),K}var n=(w)=>{if(w.options)w.options={...w.options},delete w.options.headers;if(w.headers)w.headers=Object.fromEntries((w.headers instanceof Headers?[...w.headers]:Object.entries(w.headers)).map(([N,$])=>[N,N.toLowerCase()==="x-api-key"||N.toLowerCase()==="authorization"||N.toLowerCase()==="cookie"||N.toLowerCase()==="set-cookie"?"***":$]));if("retryOfRequestLogID"in w){if(w.retryOfRequestLogID)w.retryOf=w.retryOfRequestLogID;delete w.retryOfRequestLogID}return w};var dw;class b{constructor(w,N,$){this.iterator=w,dw.set(this,void 0),this.controller=N,x(this,dw,$,"f")}static fromSSEResponse(w,N,$){let V=!1,K=$?D($):console;async function*W(){if(V)throw new Q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");V=!0;let X=!1;try{for await(let Z of Q$(w,N)){if(Z.event==="completion")try{yield JSON.parse(Z.data)}catch(Y){throw K.error("Could not parse message into JSON:",Z.data),K.error("From chunk:",Z.raw),Y}if(Z.event==="message_start"||Z.event==="message_delta"||Z.event==="message_stop"||Z.event==="content_block_start"||Z.event==="content_block_delta"||Z.event==="content_block_stop"||Z.event==="message"||Z.event==="user.message"||Z.event==="user.interrupt"||Z.event==="user.tool_confirmation"||Z.event==="user.custom_tool_result"||Z.event==="agent.message"||Z.event==="agent.thinking"||Z.event==="agent.tool_use"||Z.event==="agent.tool_result"||Z.event==="agent.mcp_tool_use"||Z.event==="agent.mcp_tool_result"||Z.event==="agent.custom_tool_use"||Z.event==="agent.thread_context_compacted"||Z.event==="session.status_running"||Z.event==="session.status_idle"||Z.event==="session.status_rescheduled"||Z.event==="session.status_terminated"||Z.event==="session.error"||Z.event==="session.deleted"||Z.event==="span.model_request_start"||Z.event==="span.model_request_end")try{yield JSON.parse(Z.data)}catch(Y){throw K.error("Could not parse message into JSON:",Z.data),K.error("From chunk:",Z.raw),Y}if(Z.event==="ping")continue;if(Z.event==="error"){let Y=f0(Z.data)??Z.data,M=Y?.error?.type;throw new L(void 0,Y,void 0,w.headers,M)}}X=!0}catch(Z){if(l(Z))return;throw Z}finally{if(!X)N.abort()}}return new b(W,N,$)}static fromReadableStream(w,N,$){let V=!1;async function*K(){let X=new e,Z=Rw(w);for await(let Y of Z)for(let M of X.decode(Y))yield M;for(let Y of X.flush())yield Y}async function*W(){if(V)throw new Q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");V=!0;let X=!1;try{for await(let Z of K()){if(X)continue;if(Z)yield JSON.parse(Z)}X=!0}catch(Z){if(l(Z))return;throw Z}finally{if(!X)N.abort()}}return new b(W,N,$)}[(dw=new WeakMap,Symbol.asyncIterator)](){return this.iterator()}tee(){let w=[],N=[],$=this.iterator(),V=(K)=>{return{next:()=>{if(K.length===0){let W=$.next();w.push(W),N.push(W)}return K.shift()}}};return[new b(()=>V(w),this.controller,z(this,dw,"f")),new b(()=>V(N),this.controller,z(this,dw,"f"))]}toReadableStream(){let w=this,N;return KN({async start(){N=w[Symbol.asyncIterator]()},async pull($){try{let{value:V,done:K}=await N.next();if(K)return $.close();let W=_w(JSON.stringify(V)+`
-`);$.enqueue(W)}catch(V){$.error(V)}},async cancel(){await N.return?.()}})}}async function*Q$(w,N){if(!w.body){if(N.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new Q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new Q("Attempted to iterate over a response with no body")}let $=new Z1,V=new e,K=Rw(w.body);for await(let W of G$(K))for(let X of V.decode(W)){let Z=$.decode(X);if(Z)yield Z}for(let W of V.flush()){let X=$.decode(W);if(X)yield X}}async function*G$(w){let N=new Uint8Array;for await(let $ of w){if($==null)continue;let V=$ instanceof ArrayBuffer?new Uint8Array($):typeof $==="string"?_w($):$,K=new Uint8Array(N.length+V.length);K.set(N),K.set(V,N.length),N=K;let W;while((W=X1(N))!==-1)yield N.slice(0,W),N=N.slice(W)}if(N.length>0)yield N}class Z1{constructor(){this.event=null,this.data=[],this.chunks=[]}decode(w){if(w.endsWith("\r"))w=w.substring(0,w.length-1);if(!w){if(!this.event&&!this.data.length)return null;let K={event:this.event,data:this.data.join(`
-`),raw:this.chunks};return this.event=null,this.data=[],this.chunks=[],K}if(this.chunks.push(w),w.startsWith(":"))return null;let[N,$,V]=T$(w,":");if(V.startsWith(" "))V=V.substring(1);if(N==="event")this.event=V;else if(N==="data")this.data.push(V);return null}}function T$(w,N){let $=w.indexOf(N);if($!==-1)return[w.substring(0,$),N,w.substring($+N.length)];return[w,"",""]}async function L0(w,N){let{response:$,requestLogID:V,retryOfRequestLogID:K,startTime:W}=N,X=await(async()=>{if(N.options.stream){if(D(w).debug("response",$.status,$.url,$.headers,$.body),N.options.__streamClass)return N.options.__streamClass.fromSSEResponse($,N.controller);return b.fromSSEResponse($,N.controller)}if($.status===204)return null;if(N.options.__binaryResponse)return $;let Y=$.headers.get("content-type")?.split(";")[0]?.trim();if(Y?.includes("application/json")||Y?.endsWith("+json")){if($.headers.get("content-length")==="0")return;let T=await $.json();return zN(T,$)}return await $.text()})();return D(w).debug(`[${V}] response parsed`,n({retryOfRequestLogID:K,url:$.url,status:$.status,body:X,durationMs:Date.now()-W})),X}function zN(w,N){if(!w||typeof w!=="object"||Array.isArray(w))return w;return Object.defineProperty(w,"_request_id",{value:N.headers.get("request-id"),enumerable:!1})}var cw;class Xw extends Promise{constructor(w,N,$=L0){super((V)=>{V(null)});this.responsePromise=N,this.parseResponse=$,cw.set(this,void 0),x(this,cw,w,"f")}_thenUnwrap(w){return new Xw(z(this,cw,"f"),this.responsePromise,async(N,$)=>zN(w(await this.parseResponse(N,$),$),$.response))}asResponse(){return this.responsePromise.then((w)=>w.response)}async withResponse(){let[w,N]=await Promise.all([this.parse(),this.asResponse()]);return{data:w,response:N,request_id:N.headers.get("request-id")}}parse(){if(!this.parsedPromise)this.parsedPromise=this.responsePromise.then((w)=>this.parseResponse(z(this,cw,"f"),w));return this.parsedPromise}then(w,N){return this.parse().then(w,N)}catch(w){return this.parse().catch(w)}finally(w){return this.parse().finally(w)}}cw=new WeakMap;var E0;class ZN{constructor(w,N,$,V){E0.set(this,void 0),x(this,E0,w,"f"),this.options=V,this.response=N,this.body=$}hasNextPage(){if(!this.getPaginatedItems().length)return!1;return this.nextPageRequestOptions()!=null}async getNextPage(){let w=this.nextPageRequestOptions();if(!w)throw new Q("No next page expected; please check `.hasNextPage()` before calling `.getNextPage()`.");return await z(this,E0,"f").requestAPIList(this.constructor,w)}async*iterPages(){let w=this;yield w;while(w.hasNextPage())w=await w.getNextPage(),yield w}async*[(E0=new WeakMap,Symbol.asyncIterator)](){for await(let w of this.iterPages())for(let N of w.getPaginatedItems())yield N}}class P0 extends Xw{constructor(w,N,$){super(w,N,async(V,K)=>new $(V,K.response,await L0(V,K),K.options))}async*[Symbol.asyncIterator](){let w=await this;for await(let N of w)yield N}}class i extends ZN{constructor(w,N,$,V){super(w,N,$,V);this.data=$.data||[],this.has_more=$.has_more||!1,this.first_id=$.first_id||null,this.last_id=$.last_id||null}getPaginatedItems(){return this.data??[]}hasNextPage(){if(this.has_more===!1)return!1;return super.hasNextPage()}nextPageRequestOptions(){if(this.options.query?.before_id){let N=this.first_id;if(!N)return null;return{...this.options,query:{...S0(this.options.query),before_id:N}}}let w=this.last_id;if(!w)return null;return{...this.options,query:{...S0(this.options.query),after_id:w}}}}class q extends ZN{constructor(w,N,$,V){super(w,N,$,V);this.data=$.data||[],this.next_page=$.next_page||null}getPaginatedItems(){return this.data??[]}nextPageRequestOptions(){let w=this.next_page;if(!w)return null;return{...this.options,query:{...S0(this.options.query),page:w}}}}var YN=()=>{if(typeof File>"u"){let{process:w}=globalThis,N=typeof w?.versions?.node==="string"&&parseInt(w.versions.node.split("."))<20;throw Error("`File` is not defined as a global, which is required for file uploads."+(N?" Update to Node 20 LTS or newer, or set `globalThis.File` to `import('node:buffer').File`.":""))}};function zw(w,N,$){return YN(),new File(w,N??"unknown_file",$)}function vw(w,N){let $=typeof w==="object"&&w!==null&&(("name"in w)&&w.name&&String(w.name)||("url"in w)&&w.url&&String(w.url)||("filename"in w)&&w.filename&&String(w.filename)||("path"in w)&&w.path&&String(w.path))||"";return N?$.split(/[\\/]/).pop()||void 0:$}var xN=(w)=>w!=null&&typeof w==="object"&&typeof w[Symbol.asyncIterator]==="function";var Qw=async(w,N,$=!0)=>{return{...w,body:await U$(w.body,N,$)}},J1=new WeakMap;function O$(w){let N=typeof w==="function"?w:w.fetch,$=J1.get(N);if($)return $;let V=(async()=>{try{let K="Response"in N?N.Response:(await N("data:,")).constructor,W=new FormData;if(W.toString()===await new K(W).text())return!1;return!0}catch{return!0}})();return J1.set(N,V),V}var U$=async(w,N,$=!0)=>{if(!await O$(N))throw TypeError("The provided fetch function does not support file uploads with the current global FormData class.");let V=new FormData;return await Promise.all(Object.entries(w||{}).map(([K,W])=>JN(V,K,W,$))),V},F$=(w)=>w instanceof Blob&&("name"in w);var JN=async(w,N,$,V)=>{if($===void 0)return;if($==null)throw TypeError(`Received null for "${N}"; to pass null in FormData, you must use the string 'null'`);if(typeof $==="string"||typeof $==="number"||typeof $==="boolean")w.append(N,String($));else if($ instanceof Response){let K={},W=$.headers.get("Content-Type");if(W)K={type:W};w.append(N,zw([await $.blob()],vw($,V),K))}else if(xN($))w.append(N,zw([await new Response(A0($)).blob()],vw($,V)));else if(F$($))w.append(N,zw([$],vw($,V),{type:$.type}));else if(Array.isArray($))await Promise.all($.map((K)=>JN(w,N+"[]",K,V)));else if(typeof $==="object")await Promise.all(Object.entries($).map(([K,W])=>JN(w,`${N}[${K}]`,W,V)));else throw TypeError(`Invalid value given to form, expected a string, number, boolean, object, Array, File or Blob but got ${$} instead`)};var Y1=(w)=>w!=null&&typeof w==="object"&&typeof w.size==="number"&&typeof w.type==="string"&&typeof w.text==="function"&&typeof w.slice==="function"&&typeof w.arrayBuffer==="function",S$=(w)=>w!=null&&typeof w==="object"&&typeof w.name==="string"&&typeof w.lastModified==="number"&&Y1(w),f$=(w)=>w!=null&&typeof w==="object"&&typeof w.url==="string"&&typeof w.blob==="function";async function D0(w,N,$){if(YN(),w=await w,N||(N=vw(w,!0)),S$(w)){if(w instanceof File&&N==null&&$==null)return w;return zw([await w.arrayBuffer()],N??w.name,{type:w.type,lastModified:w.lastModified,...$})}if(f$(w)){let K=await w.blob();return N||(N=new URL(w.url).pathname.split(/[\\/]/).pop()),zw(await MN(K),N,$)}let V=await MN(w);if(!$?.type){let K=V.find((W)=>typeof W==="object"&&("type"in W)&&W.type);if(typeof K==="string")$={...$,type:K}}return zw(V,N,$)}async function MN(w){let N=[];if(typeof w==="string"||ArrayBuffer.isView(w)||w instanceof ArrayBuffer)N.push(w);else if(Y1(w))N.push(w instanceof Blob?w:await w.arrayBuffer());else if(xN(w))for await(let $ of w)N.push(...await MN($));else{let $=w?.constructor?.name;throw Error(`Unexpected data type: ${typeof w}${$?`; constructor: ${$}`:""}${A$(w)}`)}return N}function A$(w){if(typeof w!=="object"||w===null)return"";return`; props: [${Object.getOwnPropertyNames(w).map(($)=>`"${$}"`).join(", ")}]`}class O{constructor(w){this._client=w}}var x1=Symbol.for("brand.privateNullableHeaders");function*H$(w){if(!w)return;if(x1 in w){let{values:V,nulls:K}=w;yield*V.entries();for(let W of K)yield[W,null];return}let N=!1,$;if(w instanceof Headers)$=w.entries();else if($N(w))$=w;else N=!0,$=Object.entries(w??{});for(let V of $){let K=V[0];if(typeof K!=="string")throw TypeError("expected header name to be a string");let W=$N(V[1])?V[1]:[V[1]],X=!1;for(let Z of W){if(Z===void 0)continue;if(N&&!X)X=!0,yield[K,null];yield[K,Z]}}}var J=(w)=>{let N=new Headers,$=new Set;for(let V of w){let K=new Set;for(let[W,X]of H$(V)){let Z=W.toLowerCase();if(!K.has(Z))N.delete(W),K.add(Z);if(X===null)N.delete(W),$.add(Z);else N.append(W,X),$.delete(Z)}}return{[x1]:!0,values:N,nulls:$}};function Q1(w){return w.replace(/[^A-Za-z0-9\-._~!$&'()*+,;=:@]+/g,encodeURIComponent)}var M1=Object.freeze(Object.create(null)),L$=(w=Q1)=>function($,...V){if($.length===1)return $[0];let K=!1,W=[],X=$.reduce((C,S,T)=>{if(/[?#]/.test(S))K=!0;let U=V[T],h=(K?encodeURIComponent:w)(""+U);if(T!==V.length&&(U==null||typeof U==="object"&&U.toString===Object.getPrototypeOf(Object.getPrototypeOf(U.hasOwnProperty??M1)??M1)?.toString))h=U+"",W.push({start:C.length+S.length,length:h.length,error:`Value of type ${Object.prototype.toString.call(U).slice(8,-1)} is not a valid path parameter`});return C+S+(T===V.length?"":h)},""),Z=X.split(/[?#]/,1)[0],Y=/(?<=^|\/)(?:\.|%2e){1,2}(?=\/|$)/gi,M;while((M=Y.exec(Z))!==null)W.push({start:M.index,length:M[0].length,error:`Value "${M[0]}" can't be safely passed as a path parameter`});if(W.sort((C,S)=>C.start-S.start),W.length>0){let C=0,S=W.reduce((T,U)=>{let h=" ".repeat(U.start-C),I="^".repeat(U.length);return C=U.start+U.length,T+h+I},"");throw new Q(`Path parameters result in path with invalid segments:
+`,"\r"]);e.NEWLINE_REGEXP=/\r\n|[\n\r]/g;function G1(w,N){for(let K=N??0;K<w.length;K++){if(w[K]===10)return{preceding:K,index:K+1,carriage:!1};if(w[K]===13)return{preceding:K,index:K+1,carriage:!0}}return null}function X$(w){for(let V=0;V<w.length-1;V++){if(w[V]===10&&w[V+1]===10)return V+2;if(w[V]===13&&w[V+1]===13)return V+2;if(w[V]===13&&w[V+1]===10&&V+3<w.length&&w[V+2]===13&&w[V+3]===10)return V+4}return-1}var E0={off:0,error:200,warn:300,info:400,debug:500},XN=(w,N,$)=>{if(!w)return;if(pN(E0,w))return w;D($).warn(`${N} was set to ${JSON.stringify(w)}, expected one of ${JSON.stringify(Object.keys(E0))}`);return};function bw(){}function q0(w,N,$){if(!N||E0[w]>E0[$])return bw;else return N[w].bind(N)}var T1={error:bw,warn:bw,info:bw,debug:bw},z$=new WeakMap;function D(w){let N=w.logger,$=w.logLevel??"off";if(!N)return T1;let V=z$.get(N);if(V&&V[0]===$)return V[1];let K={error:q0("error",N,$),warn:q0("warn",N,$),info:q0("info",N,$),debug:q0("debug",N,$)};return z$.set(N,[$,K]),K}var n=(w)=>{if(w.options)w.options={...w.options},delete w.options.headers;if(w.headers)w.headers=Object.fromEntries((w.headers instanceof Headers?[...w.headers]:Object.entries(w.headers)).map(([N,$])=>[N,N.toLowerCase()==="x-api-key"||N.toLowerCase()==="authorization"||N.toLowerCase()==="cookie"||N.toLowerCase()==="set-cookie"?"***":$]));if("retryOfRequestLogID"in w){if(w.retryOfRequestLogID)w.retryOf=w.retryOfRequestLogID;delete w.retryOfRequestLogID}return w};var dw;class b{constructor(w,N,$){this.iterator=w,dw.set(this,void 0),this.controller=N,x(this,dw,$,"f")}static fromSSEResponse(w,N,$){let V=!1,K=$?D($):console;async function*W(){if(V)throw new Q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");V=!0;let X=!1;try{for await(let Z of C1(w,N)){if(Z.event==="completion")try{yield JSON.parse(Z.data)}catch(Y){throw K.error("Could not parse message into JSON:",Z.data),K.error("From chunk:",Z.raw),Y}if(Z.event==="message_start"||Z.event==="message_delta"||Z.event==="message_stop"||Z.event==="content_block_start"||Z.event==="content_block_delta"||Z.event==="content_block_stop"||Z.event==="message"||Z.event==="user.message"||Z.event==="user.interrupt"||Z.event==="user.tool_confirmation"||Z.event==="user.custom_tool_result"||Z.event==="agent.message"||Z.event==="agent.thinking"||Z.event==="agent.tool_use"||Z.event==="agent.tool_result"||Z.event==="agent.mcp_tool_use"||Z.event==="agent.mcp_tool_result"||Z.event==="agent.custom_tool_use"||Z.event==="agent.thread_context_compacted"||Z.event==="session.status_running"||Z.event==="session.status_idle"||Z.event==="session.status_rescheduled"||Z.event==="session.status_terminated"||Z.event==="session.error"||Z.event==="session.deleted"||Z.event==="span.model_request_start"||Z.event==="span.model_request_end")try{yield JSON.parse(Z.data)}catch(Y){throw K.error("Could not parse message into JSON:",Z.data),K.error("From chunk:",Z.raw),Y}if(Z.event==="ping")continue;if(Z.event==="error"){let Y=S0(Z.data)??Z.data,M=Y?.error?.type;throw new L(void 0,Y,void 0,w.headers,M)}}X=!0}catch(Z){if(l(Z))return;throw Z}finally{if(!X)N.abort()}}return new b(W,N,$)}static fromReadableStream(w,N,$){let V=!1;async function*K(){let X=new e,Z=Rw(w);for await(let Y of Z)for(let M of X.decode(Y))yield M;for(let Y of X.flush())yield Y}async function*W(){if(V)throw new Q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");V=!0;let X=!1;try{for await(let Z of K()){if(X)continue;if(Z)yield JSON.parse(Z)}X=!0}catch(Z){if(l(Z))return;throw Z}finally{if(!X)N.abort()}}return new b(W,N,$)}[(dw=new WeakMap,Symbol.asyncIterator)](){return this.iterator()}tee(){let w=[],N=[],$=this.iterator(),V=(K)=>{return{next:()=>{if(K.length===0){let W=$.next();w.push(W),N.push(W)}return K.shift()}}};return[new b(()=>V(w),this.controller,z(this,dw,"f")),new b(()=>V(N),this.controller,z(this,dw,"f"))]}toReadableStream(){let w=this,N;return KN({async start(){N=w[Symbol.asyncIterator]()},async pull($){try{let{value:V,done:K}=await N.next();if(K)return $.close();let W=_w(JSON.stringify(V)+`
+`);$.enqueue(W)}catch(V){$.error(V)}},async cancel(){await N.return?.()}})}}async function*C1(w,N){if(!w.body){if(N.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new Q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new Q("Attempted to iterate over a response with no body")}let $=new Z$,V=new e,K=Rw(w.body);for await(let W of O1(K))for(let X of V.decode(W)){let Z=$.decode(X);if(Z)yield Z}for(let W of V.flush()){let X=$.decode(W);if(X)yield X}}async function*O1(w){let N=new Uint8Array;for await(let $ of w){if($==null)continue;let V=$ instanceof ArrayBuffer?new Uint8Array($):typeof $==="string"?_w($):$,K=new Uint8Array(N.length+V.length);K.set(N),K.set(V,N.length),N=K;let W;while((W=X$(N))!==-1)yield N.slice(0,W),N=N.slice(W)}if(N.length>0)yield N}class Z${constructor(){this.event=null,this.data=[],this.chunks=[]}decode(w){if(w.endsWith("\r"))w=w.substring(0,w.length-1);if(!w){if(!this.event&&!this.data.length)return null;let K={event:this.event,data:this.data.join(`
+`),raw:this.chunks};return this.event=null,this.data=[],this.chunks=[],K}if(this.chunks.push(w),w.startsWith(":"))return null;let[N,$,V]=U1(w,":");if(V.startsWith(" "))V=V.substring(1);if(N==="event")this.event=V;else if(N==="data")this.data.push(V);return null}}function U1(w,N){let $=w.indexOf(N);if($!==-1)return[w.substring(0,$),N,w.substring($+N.length)];return[w,"",""]}async function L0(w,N){let{response:$,requestLogID:V,retryOfRequestLogID:K,startTime:W}=N,X=await(async()=>{if(N.options.stream){if(D(w).debug("response",$.status,$.url,$.headers,$.body),N.options.__streamClass)return N.options.__streamClass.fromSSEResponse($,N.controller);return b.fromSSEResponse($,N.controller)}if($.status===204)return null;if(N.options.__binaryResponse)return $;let Y=$.headers.get("content-type")?.split(";")[0]?.trim();if(Y?.includes("application/json")||Y?.endsWith("+json")){if($.headers.get("content-length")==="0")return;let T=await $.json();return zN(T,$)}return await $.text()})();return D(w).debug(`[${V}] response parsed`,n({retryOfRequestLogID:K,url:$.url,status:$.status,body:X,durationMs:Date.now()-W})),X}function zN(w,N){if(!w||typeof w!=="object"||Array.isArray(w))return w;return Object.defineProperty(w,"_request_id",{value:N.headers.get("request-id"),enumerable:!1})}var cw;class Xw extends Promise{constructor(w,N,$=L0){super((V)=>{V(null)});this.responsePromise=N,this.parseResponse=$,cw.set(this,void 0),x(this,cw,w,"f")}_thenUnwrap(w){return new Xw(z(this,cw,"f"),this.responsePromise,async(N,$)=>zN(w(await this.parseResponse(N,$),$),$.response))}asResponse(){return this.responsePromise.then((w)=>w.response)}async withResponse(){let[w,N]=await Promise.all([this.parse(),this.asResponse()]);return{data:w,response:N,request_id:N.headers.get("request-id")}}parse(){if(!this.parsedPromise)this.parsedPromise=this.responsePromise.then((w)=>this.parseResponse(z(this,cw,"f"),w));return this.parsedPromise}then(w,N){return this.parse().then(w,N)}catch(w){return this.parse().catch(w)}finally(w){return this.parse().finally(w)}}cw=new WeakMap;var H0;class ZN{constructor(w,N,$,V){H0.set(this,void 0),x(this,H0,w,"f"),this.options=V,this.response=N,this.body=$}hasNextPage(){if(!this.getPaginatedItems().length)return!1;return this.nextPageRequestOptions()!=null}async getNextPage(){let w=this.nextPageRequestOptions();if(!w)throw new Q("No next page expected; please check `.hasNextPage()` before calling `.getNextPage()`.");return await z(this,H0,"f").requestAPIList(this.constructor,w)}async*iterPages(){let w=this;yield w;while(w.hasNextPage())w=await w.getNextPage(),yield w}async*[(H0=new WeakMap,Symbol.asyncIterator)](){for await(let w of this.iterPages())for(let N of w.getPaginatedItems())yield N}}class P0 extends Xw{constructor(w,N,$){super(w,N,async(V,K)=>new $(V,K.response,await L0(V,K),K.options))}async*[Symbol.asyncIterator](){let w=await this;for await(let N of w)yield N}}class i extends ZN{constructor(w,N,$,V){super(w,N,$,V);this.data=$.data||[],this.has_more=$.has_more||!1,this.first_id=$.first_id||null,this.last_id=$.last_id||null}getPaginatedItems(){return this.data??[]}hasNextPage(){if(this.has_more===!1)return!1;return super.hasNextPage()}nextPageRequestOptions(){if(this.options.query?.before_id){let N=this.first_id;if(!N)return null;return{...this.options,query:{...F0(this.options.query),before_id:N}}}let w=this.last_id;if(!w)return null;return{...this.options,query:{...F0(this.options.query),after_id:w}}}}class q extends ZN{constructor(w,N,$,V){super(w,N,$,V);this.data=$.data||[],this.next_page=$.next_page||null}getPaginatedItems(){return this.data??[]}nextPageRequestOptions(){let w=this.next_page;if(!w)return null;return{...this.options,query:{...F0(this.options.query),page:w}}}}var YN=()=>{if(typeof File>"u"){let{process:w}=globalThis,N=typeof w?.versions?.node==="string"&&parseInt(w.versions.node.split("."))<20;throw Error("`File` is not defined as a global, which is required for file uploads."+(N?" Update to Node 20 LTS or newer, or set `globalThis.File` to `import('node:buffer').File`.":""))}};function zw(w,N,$){return YN(),new File(w,N??"unknown_file",$)}function vw(w,N){let $=typeof w==="object"&&w!==null&&(("name"in w)&&w.name&&String(w.name)||("url"in w)&&w.url&&String(w.url)||("filename"in w)&&w.filename&&String(w.filename)||("path"in w)&&w.path&&String(w.path))||"";return N?$.split(/[\\/]/).pop()||void 0:$}var xN=(w)=>w!=null&&typeof w==="object"&&typeof w[Symbol.asyncIterator]==="function";var Qw=async(w,N,$=!0)=>{return{...w,body:await S1(w.body,N,$)}},J$=new WeakMap;function F1(w){let N=typeof w==="function"?w:w.fetch,$=J$.get(N);if($)return $;let V=(async()=>{try{let K="Response"in N?N.Response:(await N("data:,")).constructor,W=new FormData;if(W.toString()===await new K(W).text())return!1;return!0}catch{return!0}})();return J$.set(N,V),V}var S1=async(w,N,$=!0)=>{if(!await F1(N))throw TypeError("The provided fetch function does not support file uploads with the current global FormData class.");let V=new FormData;return await Promise.all(Object.entries(w||{}).map(([K,W])=>JN(V,K,W,$))),V},A1=(w)=>w instanceof Blob&&("name"in w);var JN=async(w,N,$,V)=>{if($===void 0)return;if($==null)throw TypeError(`Received null for "${N}"; to pass null in FormData, you must use the string 'null'`);if(typeof $==="string"||typeof $==="number"||typeof $==="boolean")w.append(N,String($));else if($ instanceof Response){let K={},W=$.headers.get("Content-Type");if(W)K={type:W};w.append(N,zw([await $.blob()],vw($,V),K))}else if(xN($))w.append(N,zw([await new Response(A0($)).blob()],vw($,V)));else if(A1($))w.append(N,zw([$],vw($,V),{type:$.type}));else if(Array.isArray($))await Promise.all($.map((K)=>JN(w,N+"[]",K,V)));else if(typeof $==="object")await Promise.all(Object.entries($).map(([K,W])=>JN(w,`${N}[${K}]`,W,V)));else throw TypeError(`Invalid value given to form, expected a string, number, boolean, object, Array, File or Blob but got ${$} instead`)};var Y$=(w)=>w!=null&&typeof w==="object"&&typeof w.size==="number"&&typeof w.type==="string"&&typeof w.text==="function"&&typeof w.slice==="function"&&typeof w.arrayBuffer==="function",q1=(w)=>w!=null&&typeof w==="object"&&typeof w.name==="string"&&typeof w.lastModified==="number"&&Y$(w),E1=(w)=>w!=null&&typeof w==="object"&&typeof w.url==="string"&&typeof w.blob==="function";async function D0(w,N,$){if(YN(),w=await w,N||(N=vw(w,!0)),q1(w)){if(w instanceof File&&N==null&&$==null)return w;return zw([await w.arrayBuffer()],N??w.name,{type:w.type,lastModified:w.lastModified,...$})}if(E1(w)){let K=await w.blob();return N||(N=new URL(w.url).pathname.split(/[\\/]/).pop()),zw(await MN(K),N,$)}let V=await MN(w);if(!$?.type){let K=V.find((W)=>typeof W==="object"&&("type"in W)&&W.type);if(typeof K==="string")$={...$,type:K}}return zw(V,N,$)}async function MN(w){let N=[];if(typeof w==="string"||ArrayBuffer.isView(w)||w instanceof ArrayBuffer)N.push(w);else if(Y$(w))N.push(w instanceof Blob?w:await w.arrayBuffer());else if(xN(w))for await(let $ of w)N.push(...await MN($));else{let $=w?.constructor?.name;throw Error(`Unexpected data type: ${typeof w}${$?`; constructor: ${$}`:""}${L1(w)}`)}return N}function L1(w){if(typeof w!=="object"||w===null)return"";return`; props: [${Object.getOwnPropertyNames(w).map(($)=>`"${$}"`).join(", ")}]`}class O{constructor(w){this._client=w}}var x$=Symbol.for("brand.privateNullableHeaders");function*P1(w){if(!w)return;if(x$ in w){let{values:V,nulls:K}=w;yield*V.entries();for(let W of K)yield[W,null];return}let N=!1,$;if(w instanceof Headers)$=w.entries();else if($N(w))$=w;else N=!0,$=Object.entries(w??{});for(let V of $){let K=V[0];if(typeof K!=="string")throw TypeError("expected header name to be a string");let W=$N(V[1])?V[1]:[V[1]],X=!1;for(let Z of W){if(Z===void 0)continue;if(N&&!X)X=!0,yield[K,null];yield[K,Z]}}}var J=(w)=>{let N=new Headers,$=new Set;for(let V of w){let K=new Set;for(let[W,X]of P1(V)){let Z=W.toLowerCase();if(!K.has(Z))N.delete(W),K.add(Z);if(X===null)N.delete(W),$.add(Z);else N.append(W,X),$.delete(Z)}}return{[x$]:!0,values:N,nulls:$}};function Q$(w){return w.replace(/[^A-Za-z0-9\-._~!$&'()*+,;=:@]+/g,encodeURIComponent)}var M$=Object.freeze(Object.create(null)),D1=(w=Q$)=>function($,...V){if($.length===1)return $[0];let K=!1,W=[],X=$.reduce((C,F,T)=>{if(/[?#]/.test(F))K=!0;let U=V[T],h=(K?encodeURIComponent:w)(""+U);if(T!==V.length&&(U==null||typeof U==="object"&&U.toString===Object.getPrototypeOf(Object.getPrototypeOf(U.hasOwnProperty??M$)??M$)?.toString))h=U+"",W.push({start:C.length+F.length,length:h.length,error:`Value of type ${Object.prototype.toString.call(U).slice(8,-1)} is not a valid path parameter`});return C+F+(T===V.length?"":h)},""),Z=X.split(/[?#]/,1)[0],Y=/(?<=^|\/)(?:\.|%2e){1,2}(?=\/|$)/gi,M;while((M=Y.exec(Z))!==null)W.push({start:M.index,length:M[0].length,error:`Value "${M[0]}" can't be safely passed as a path parameter`});if(W.sort((C,F)=>C.start-F.start),W.length>0){let C=0,F=W.reduce((T,U)=>{let h=" ".repeat(U.start-C),I="^".repeat(U.length);return C=U.start+U.length,T+h+I},"");throw new Q(`Path parameters result in path with invalid segments:
 ${W.map((T)=>T.error).join(`
 `)}
 ${X}
-${S}`)}return X},G=L$(Q1);class mw extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/environments?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/environments/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/environments/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/environments?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/environments/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/environments/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}var uw=Symbol("anthropic.sdk.stainlessHelper");function g0(w){return typeof w==="object"&&w!==null&&uw in w}function QN(w,N){let $=new Set;if(w){for(let V of w)if(g0(V))$.add(V[uw])}if(N)for(let V of N){if(g0(V))$.add(V[uw]);if(Array.isArray(V.content)){for(let K of V.content)if(g0(K))$.add(K[uw])}}return Array.from($)}function B0(w,N){let $=QN(w,N);if($.length===0)return{};return{"x-stainless-helper":$.join(", ")}}function G1(w){if(g0(w))return{"x-stainless-helper":w[uw]};return{}}class iw extends O{list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/files",i,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"files-api-2025-04-14"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/files/${w}`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString()},$?.headers])})}download(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/files/${w}/content`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString(),Accept:"application/binary"},$?.headers]),__binaryResponse:!0})}retrieveMetadata(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/files/${w}`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString()},$?.headers])})}upload(w,N){let{betas:$,...V}=w;return this._client.post("/v1/files",Qw({body:V,...N,headers:J([{"anthropic-beta":[...$??[],"files-api-2025-04-14"].toString()},G1(V.file),N?.headers])},this._client))}}class pw extends O{retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/models/${w}?beta=true`,{...$,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/models?beta=true",i,{query:V,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers])})}}class lw extends O{list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/agents/${w}/versions?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class Gw extends O{constructor(){super(...arguments);this.versions=new lw(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/agents?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V,...K}=N??{};return this._client.get(G`/v1/agents/${w}?beta=true`,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/agents/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/agents?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/agents/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}Gw.Versions=lw;var I0={"claude-opus-4-20250514":8192,"claude-opus-4-0":8192,"claude-4-opus-20250514":8192,"anthropic.claude-opus-4-20250514-v1:0":8192,"claude-opus-4@20250514":8192,"claude-opus-4-1-20250805":8192,"anthropic.claude-opus-4-1-20250805-v1:0":8192,"claude-opus-4-1@20250805":8192};function T1(w){return w?.output_format??w?.output_config?.format}function GN(w,N,$){let V=T1(N);if(!N||!("parse"in(V??{})))return{...w,content:w.content.map((K)=>{if(K.type==="text"){let W=Object.defineProperty({...K},"parsed_output",{value:null,enumerable:!1});return Object.defineProperty(W,"parsed",{get(){return $.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),null},enumerable:!1})}return K}),parsed_output:null};return TN(w,N,$)}function TN(w,N,$){let V=null,K=w.content.map((W)=>{if(W.type==="text"){let X=I$(N,W.text);if(V===null)V=X;let Z=Object.defineProperty({...W},"parsed_output",{value:X,enumerable:!1});return Object.defineProperty(Z,"parsed",{get(){return $.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),X},enumerable:!1})}return W});return{...w,content:K,parsed_output:V}}function I$(w,N){let $=T1(w);if($?.type!=="json_schema")return null;try{if("parse"in $)return $.parse(N);return JSON.parse(N)}catch(V){throw new Q(`Failed to parse structured output: ${V}`)}}var j$=(w)=>{let N=0,$=[];while(N<w.length){let V=w[N];if(V==="\\"){N++;continue}if(V==="{"){$.push({type:"brace",value:"{"}),N++;continue}if(V==="}"){$.push({type:"brace",value:"}"}),N++;continue}if(V==="["){$.push({type:"paren",value:"["}),N++;continue}if(V==="]"){$.push({type:"paren",value:"]"}),N++;continue}if(V===":"){$.push({type:"separator",value:":"}),N++;continue}if(V===","){$.push({type:"delimiter",value:","}),N++;continue}if(V==='"'){let Z="",Y=!1;V=w[++N];while(V!=='"'){if(N===w.length){Y=!0;break}if(V==="\\"){if(N++,N===w.length){Y=!0;break}Z+=V+w[N],V=w[++N]}else Z+=V,V=w[++N]}if(V=w[++N],!Y)$.push({type:"string",value:Z});continue}if(V&&/\s/.test(V)){N++;continue}let W=/[0-9]/;if(V&&W.test(V)||V==="-"||V==="."){let Z="";if(V==="-")Z+=V,V=w[++N];while(V&&W.test(V)||V===".")Z+=V,V=w[++N];$.push({type:"number",value:Z});continue}let X=/[a-z]/i;if(V&&X.test(V)){let Z="";while(V&&X.test(V)){if(N===w.length)break;Z+=V,V=w[++N]}if(Z=="true"||Z=="false"||Z==="null")$.push({type:"name",value:Z});else{N++;continue}continue}N++}return $},Tw=(w)=>{if(w.length===0)return w;let N=w[w.length-1];switch(N.type){case"separator":return w=w.slice(0,w.length-1),Tw(w);break;case"number":let $=N.value[N.value.length-1];if($==="."||$==="-")return w=w.slice(0,w.length-1),Tw(w);case"string":let V=w[w.length-2];if(V?.type==="delimiter")return w=w.slice(0,w.length-1),Tw(w);else if(V?.type==="brace"&&V.value==="{")return w=w.slice(0,w.length-1),Tw(w);break;case"delimiter":return w=w.slice(0,w.length-1),Tw(w);break}return w},y$=(w)=>{let N=[];if(w.map(($)=>{if($.type==="brace")if($.value==="{")N.push("}");else N.splice(N.lastIndexOf("}"),1);if($.type==="paren")if($.value==="[")N.push("]");else N.splice(N.lastIndexOf("]"),1)}),N.length>0)N.reverse().map(($)=>{if($==="}")w.push({type:"brace",value:"}"});else if($==="]")w.push({type:"paren",value:"]"})});return w},k$=(w)=>{let N="";return w.map(($)=>{switch($.type){case"string":N+='"'+$.value+'"';break;default:N+=$.value;break}}),N},j0=(w)=>JSON.parse(k$(y$(Tw(j$(w)))));var m,ww,Cw,nw,y0,tw,aw,k0,ow,t,rw,h0,R0,Zw,_0,b0,sw,CN,C1,d0,ON,UN,FN,O1,U1="__json_buf";function F1(w){return w.type==="tool_use"||w.type==="server_tool_use"||w.type==="mcp_tool_use"}class ew{constructor(w,N){m.add(this),this.messages=[],this.receivedMessages=[],ww.set(this,void 0),Cw.set(this,null),this.controller=new AbortController,nw.set(this,void 0),y0.set(this,()=>{}),tw.set(this,()=>{}),aw.set(this,void 0),k0.set(this,()=>{}),ow.set(this,()=>{}),t.set(this,{}),rw.set(this,!1),h0.set(this,!1),R0.set(this,!1),Zw.set(this,!1),_0.set(this,void 0),b0.set(this,void 0),sw.set(this,void 0),d0.set(this,($)=>{if(x(this,h0,!0,"f"),l($))$=new B;if($ instanceof B)return x(this,R0,!0,"f"),this._emit("abort",$);if($ instanceof Q)return this._emit("error",$);if($ instanceof Error){let V=new Q($.message);return V.cause=$,this._emit("error",V)}return this._emit("error",new Q(String($)))}),x(this,nw,new Promise(($,V)=>{x(this,y0,$,"f"),x(this,tw,V,"f")}),"f"),x(this,aw,new Promise(($,V)=>{x(this,k0,$,"f"),x(this,ow,V,"f")}),"f"),z(this,nw,"f").catch(()=>{}),z(this,aw,"f").catch(()=>{}),x(this,Cw,w,"f"),x(this,sw,N?.logger??console,"f")}get response(){return z(this,_0,"f")}get request_id(){return z(this,b0,"f")}async withResponse(){x(this,Zw,!0,"f");let w=await z(this,nw,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let N=new ew(null);return N._run(()=>N._fromReadableStream(w)),N}static createMessage(w,N,$,{logger:V}={}){let K=new ew(N,{logger:V});for(let W of N.messages)K._addMessageParam(W);return x(K,Cw,{...N,stream:!0},"f"),K._run(()=>K._createMessage(w,{...N,stream:!0},{...$,headers:{...$?.headers,"X-Stainless-Helper-Method":"stream"}})),K}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},z(this,d0,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,N=!0){if(this.receivedMessages.push(w),N)this._emit("message",w)}async _createMessage(w,N,$){let V=$?.signal,K;if(V){if(V.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),V.addEventListener("abort",K)}try{z(this,m,"m",ON).call(this);let{response:W,data:X}=await w.create({...N,stream:!0},{...$,signal:this.controller.signal}).withResponse();this._connected(W);for await(let Z of X)z(this,m,"m",UN).call(this,Z);if(X.controller.signal?.aborted)throw new B;z(this,m,"m",FN).call(this)}finally{if(V&&K)V.removeEventListener("abort",K)}}_connected(w){if(this.ended)return;x(this,_0,w,"f"),x(this,b0,w?.headers.get("request-id"),"f"),z(this,y0,"f").call(this,w),this._emit("connect")}get ended(){return z(this,rw,"f")}get errored(){return z(this,h0,"f")}get aborted(){return z(this,R0,"f")}abort(){this.controller.abort()}on(w,N){return(z(this,t,"f")[w]||(z(this,t,"f")[w]=[])).push({listener:N}),this}off(w,N){let $=z(this,t,"f")[w];if(!$)return this;let V=$.findIndex((K)=>K.listener===N);if(V>=0)$.splice(V,1);return this}once(w,N){return(z(this,t,"f")[w]||(z(this,t,"f")[w]=[])).push({listener:N,once:!0}),this}emitted(w){return new Promise((N,$)=>{if(x(this,Zw,!0,"f"),w!=="error")this.once("error",$);this.once(w,N)})}async done(){x(this,Zw,!0,"f"),await z(this,aw,"f")}get currentMessage(){return z(this,ww,"f")}async finalMessage(){return await this.done(),z(this,m,"m",CN).call(this)}async finalText(){return await this.done(),z(this,m,"m",C1).call(this)}_emit(w,...N){if(z(this,rw,"f"))return;if(w==="end")x(this,rw,!0,"f"),z(this,k0,"f").call(this);let $=z(this,t,"f")[w];if($)z(this,t,"f")[w]=$.filter((V)=>!V.once),$.forEach(({listener:V})=>V(...N));if(w==="abort"){let V=N[0];if(!z(this,Zw,"f")&&!$?.length)Promise.reject(V);z(this,tw,"f").call(this,V),z(this,ow,"f").call(this,V),this._emit("end");return}if(w==="error"){let V=N[0];if(!z(this,Zw,"f")&&!$?.length)Promise.reject(V);z(this,tw,"f").call(this,V),z(this,ow,"f").call(this,V),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",z(this,m,"m",CN).call(this))}async _fromReadableStream(w,N){let $=N?.signal,V;if($){if($.aborted)this.controller.abort();V=this.controller.abort.bind(this.controller),$.addEventListener("abort",V)}try{z(this,m,"m",ON).call(this),this._connected(null);let K=b.fromReadableStream(w,this.controller);for await(let W of K)z(this,m,"m",UN).call(this,W);if(K.controller.signal?.aborted)throw new B;z(this,m,"m",FN).call(this)}finally{if($&&V)$.removeEventListener("abort",V)}}[(ww=new WeakMap,Cw=new WeakMap,nw=new WeakMap,y0=new WeakMap,tw=new WeakMap,aw=new WeakMap,k0=new WeakMap,ow=new WeakMap,t=new WeakMap,rw=new WeakMap,h0=new WeakMap,R0=new WeakMap,Zw=new WeakMap,_0=new WeakMap,b0=new WeakMap,sw=new WeakMap,d0=new WeakMap,m=new WeakSet,CN=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},C1=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");let N=this.receivedMessages.at(-1).content.filter(($)=>$.type==="text").map(($)=>$.text);if(N.length===0)throw new Q("stream ended without producing a content block with type=text");return N.join(" ")},ON=function(){if(this.ended)return;x(this,ww,void 0,"f")},UN=function(N){if(this.ended)return;let $=z(this,m,"m",O1).call(this,N);switch(this._emit("streamEvent",N,$),N.type){case"content_block_delta":{let V=$.content.at(-1);switch(N.delta.type){case"text_delta":{if(V.type==="text")this._emit("text",N.delta.text,V.text||"");break}case"citations_delta":{if(V.type==="text")this._emit("citation",N.delta.citation,V.citations??[]);break}case"input_json_delta":{if(F1(V)&&V.input)this._emit("inputJson",N.delta.partial_json,V.input);break}case"thinking_delta":{if(V.type==="thinking")this._emit("thinking",N.delta.thinking,V.thinking);break}case"signature_delta":{if(V.type==="thinking")this._emit("signature",V.signature);break}case"compaction_delta":{if(V.type==="compaction"&&V.content)this._emit("compaction",V.content);break}default:S1(N.delta)}break}case"message_stop":{this._addMessageParam($),this._addMessage(GN($,z(this,Cw,"f"),{logger:z(this,sw,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",$.content.at(-1));break}case"message_start":{x(this,ww,$,"f");break}case"content_block_start":case"message_delta":break}},FN=function(){if(this.ended)throw new Q("stream has ended, this shouldn't happen");let N=z(this,ww,"f");if(!N)throw new Q("request ended without sending any chunks");return x(this,ww,void 0,"f"),GN(N,z(this,Cw,"f"),{logger:z(this,sw,"f")})},O1=function(N){let $=z(this,ww,"f");if(N.type==="message_start"){if($)throw new Q(`Unexpected event order, got ${N.type} before receiving "message_stop"`);return N.message}if(!$)throw new Q(`Unexpected event order, got ${N.type} before "message_start"`);switch(N.type){case"message_stop":return $;case"message_delta":if($.container=N.delta.container,$.stop_reason=N.delta.stop_reason,$.stop_sequence=N.delta.stop_sequence,$.usage.output_tokens=N.usage.output_tokens,$.context_management=N.context_management,N.usage.input_tokens!=null)$.usage.input_tokens=N.usage.input_tokens;if(N.usage.cache_creation_input_tokens!=null)$.usage.cache_creation_input_tokens=N.usage.cache_creation_input_tokens;if(N.usage.cache_read_input_tokens!=null)$.usage.cache_read_input_tokens=N.usage.cache_read_input_tokens;if(N.usage.server_tool_use!=null)$.usage.server_tool_use=N.usage.server_tool_use;if(N.usage.iterations!=null)$.usage.iterations=N.usage.iterations;return $;case"content_block_start":return $.content.push(N.content_block),$;case"content_block_delta":{let V=$.content.at(N.index);switch(N.delta.type){case"text_delta":{if(V?.type==="text")$.content[N.index]={...V,text:(V.text||"")+N.delta.text};break}case"citations_delta":{if(V?.type==="text")$.content[N.index]={...V,citations:[...V.citations??[],N.delta.citation]};break}case"input_json_delta":{if(V&&F1(V)){let K=V[U1]||"";K+=N.delta.partial_json;let W={...V};if(Object.defineProperty(W,U1,{value:K,enumerable:!1,writable:!0}),K)try{W.input=j0(K)}catch(X){let Z=new Q(`Unable to parse tool parameter JSON from model. Please retry your request or adjust your prompt. Error: ${X}. JSON: ${K}`);z(this,d0,"f").call(this,Z)}$.content[N.index]=W}break}case"thinking_delta":{if(V?.type==="thinking")$.content[N.index]={...V,thinking:V.thinking+N.delta.thinking};break}case"signature_delta":{if(V?.type==="thinking")$.content[N.index]={...V,signature:N.delta.signature};break}case"compaction_delta":{if(V?.type==="compaction")$.content[N.index]={...V,content:(V.content||"")+N.delta.content};break}default:S1(N.delta)}return $}case"content_block_stop":return $}},Symbol.asyncIterator)](){let w=[],N=[],$=!1;return this.on("streamEvent",(V)=>{let K=N.shift();if(K)K.resolve(V);else w.push(V)}),this.on("end",()=>{$=!0;for(let V of N)V.resolve(void 0);N.length=0}),this.on("abort",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),this.on("error",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),{next:async()=>{if(!w.length){if($)return{value:void 0,done:!0};return new Promise((K,W)=>N.push({resolve:K,reject:W})).then((K)=>K?{value:K,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new b(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function S1(w){}class Ow extends Error{constructor(w){let N=typeof w==="string"?w:w.map(($)=>{if($.type==="text")return $.text;return`[${$.type}]`}).join(" ");super(N);this.name="ToolError",this.content=w}}var f1=1e5,A1=`You have been working on the task described above but have not yet completed it. Write a continuation summary that will allow you (or another instance of yourself) to resume work efficiently in a future context window where the conversation history will be replaced with this summary. Your summary should be structured, concise, and actionable. Include:
+${F}`)}return X},G=D1(Q$);class mw extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/environments?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/environments/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/environments/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/environments?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/environments/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/environments/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}var uw=Symbol("anthropic.sdk.stainlessHelper");function g0(w){return typeof w==="object"&&w!==null&&uw in w}function QN(w,N){let $=new Set;if(w){for(let V of w)if(g0(V))$.add(V[uw])}if(N)for(let V of N){if(g0(V))$.add(V[uw]);if(Array.isArray(V.content)){for(let K of V.content)if(g0(K))$.add(K[uw])}}return Array.from($)}function B0(w,N){let $=QN(w,N);if($.length===0)return{};return{"x-stainless-helper":$.join(", ")}}function G$(w){if(g0(w))return{"x-stainless-helper":w[uw]};return{}}class iw extends O{list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/files",i,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"files-api-2025-04-14"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/files/${w}`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString()},$?.headers])})}download(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/files/${w}/content`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString(),Accept:"application/binary"},$?.headers]),__binaryResponse:!0})}retrieveMetadata(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/files/${w}`,{...$,headers:J([{"anthropic-beta":[...V??[],"files-api-2025-04-14"].toString()},$?.headers])})}upload(w,N){let{betas:$,...V}=w;return this._client.post("/v1/files",Qw({body:V,...N,headers:J([{"anthropic-beta":[...$??[],"files-api-2025-04-14"].toString()},G$(V.file),N?.headers])},this._client))}}class pw extends O{retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/models/${w}?beta=true`,{...$,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/models?beta=true",i,{query:V,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers])})}}class lw extends O{list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/agents/${w}/versions?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class Gw extends O{constructor(){super(...arguments);this.versions=new lw(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/agents?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V,...K}=N??{};return this._client.get(G`/v1/agents/${w}?beta=true`,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/agents/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/agents?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/agents/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}Gw.Versions=lw;var I0={"claude-opus-4-20250514":8192,"claude-opus-4-0":8192,"claude-4-opus-20250514":8192,"anthropic.claude-opus-4-20250514-v1:0":8192,"claude-opus-4@20250514":8192,"claude-opus-4-1-20250805":8192,"anthropic.claude-opus-4-1-20250805-v1:0":8192,"claude-opus-4-1@20250805":8192};function T$(w){return w?.output_format??w?.output_config?.format}function GN(w,N,$){let V=T$(N);if(!N||!("parse"in(V??{})))return{...w,content:w.content.map((K)=>{if(K.type==="text"){let W=Object.defineProperty({...K},"parsed_output",{value:null,enumerable:!1});return Object.defineProperty(W,"parsed",{get(){return $.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),null},enumerable:!1})}return K}),parsed_output:null};return TN(w,N,$)}function TN(w,N,$){let V=null,K=w.content.map((W)=>{if(W.type==="text"){let X=k1(N,W.text);if(V===null)V=X;let Z=Object.defineProperty({...W},"parsed_output",{value:X,enumerable:!1});return Object.defineProperty(Z,"parsed",{get(){return $.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),X},enumerable:!1})}return W});return{...w,content:K,parsed_output:V}}function k1(w,N){let $=T$(w);if($?.type!=="json_schema")return null;try{if("parse"in $)return $.parse(N);return JSON.parse(N)}catch(V){throw new Q(`Failed to parse structured output: ${V}`)}}var h1=(w)=>{let N=0,$=[];while(N<w.length){let V=w[N];if(V==="\\"){N++;continue}if(V==="{"){$.push({type:"brace",value:"{"}),N++;continue}if(V==="}"){$.push({type:"brace",value:"}"}),N++;continue}if(V==="["){$.push({type:"paren",value:"["}),N++;continue}if(V==="]"){$.push({type:"paren",value:"]"}),N++;continue}if(V===":"){$.push({type:"separator",value:":"}),N++;continue}if(V===","){$.push({type:"delimiter",value:","}),N++;continue}if(V==='"'){let Z="",Y=!1;V=w[++N];while(V!=='"'){if(N===w.length){Y=!0;break}if(V==="\\"){if(N++,N===w.length){Y=!0;break}Z+=V+w[N],V=w[++N]}else Z+=V,V=w[++N]}if(V=w[++N],!Y)$.push({type:"string",value:Z});continue}if(V&&/\s/.test(V)){N++;continue}let W=/[0-9]/;if(V&&W.test(V)||V==="-"||V==="."){let Z="";if(V==="-")Z+=V,V=w[++N];while(V&&W.test(V)||V===".")Z+=V,V=w[++N];$.push({type:"number",value:Z});continue}let X=/[a-z]/i;if(V&&X.test(V)){let Z="";while(V&&X.test(V)){if(N===w.length)break;Z+=V,V=w[++N]}if(Z=="true"||Z=="false"||Z==="null")$.push({type:"name",value:Z});else{N++;continue}continue}N++}return $},Tw=(w)=>{if(w.length===0)return w;let N=w[w.length-1];switch(N.type){case"separator":return w=w.slice(0,w.length-1),Tw(w);break;case"number":let $=N.value[N.value.length-1];if($==="."||$==="-")return w=w.slice(0,w.length-1),Tw(w);case"string":let V=w[w.length-2];if(V?.type==="delimiter")return w=w.slice(0,w.length-1),Tw(w);else if(V?.type==="brace"&&V.value==="{")return w=w.slice(0,w.length-1),Tw(w);break;case"delimiter":return w=w.slice(0,w.length-1),Tw(w);break}return w},R1=(w)=>{let N=[];if(w.map(($)=>{if($.type==="brace")if($.value==="{")N.push("}");else N.splice(N.lastIndexOf("}"),1);if($.type==="paren")if($.value==="[")N.push("]");else N.splice(N.lastIndexOf("]"),1)}),N.length>0)N.reverse().map(($)=>{if($==="}")w.push({type:"brace",value:"}"});else if($==="]")w.push({type:"paren",value:"]"})});return w},_1=(w)=>{let N="";return w.map(($)=>{switch($.type){case"string":N+='"'+$.value+'"';break;default:N+=$.value;break}}),N},y0=(w)=>JSON.parse(_1(R1(Tw(h1(w)))));var m,ww,Cw,nw,j0,tw,aw,k0,ow,t,rw,h0,R0,Zw,_0,b0,sw,CN,C$,d0,ON,UN,fN,O$,U$="__json_buf";function f$(w){return w.type==="tool_use"||w.type==="server_tool_use"||w.type==="mcp_tool_use"}class ew{constructor(w,N){m.add(this),this.messages=[],this.receivedMessages=[],ww.set(this,void 0),Cw.set(this,null),this.controller=new AbortController,nw.set(this,void 0),j0.set(this,()=>{}),tw.set(this,()=>{}),aw.set(this,void 0),k0.set(this,()=>{}),ow.set(this,()=>{}),t.set(this,{}),rw.set(this,!1),h0.set(this,!1),R0.set(this,!1),Zw.set(this,!1),_0.set(this,void 0),b0.set(this,void 0),sw.set(this,void 0),d0.set(this,($)=>{if(x(this,h0,!0,"f"),l($))$=new B;if($ instanceof B)return x(this,R0,!0,"f"),this._emit("abort",$);if($ instanceof Q)return this._emit("error",$);if($ instanceof Error){let V=new Q($.message);return V.cause=$,this._emit("error",V)}return this._emit("error",new Q(String($)))}),x(this,nw,new Promise(($,V)=>{x(this,j0,$,"f"),x(this,tw,V,"f")}),"f"),x(this,aw,new Promise(($,V)=>{x(this,k0,$,"f"),x(this,ow,V,"f")}),"f"),z(this,nw,"f").catch(()=>{}),z(this,aw,"f").catch(()=>{}),x(this,Cw,w,"f"),x(this,sw,N?.logger??console,"f")}get response(){return z(this,_0,"f")}get request_id(){return z(this,b0,"f")}async withResponse(){x(this,Zw,!0,"f");let w=await z(this,nw,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let N=new ew(null);return N._run(()=>N._fromReadableStream(w)),N}static createMessage(w,N,$,{logger:V}={}){let K=new ew(N,{logger:V});for(let W of N.messages)K._addMessageParam(W);return x(K,Cw,{...N,stream:!0},"f"),K._run(()=>K._createMessage(w,{...N,stream:!0},{...$,headers:{...$?.headers,"X-Stainless-Helper-Method":"stream"}})),K}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},z(this,d0,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,N=!0){if(this.receivedMessages.push(w),N)this._emit("message",w)}async _createMessage(w,N,$){let V=$?.signal,K;if(V){if(V.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),V.addEventListener("abort",K)}try{z(this,m,"m",ON).call(this);let{response:W,data:X}=await w.create({...N,stream:!0},{...$,signal:this.controller.signal}).withResponse();this._connected(W);for await(let Z of X)z(this,m,"m",UN).call(this,Z);if(X.controller.signal?.aborted)throw new B;z(this,m,"m",fN).call(this)}finally{if(V&&K)V.removeEventListener("abort",K)}}_connected(w){if(this.ended)return;x(this,_0,w,"f"),x(this,b0,w?.headers.get("request-id"),"f"),z(this,j0,"f").call(this,w),this._emit("connect")}get ended(){return z(this,rw,"f")}get errored(){return z(this,h0,"f")}get aborted(){return z(this,R0,"f")}abort(){this.controller.abort()}on(w,N){return(z(this,t,"f")[w]||(z(this,t,"f")[w]=[])).push({listener:N}),this}off(w,N){let $=z(this,t,"f")[w];if(!$)return this;let V=$.findIndex((K)=>K.listener===N);if(V>=0)$.splice(V,1);return this}once(w,N){return(z(this,t,"f")[w]||(z(this,t,"f")[w]=[])).push({listener:N,once:!0}),this}emitted(w){return new Promise((N,$)=>{if(x(this,Zw,!0,"f"),w!=="error")this.once("error",$);this.once(w,N)})}async done(){x(this,Zw,!0,"f"),await z(this,aw,"f")}get currentMessage(){return z(this,ww,"f")}async finalMessage(){return await this.done(),z(this,m,"m",CN).call(this)}async finalText(){return await this.done(),z(this,m,"m",C$).call(this)}_emit(w,...N){if(z(this,rw,"f"))return;if(w==="end")x(this,rw,!0,"f"),z(this,k0,"f").call(this);let $=z(this,t,"f")[w];if($)z(this,t,"f")[w]=$.filter((V)=>!V.once),$.forEach(({listener:V})=>V(...N));if(w==="abort"){let V=N[0];if(!z(this,Zw,"f")&&!$?.length)Promise.reject(V);z(this,tw,"f").call(this,V),z(this,ow,"f").call(this,V),this._emit("end");return}if(w==="error"){let V=N[0];if(!z(this,Zw,"f")&&!$?.length)Promise.reject(V);z(this,tw,"f").call(this,V),z(this,ow,"f").call(this,V),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",z(this,m,"m",CN).call(this))}async _fromReadableStream(w,N){let $=N?.signal,V;if($){if($.aborted)this.controller.abort();V=this.controller.abort.bind(this.controller),$.addEventListener("abort",V)}try{z(this,m,"m",ON).call(this),this._connected(null);let K=b.fromReadableStream(w,this.controller);for await(let W of K)z(this,m,"m",UN).call(this,W);if(K.controller.signal?.aborted)throw new B;z(this,m,"m",fN).call(this)}finally{if($&&V)$.removeEventListener("abort",V)}}[(ww=new WeakMap,Cw=new WeakMap,nw=new WeakMap,j0=new WeakMap,tw=new WeakMap,aw=new WeakMap,k0=new WeakMap,ow=new WeakMap,t=new WeakMap,rw=new WeakMap,h0=new WeakMap,R0=new WeakMap,Zw=new WeakMap,_0=new WeakMap,b0=new WeakMap,sw=new WeakMap,d0=new WeakMap,m=new WeakSet,CN=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},C$=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");let N=this.receivedMessages.at(-1).content.filter(($)=>$.type==="text").map(($)=>$.text);if(N.length===0)throw new Q("stream ended without producing a content block with type=text");return N.join(" ")},ON=function(){if(this.ended)return;x(this,ww,void 0,"f")},UN=function(N){if(this.ended)return;let $=z(this,m,"m",O$).call(this,N);switch(this._emit("streamEvent",N,$),N.type){case"content_block_delta":{let V=$.content.at(-1);switch(N.delta.type){case"text_delta":{if(V.type==="text")this._emit("text",N.delta.text,V.text||"");break}case"citations_delta":{if(V.type==="text")this._emit("citation",N.delta.citation,V.citations??[]);break}case"input_json_delta":{if(f$(V)&&V.input)this._emit("inputJson",N.delta.partial_json,V.input);break}case"thinking_delta":{if(V.type==="thinking")this._emit("thinking",N.delta.thinking,V.thinking);break}case"signature_delta":{if(V.type==="thinking")this._emit("signature",V.signature);break}case"compaction_delta":{if(V.type==="compaction"&&V.content)this._emit("compaction",V.content);break}default:F$(N.delta)}break}case"message_stop":{this._addMessageParam($),this._addMessage(GN($,z(this,Cw,"f"),{logger:z(this,sw,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",$.content.at(-1));break}case"message_start":{x(this,ww,$,"f");break}case"content_block_start":case"message_delta":break}},fN=function(){if(this.ended)throw new Q("stream has ended, this shouldn't happen");let N=z(this,ww,"f");if(!N)throw new Q("request ended without sending any chunks");return x(this,ww,void 0,"f"),GN(N,z(this,Cw,"f"),{logger:z(this,sw,"f")})},O$=function(N){let $=z(this,ww,"f");if(N.type==="message_start"){if($)throw new Q(`Unexpected event order, got ${N.type} before receiving "message_stop"`);return N.message}if(!$)throw new Q(`Unexpected event order, got ${N.type} before "message_start"`);switch(N.type){case"message_stop":return $;case"message_delta":if($.container=N.delta.container,$.stop_reason=N.delta.stop_reason,$.stop_sequence=N.delta.stop_sequence,$.usage.output_tokens=N.usage.output_tokens,$.context_management=N.context_management,N.usage.input_tokens!=null)$.usage.input_tokens=N.usage.input_tokens;if(N.usage.cache_creation_input_tokens!=null)$.usage.cache_creation_input_tokens=N.usage.cache_creation_input_tokens;if(N.usage.cache_read_input_tokens!=null)$.usage.cache_read_input_tokens=N.usage.cache_read_input_tokens;if(N.usage.server_tool_use!=null)$.usage.server_tool_use=N.usage.server_tool_use;if(N.usage.iterations!=null)$.usage.iterations=N.usage.iterations;return $;case"content_block_start":return $.content.push(N.content_block),$;case"content_block_delta":{let V=$.content.at(N.index);switch(N.delta.type){case"text_delta":{if(V?.type==="text")$.content[N.index]={...V,text:(V.text||"")+N.delta.text};break}case"citations_delta":{if(V?.type==="text")$.content[N.index]={...V,citations:[...V.citations??[],N.delta.citation]};break}case"input_json_delta":{if(V&&f$(V)){let K=V[U$]||"";K+=N.delta.partial_json;let W={...V};if(Object.defineProperty(W,U$,{value:K,enumerable:!1,writable:!0}),K)try{W.input=y0(K)}catch(X){let Z=new Q(`Unable to parse tool parameter JSON from model. Please retry your request or adjust your prompt. Error: ${X}. JSON: ${K}`);z(this,d0,"f").call(this,Z)}$.content[N.index]=W}break}case"thinking_delta":{if(V?.type==="thinking")$.content[N.index]={...V,thinking:V.thinking+N.delta.thinking};break}case"signature_delta":{if(V?.type==="thinking")$.content[N.index]={...V,signature:N.delta.signature};break}case"compaction_delta":{if(V?.type==="compaction")$.content[N.index]={...V,content:(V.content||"")+N.delta.content};break}default:F$(N.delta)}return $}case"content_block_stop":return $}},Symbol.asyncIterator)](){let w=[],N=[],$=!1;return this.on("streamEvent",(V)=>{let K=N.shift();if(K)K.resolve(V);else w.push(V)}),this.on("end",()=>{$=!0;for(let V of N)V.resolve(void 0);N.length=0}),this.on("abort",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),this.on("error",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),{next:async()=>{if(!w.length){if($)return{value:void 0,done:!0};return new Promise((K,W)=>N.push({resolve:K,reject:W})).then((K)=>K?{value:K,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new b(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function F$(w){}class Ow extends Error{constructor(w){let N=typeof w==="string"?w:w.map(($)=>{if($.type==="text")return $.text;return`[${$.type}]`}).join(" ");super(N);this.name="ToolError",this.content=w}}var S$=1e5,A$=`You have been working on the task described above but have not yet completed it. Write a continuation summary that will allow you (or another instance of yourself) to resume work efficiently in a future context window where the conversation history will be replaced with this summary. Your summary should be structured, concise, and actionable. Include:
 1. Task Overview
 The user's core request and success criteria
 Any clarifications or constraints they specified
@@ -28,26 +28,26 @@ User preferences or style requirements
 Domain-specific details that aren't obvious
 Any promises made to the user
 Be concise but complete—err on the side of including information that would prevent duplicate work or repeated mistakes. Write in a way that enables immediate resumption of the task.
-Wrap your summary in <summary></summary> tags.`;var w0,Uw,Jw,E,j,d,a,Nw,N0,q1,SN;function H1(){let w,N;return{promise:new Promise((V,K)=>{w=V,N=K}),resolve:w,reject:N}}class $0{constructor(w,N,$){w0.add(this),this.client=w,Uw.set(this,!1),Jw.set(this,!1),E.set(this,void 0),j.set(this,void 0),d.set(this,void 0),a.set(this,void 0),Nw.set(this,void 0),N0.set(this,0),x(this,E,{params:{...N,messages:structuredClone(N.messages)}},"f");let K=["BetaToolRunner",...QN(N.tools,N.messages)].join(", ");if(x(this,j,{...$,headers:J([{"x-stainless-helper":K},$?.headers])},"f"),x(this,Nw,H1(),"f"),N.compactionControl?.enabled)console.warn('Anthropic: The `compactionControl` parameter is deprecated and will be removed in a future version. Use server-side compaction instead by passing `edits: [{ type: "compact_20260112" }]` in the params passed to `toolRunner()`. See https://platform.claude.com/docs/en/build-with-claude/compaction')}async*[(Uw=new WeakMap,Jw=new WeakMap,E=new WeakMap,j=new WeakMap,d=new WeakMap,a=new WeakMap,Nw=new WeakMap,N0=new WeakMap,w0=new WeakSet,q1=async function(){let N=z(this,E,"f").params.compactionControl;if(!N||!N.enabled)return!1;let $=0;if(z(this,d,"f")!==void 0)try{let Y=await z(this,d,"f");$=Y.usage.input_tokens+(Y.usage.cache_creation_input_tokens??0)+(Y.usage.cache_read_input_tokens??0)+Y.usage.output_tokens}catch{return!1}let V=N.contextTokenThreshold??f1;if($<V)return!1;let K=N.model??z(this,E,"f").params.model,W=N.summaryPrompt??A1,X=z(this,E,"f").params.messages;if(X[X.length-1].role==="assistant"){let Y=X[X.length-1];if(Array.isArray(Y.content)){let M=Y.content.filter((C)=>C.type!=="tool_use");if(M.length===0)X.pop();else Y.content=M}}let Z=await this.client.beta.messages.create({model:K,messages:[...X,{role:"user",content:[{type:"text",text:W}]}],max_tokens:z(this,E,"f").params.max_tokens},{signal:z(this,j,"f").signal,headers:J([z(this,j,"f").headers,{"x-stainless-helper":"compaction"}])});if(Z.content[0]?.type!=="text")throw new Q("Expected text response for compaction");return z(this,E,"f").params.messages=[{role:"user",content:Z.content}],!0},Symbol.asyncIterator)](){var w;if(z(this,Uw,"f"))throw new Q("Cannot iterate over a consumed stream");x(this,Uw,!0,"f"),x(this,Jw,!0,"f"),x(this,a,void 0,"f");try{while(!0){let N;try{if(z(this,E,"f").params.max_iterations&&z(this,N0,"f")>=z(this,E,"f").params.max_iterations)break;x(this,Jw,!1,"f"),x(this,a,void 0,"f"),x(this,N0,(w=z(this,N0,"f"),w++,w),"f"),x(this,d,void 0,"f");let{max_iterations:$,compactionControl:V,...K}=z(this,E,"f").params;if(K.stream)N=this.client.beta.messages.stream({...K},z(this,j,"f")),x(this,d,N.finalMessage(),"f"),z(this,d,"f").catch(()=>{}),yield N;else x(this,d,this.client.beta.messages.create({...K,stream:!1},z(this,j,"f")),"f"),yield z(this,d,"f");if(!await z(this,w0,"m",q1).call(this)){if(!z(this,Jw,"f")){let{role:Z,content:Y}=await z(this,d,"f");z(this,E,"f").params.messages.push({role:Z,content:Y})}let X=await z(this,w0,"m",SN).call(this,z(this,E,"f").params.messages.at(-1));if(X)z(this,E,"f").params.messages.push(X);else if(!z(this,Jw,"f"))break}}finally{if(N)N.abort()}}if(!z(this,d,"f"))throw new Q("ToolRunner concluded without a message from the server");z(this,Nw,"f").resolve(await z(this,d,"f"))}catch(N){throw x(this,Uw,!1,"f"),z(this,Nw,"f").promise.catch(()=>{}),z(this,Nw,"f").reject(N),x(this,Nw,H1(),"f"),N}}setMessagesParams(w){if(typeof w==="function")z(this,E,"f").params=w(z(this,E,"f").params);else z(this,E,"f").params=w;x(this,Jw,!0,"f"),x(this,a,void 0,"f")}setRequestOptions(w){if(typeof w==="function")x(this,j,w(z(this,j,"f")),"f");else x(this,j,{...z(this,j,"f"),...w},"f")}async generateToolResponse(w=z(this,j,"f").signal){let N=await z(this,d,"f")??this.params.messages.at(-1);if(!N)return null;return z(this,w0,"m",SN).call(this,N,w)}done(){return z(this,Nw,"f").promise}async runUntilDone(){if(!z(this,Uw,"f"))for await(let w of this);return this.done()}get params(){return z(this,E,"f").params}pushMessages(...w){this.setMessagesParams((N)=>({...N,messages:[...N.messages,...w]}))}then(w,N){return this.runUntilDone().then(w,N)}}SN=async function(N,$=z(this,j,"f").signal){if(z(this,a,"f")!==void 0)return z(this,a,"f");return x(this,a,h$(z(this,E,"f").params,N,{...z(this,j,"f"),signal:$}),"f"),z(this,a,"f")};async function h$(w,N=w.messages.at(-1),$){if(!N||N.role!=="assistant"||!N.content||typeof N.content==="string")return null;let V=N.content.filter((W)=>W.type==="tool_use");if(V.length===0)return null;return{role:"user",content:await Promise.all(V.map(async(W)=>{let X=w.tools.find((Z)=>("name"in Z?Z.name:Z.mcp_server_name)===W.name);if(!X||!("run"in X))return{type:"tool_result",tool_use_id:W.id,content:`Error: Tool '${W.name}' not found`,is_error:!0};try{let Z=W.input;if("parse"in X&&X.parse)Z=X.parse(Z);let Y=await X.run(Z,{toolUseBlock:W,signal:$?.signal});return{type:"tool_result",tool_use_id:W.id,content:Y}}catch(Z){return{type:"tool_result",tool_use_id:W.id,content:Z instanceof Ow?Z.content:`Error: ${Z instanceof Error?Z.message:String(Z)}`,is_error:!0}}}))}}class Fw{constructor(w,N){this.iterator=w,this.controller=N}async*decoder(){let w=new e;for await(let N of this.iterator)for(let $ of w.decode(N))yield JSON.parse($);for(let N of w.flush())yield JSON.parse(N)}[Symbol.asyncIterator](){return this.decoder()}static fromResponse(w,N){if(!w.body){if(N.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new Q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new Q("Attempted to iterate over a response with no body")}return new Fw(Rw(w.body),N)}}class V0 extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/messages/batches?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"message-batches-2024-09-24"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/messages/batches/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/messages/batches?beta=true",i,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"message-batches-2024-09-24"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/messages/batches/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}cancel(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/messages/batches/${w}/cancel?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}async results(w,N={},$){let V=await this.retrieve(w);if(!V.results_url)throw new Q(`No batch \`results_url\`; Has it finished processing? ${V.processing_status} - ${V.id}`);let{betas:K}=N??{};return this._client.get(V.results_url,{...$,headers:J([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString(),Accept:"application/binary"},$?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((W,X)=>Fw.fromResponse(X.response,X.controller))}}var L1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026"},_$=["claude-opus-4-6"];class $w extends O{constructor(){super(...arguments);this.batches=new V0(this._client)}create(w,N){let $=E1(w),{betas:V,...K}=$;if(K.model in L1)console.warn(`The model '${K.model}' is deprecated and will reach end-of-life on ${L1[K.model]}
-Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(K.model in _$&&K.thinking&&K.thinking.type==="enabled")console.warn(`Using Claude with ${K.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let W=this._client._options.timeout;if(!K.stream&&W==null){let Z=I0[K.model]??void 0;W=this._client.calculateNonstreamingTimeout(K.max_tokens,Z)}let X=B0(K.tools,K.messages);return this._client.post("/v1/messages?beta=true",{body:K,timeout:W??600000,...N,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},X,N?.headers]),stream:$.stream??!1})}parse(w,N){return N={...N,headers:J([{"anthropic-beta":[...w.betas??[],"structured-outputs-2025-12-15"].toString()},N?.headers])},this.create(w,N).then(($)=>TN($,w,{logger:this._client.logger??console}))}stream(w,N){return ew.createMessage(this,w,N)}countTokens(w,N){let $=E1(w),{betas:V,...K}=$;return this._client.post("/v1/messages/count_tokens?beta=true",{body:K,...N,headers:J([{"anthropic-beta":[...V??[],"token-counting-2024-11-01"].toString()},N?.headers])})}toolRunner(w,N){return new $0(this._client,w,N)}}function E1(w){if(!w.output_format)return w;if(w.output_config?.format)throw new Q("Both output_format and output_config.format were provided. Please use only output_config.format (output_format is deprecated).");let{output_format:N,...$}=w;return{...$,output_config:{...w.output_config,format:N}}}$w.Batches=V0;$w.BetaToolRunner=$0;$w.ToolError=Ow;class K0 extends O{list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/sessions/${w}/events?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}send(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}/events?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}stream(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/sessions/${w}/events/stream?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers]),stream:!0})}}class W0 extends O{retrieve(w,N,$){let{session_id:V,betas:K}=N;return this._client.get(G`/v1/sessions/${V}/resources/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{session_id:V,betas:K,...W}=N;return this._client.post(G`/v1/sessions/${V}/resources/${w}?beta=true`,{body:W,...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/sessions/${w}/resources?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,N,$){let{session_id:V,betas:K}=N;return this._client.delete(G`/v1/sessions/${V}/resources/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}add(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}/resources?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class Yw extends O{constructor(){super(...arguments);this.events=new K0(this._client),this.resources=new W0(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/sessions?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/sessions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/sessions?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/sessions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/sessions/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}Yw.Events=K0;Yw.Resources=W0;class X0 extends O{create(w,N={},$){let{betas:V,...K}=N??{};return this._client.post(G`/v1/skills/${w}/versions?beta=true`,Qw({body:K,...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])},this._client))}retrieve(w,N,$){let{skill_id:V,betas:K}=N;return this._client.get(G`/v1/skills/${V}/versions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/skills/${w}/versions?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}delete(w,N,$){let{skill_id:V,betas:K}=N;return this._client.delete(G`/v1/skills/${V}/versions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},$?.headers])})}}class Sw extends O{constructor(){super(...arguments);this.versions=new X0(this._client)}create(w={},N){let{betas:$,...V}=w??{};return this._client.post("/v1/skills?beta=true",Qw({body:V,...N,headers:J([{"anthropic-beta":[...$??[],"skills-2025-10-02"].toString()},N?.headers])},this._client,!1))}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/skills/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/skills?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"skills-2025-10-02"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/skills/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}}Sw.Versions=X0;class z0 extends O{create(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/vaults/${w}/credentials?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,N,$){let{vault_id:V,betas:K}=N;return this._client.get(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{vault_id:V,betas:K,...W}=N;return this._client.post(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{body:W,...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/vaults/${w}/credentials?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,N,$){let{vault_id:V,betas:K}=N;return this._client.delete(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N,$){let{vault_id:V,betas:K}=N;return this._client.post(G`/v1/vaults/${V}/credentials/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class fw extends O{constructor(){super(...arguments);this.credentials=new z0(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/vaults?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/vaults/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/vaults/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/vaults?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/vaults/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/vaults/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}fw.Credentials=z0;class y extends O{constructor(){super(...arguments);this.models=new pw(this._client),this.messages=new $w(this._client),this.agents=new Gw(this._client),this.environments=new mw(this._client),this.sessions=new Yw(this._client),this.vaults=new fw(this._client),this.files=new iw(this._client),this.skills=new Sw(this._client)}}y.Models=pw;y.Messages=$w;y.Agents=Gw;y.Environments=mw;y.Sessions=Yw;y.Vaults=fw;y.Files=iw;y.Skills=Sw;class Aw extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/complete",{body:V,timeout:this._client._options.timeout??600000,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers]),stream:w.stream??!1})}}function P1(w){return w?.output_config?.format}function fN(w,N,$){let V=P1(N);if(!N||!("parse"in(V??{})))return{...w,content:w.content.map((K)=>{if(K.type==="text")return Object.defineProperty({...K},"parsed_output",{value:null,enumerable:!1});return K}),parsed_output:null};return AN(w,N,$)}function AN(w,N,$){let V=null,K=w.content.map((W)=>{if(W.type==="text"){let X=l$(N,W.text);if(V===null)V=X;return Object.defineProperty({...W},"parsed_output",{value:X,enumerable:!1})}return W});return{...w,content:K,parsed_output:V}}function l$(w,N){let $=P1(w);if($?.type!=="json_schema")return null;try{if("parse"in $)return $.parse(N);return JSON.parse(N)}catch(V){throw new Q(`Failed to parse structured output: ${V}`)}}var u,Vw,qw,Z0,c0,J0,Y0,v0,x0,o,M0,m0,u0,xw,i0,p0,Q0,qN,D1,HN,LN,EN,PN,g1,B1="__json_buf";function I1(w){return w.type==="tool_use"||w.type==="server_tool_use"}class G0{constructor(w,N){u.add(this),this.messages=[],this.receivedMessages=[],Vw.set(this,void 0),qw.set(this,null),this.controller=new AbortController,Z0.set(this,void 0),c0.set(this,()=>{}),J0.set(this,()=>{}),Y0.set(this,void 0),v0.set(this,()=>{}),x0.set(this,()=>{}),o.set(this,{}),M0.set(this,!1),m0.set(this,!1),u0.set(this,!1),xw.set(this,!1),i0.set(this,void 0),p0.set(this,void 0),Q0.set(this,void 0),HN.set(this,($)=>{if(x(this,m0,!0,"f"),l($))$=new B;if($ instanceof B)return x(this,u0,!0,"f"),this._emit("abort",$);if($ instanceof Q)return this._emit("error",$);if($ instanceof Error){let V=new Q($.message);return V.cause=$,this._emit("error",V)}return this._emit("error",new Q(String($)))}),x(this,Z0,new Promise(($,V)=>{x(this,c0,$,"f"),x(this,J0,V,"f")}),"f"),x(this,Y0,new Promise(($,V)=>{x(this,v0,$,"f"),x(this,x0,V,"f")}),"f"),z(this,Z0,"f").catch(()=>{}),z(this,Y0,"f").catch(()=>{}),x(this,qw,w,"f"),x(this,Q0,N?.logger??console,"f")}get response(){return z(this,i0,"f")}get request_id(){return z(this,p0,"f")}async withResponse(){x(this,xw,!0,"f");let w=await z(this,Z0,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let N=new G0(null);return N._run(()=>N._fromReadableStream(w)),N}static createMessage(w,N,$,{logger:V}={}){let K=new G0(N,{logger:V});for(let W of N.messages)K._addMessageParam(W);return x(K,qw,{...N,stream:!0},"f"),K._run(()=>K._createMessage(w,{...N,stream:!0},{...$,headers:{...$?.headers,"X-Stainless-Helper-Method":"stream"}})),K}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},z(this,HN,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,N=!0){if(this.receivedMessages.push(w),N)this._emit("message",w)}async _createMessage(w,N,$){let V=$?.signal,K;if(V){if(V.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),V.addEventListener("abort",K)}try{z(this,u,"m",LN).call(this);let{response:W,data:X}=await w.create({...N,stream:!0},{...$,signal:this.controller.signal}).withResponse();this._connected(W);for await(let Z of X)z(this,u,"m",EN).call(this,Z);if(X.controller.signal?.aborted)throw new B;z(this,u,"m",PN).call(this)}finally{if(V&&K)V.removeEventListener("abort",K)}}_connected(w){if(this.ended)return;x(this,i0,w,"f"),x(this,p0,w?.headers.get("request-id"),"f"),z(this,c0,"f").call(this,w),this._emit("connect")}get ended(){return z(this,M0,"f")}get errored(){return z(this,m0,"f")}get aborted(){return z(this,u0,"f")}abort(){this.controller.abort()}on(w,N){return(z(this,o,"f")[w]||(z(this,o,"f")[w]=[])).push({listener:N}),this}off(w,N){let $=z(this,o,"f")[w];if(!$)return this;let V=$.findIndex((K)=>K.listener===N);if(V>=0)$.splice(V,1);return this}once(w,N){return(z(this,o,"f")[w]||(z(this,o,"f")[w]=[])).push({listener:N,once:!0}),this}emitted(w){return new Promise((N,$)=>{if(x(this,xw,!0,"f"),w!=="error")this.once("error",$);this.once(w,N)})}async done(){x(this,xw,!0,"f"),await z(this,Y0,"f")}get currentMessage(){return z(this,Vw,"f")}async finalMessage(){return await this.done(),z(this,u,"m",qN).call(this)}async finalText(){return await this.done(),z(this,u,"m",D1).call(this)}_emit(w,...N){if(z(this,M0,"f"))return;if(w==="end")x(this,M0,!0,"f"),z(this,v0,"f").call(this);let $=z(this,o,"f")[w];if($)z(this,o,"f")[w]=$.filter((V)=>!V.once),$.forEach(({listener:V})=>V(...N));if(w==="abort"){let V=N[0];if(!z(this,xw,"f")&&!$?.length)Promise.reject(V);z(this,J0,"f").call(this,V),z(this,x0,"f").call(this,V),this._emit("end");return}if(w==="error"){let V=N[0];if(!z(this,xw,"f")&&!$?.length)Promise.reject(V);z(this,J0,"f").call(this,V),z(this,x0,"f").call(this,V),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",z(this,u,"m",qN).call(this))}async _fromReadableStream(w,N){let $=N?.signal,V;if($){if($.aborted)this.controller.abort();V=this.controller.abort.bind(this.controller),$.addEventListener("abort",V)}try{z(this,u,"m",LN).call(this),this._connected(null);let K=b.fromReadableStream(w,this.controller);for await(let W of K)z(this,u,"m",EN).call(this,W);if(K.controller.signal?.aborted)throw new B;z(this,u,"m",PN).call(this)}finally{if($&&V)$.removeEventListener("abort",V)}}[(Vw=new WeakMap,qw=new WeakMap,Z0=new WeakMap,c0=new WeakMap,J0=new WeakMap,Y0=new WeakMap,v0=new WeakMap,x0=new WeakMap,o=new WeakMap,M0=new WeakMap,m0=new WeakMap,u0=new WeakMap,xw=new WeakMap,i0=new WeakMap,p0=new WeakMap,Q0=new WeakMap,HN=new WeakMap,u=new WeakSet,qN=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},D1=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");let N=this.receivedMessages.at(-1).content.filter(($)=>$.type==="text").map(($)=>$.text);if(N.length===0)throw new Q("stream ended without producing a content block with type=text");return N.join(" ")},LN=function(){if(this.ended)return;x(this,Vw,void 0,"f")},EN=function(N){if(this.ended)return;let $=z(this,u,"m",g1).call(this,N);switch(this._emit("streamEvent",N,$),N.type){case"content_block_delta":{let V=$.content.at(-1);switch(N.delta.type){case"text_delta":{if(V.type==="text")this._emit("text",N.delta.text,V.text||"");break}case"citations_delta":{if(V.type==="text")this._emit("citation",N.delta.citation,V.citations??[]);break}case"input_json_delta":{if(I1(V)&&V.input)this._emit("inputJson",N.delta.partial_json,V.input);break}case"thinking_delta":{if(V.type==="thinking")this._emit("thinking",N.delta.thinking,V.thinking);break}case"signature_delta":{if(V.type==="thinking")this._emit("signature",V.signature);break}default:j1(N.delta)}break}case"message_stop":{this._addMessageParam($),this._addMessage(fN($,z(this,qw,"f"),{logger:z(this,Q0,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",$.content.at(-1));break}case"message_start":{x(this,Vw,$,"f");break}case"content_block_start":case"message_delta":break}},PN=function(){if(this.ended)throw new Q("stream has ended, this shouldn't happen");let N=z(this,Vw,"f");if(!N)throw new Q("request ended without sending any chunks");return x(this,Vw,void 0,"f"),fN(N,z(this,qw,"f"),{logger:z(this,Q0,"f")})},g1=function(N){let $=z(this,Vw,"f");if(N.type==="message_start"){if($)throw new Q(`Unexpected event order, got ${N.type} before receiving "message_stop"`);return N.message}if(!$)throw new Q(`Unexpected event order, got ${N.type} before "message_start"`);switch(N.type){case"message_stop":return $;case"message_delta":if($.stop_reason=N.delta.stop_reason,$.stop_sequence=N.delta.stop_sequence,$.usage.output_tokens=N.usage.output_tokens,N.usage.input_tokens!=null)$.usage.input_tokens=N.usage.input_tokens;if(N.usage.cache_creation_input_tokens!=null)$.usage.cache_creation_input_tokens=N.usage.cache_creation_input_tokens;if(N.usage.cache_read_input_tokens!=null)$.usage.cache_read_input_tokens=N.usage.cache_read_input_tokens;if(N.usage.server_tool_use!=null)$.usage.server_tool_use=N.usage.server_tool_use;return $;case"content_block_start":return $.content.push({...N.content_block}),$;case"content_block_delta":{let V=$.content.at(N.index);switch(N.delta.type){case"text_delta":{if(V?.type==="text")$.content[N.index]={...V,text:(V.text||"")+N.delta.text};break}case"citations_delta":{if(V?.type==="text")$.content[N.index]={...V,citations:[...V.citations??[],N.delta.citation]};break}case"input_json_delta":{if(V&&I1(V)){let K=V[B1]||"";K+=N.delta.partial_json;let W={...V};if(Object.defineProperty(W,B1,{value:K,enumerable:!1,writable:!0}),K)W.input=j0(K);$.content[N.index]=W}break}case"thinking_delta":{if(V?.type==="thinking")$.content[N.index]={...V,thinking:V.thinking+N.delta.thinking};break}case"signature_delta":{if(V?.type==="thinking")$.content[N.index]={...V,signature:N.delta.signature};break}default:j1(N.delta)}return $}case"content_block_stop":return $}},Symbol.asyncIterator)](){let w=[],N=[],$=!1;return this.on("streamEvent",(V)=>{let K=N.shift();if(K)K.resolve(V);else w.push(V)}),this.on("end",()=>{$=!0;for(let V of N)V.resolve(void 0);N.length=0}),this.on("abort",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),this.on("error",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),{next:async()=>{if(!w.length){if($)return{value:void 0,done:!0};return new Promise((K,W)=>N.push({resolve:K,reject:W})).then((K)=>K?{value:K,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new b(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function j1(w){}class T0 extends O{create(w,N){return this._client.post("/v1/messages/batches",{body:w,...N})}retrieve(w,N){return this._client.get(G`/v1/messages/batches/${w}`,N)}list(w={},N){return this._client.getAPIList("/v1/messages/batches",i,{query:w,...N})}delete(w,N){return this._client.delete(G`/v1/messages/batches/${w}`,N)}cancel(w,N){return this._client.post(G`/v1/messages/batches/${w}/cancel`,N)}async results(w,N){let $=await this.retrieve(w);if(!$.results_url)throw new Q(`No batch \`results_url\`; Has it finished processing? ${$.processing_status} - ${$.id}`);return this._client.get($.results_url,{...N,headers:J([{Accept:"application/binary"},N?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((V,K)=>Fw.fromResponse(K.response,K.controller))}}class Mw extends O{constructor(){super(...arguments);this.batches=new T0(this._client)}create(w,N){if(w.model in y1)console.warn(`The model '${w.model}' is deprecated and will reach end-of-life on ${y1[w.model]}
-Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(w.model in t$&&w.thinking&&w.thinking.type==="enabled")console.warn(`Using Claude with ${w.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let $=this._client._options.timeout;if(!w.stream&&$==null){let K=I0[w.model]??void 0;$=this._client.calculateNonstreamingTimeout(w.max_tokens,K)}let V=B0(w.tools,w.messages);return this._client.post("/v1/messages",{body:w,timeout:$??600000,...N,headers:J([V,N?.headers]),stream:w.stream??!1})}parse(w,N){return this.create(w,N).then(($)=>AN($,w,{logger:this._client.logger??console}))}stream(w,N){return G0.createMessage(this,w,N,{logger:this._client.logger??console})}countTokens(w,N){return this._client.post("/v1/messages/count_tokens",{body:w,...N})}}var y1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026","claude-3-5-haiku-latest":"February 19th, 2026","claude-3-5-haiku-20241022":"February 19th, 2026","claude-opus-4-0":"June 15th, 2026","claude-opus-4-20250514":"June 15th, 2026","claude-sonnet-4-0":"June 15th, 2026","claude-sonnet-4-20250514":"June 15th, 2026"},t$=["claude-opus-4-6"];Mw.Batches=T0;class Hw extends O{retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/models/${w}`,{...$,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/models",i,{query:V,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers])})}}var C0=(w)=>{if(typeof globalThis.process<"u")return globalThis.process.env?.[w]?.trim()||void 0;if(typeof globalThis.Deno<"u")return globalThis.Deno.env?.get?.(w)?.trim()||void 0;return};var DN,gN,l0,k1,h1="\\n\\nHuman:",R1="\\n\\nAssistant:";class A{constructor({baseURL:w=C0("ANTHROPIC_BASE_URL"),apiKey:N=C0("ANTHROPIC_API_KEY")??null,authToken:$=C0("ANTHROPIC_AUTH_TOKEN")??null,...V}={}){DN.add(this),l0.set(this,void 0);let K={apiKey:N,authToken:$,...V,baseURL:w||"https://api.anthropic.com"};if(!K.dangerouslyAllowBrowser&&rN())throw new Q(`It looks like you're running in a browser-like environment.
+Wrap your summary in <summary></summary> tags.`;var w0,Uw,Jw,H,y,d,a,Nw,N0,q$,FN;function E$(){let w,N;return{promise:new Promise((V,K)=>{w=V,N=K}),resolve:w,reject:N}}class $0{constructor(w,N,$){w0.add(this),this.client=w,Uw.set(this,!1),Jw.set(this,!1),H.set(this,void 0),y.set(this,void 0),d.set(this,void 0),a.set(this,void 0),Nw.set(this,void 0),N0.set(this,0),x(this,H,{params:{...N,messages:structuredClone(N.messages)}},"f");let K=["BetaToolRunner",...QN(N.tools,N.messages)].join(", ");if(x(this,y,{...$,headers:J([{"x-stainless-helper":K},$?.headers])},"f"),x(this,Nw,E$(),"f"),N.compactionControl?.enabled)console.warn('Anthropic: The `compactionControl` parameter is deprecated and will be removed in a future version. Use server-side compaction instead by passing `edits: [{ type: "compact_20260112" }]` in the params passed to `toolRunner()`. See https://platform.claude.com/docs/en/build-with-claude/compaction')}async*[(Uw=new WeakMap,Jw=new WeakMap,H=new WeakMap,y=new WeakMap,d=new WeakMap,a=new WeakMap,Nw=new WeakMap,N0=new WeakMap,w0=new WeakSet,q$=async function(){let N=z(this,H,"f").params.compactionControl;if(!N||!N.enabled)return!1;let $=0;if(z(this,d,"f")!==void 0)try{let Y=await z(this,d,"f");$=Y.usage.input_tokens+(Y.usage.cache_creation_input_tokens??0)+(Y.usage.cache_read_input_tokens??0)+Y.usage.output_tokens}catch{return!1}let V=N.contextTokenThreshold??S$;if($<V)return!1;let K=N.model??z(this,H,"f").params.model,W=N.summaryPrompt??A$,X=z(this,H,"f").params.messages;if(X[X.length-1].role==="assistant"){let Y=X[X.length-1];if(Array.isArray(Y.content)){let M=Y.content.filter((C)=>C.type!=="tool_use");if(M.length===0)X.pop();else Y.content=M}}let Z=await this.client.beta.messages.create({model:K,messages:[...X,{role:"user",content:[{type:"text",text:W}]}],max_tokens:z(this,H,"f").params.max_tokens},{signal:z(this,y,"f").signal,headers:J([z(this,y,"f").headers,{"x-stainless-helper":"compaction"}])});if(Z.content[0]?.type!=="text")throw new Q("Expected text response for compaction");return z(this,H,"f").params.messages=[{role:"user",content:Z.content}],!0},Symbol.asyncIterator)](){var w;if(z(this,Uw,"f"))throw new Q("Cannot iterate over a consumed stream");x(this,Uw,!0,"f"),x(this,Jw,!0,"f"),x(this,a,void 0,"f");try{while(!0){let N;try{if(z(this,H,"f").params.max_iterations&&z(this,N0,"f")>=z(this,H,"f").params.max_iterations)break;x(this,Jw,!1,"f"),x(this,a,void 0,"f"),x(this,N0,(w=z(this,N0,"f"),w++,w),"f"),x(this,d,void 0,"f");let{max_iterations:$,compactionControl:V,...K}=z(this,H,"f").params;if(K.stream)N=this.client.beta.messages.stream({...K},z(this,y,"f")),x(this,d,N.finalMessage(),"f"),z(this,d,"f").catch(()=>{}),yield N;else x(this,d,this.client.beta.messages.create({...K,stream:!1},z(this,y,"f")),"f"),yield z(this,d,"f");if(!await z(this,w0,"m",q$).call(this)){if(!z(this,Jw,"f")){let{role:Z,content:Y}=await z(this,d,"f");z(this,H,"f").params.messages.push({role:Z,content:Y})}let X=await z(this,w0,"m",FN).call(this,z(this,H,"f").params.messages.at(-1));if(X)z(this,H,"f").params.messages.push(X);else if(!z(this,Jw,"f"))break}}finally{if(N)N.abort()}}if(!z(this,d,"f"))throw new Q("ToolRunner concluded without a message from the server");z(this,Nw,"f").resolve(await z(this,d,"f"))}catch(N){throw x(this,Uw,!1,"f"),z(this,Nw,"f").promise.catch(()=>{}),z(this,Nw,"f").reject(N),x(this,Nw,E$(),"f"),N}}setMessagesParams(w){if(typeof w==="function")z(this,H,"f").params=w(z(this,H,"f").params);else z(this,H,"f").params=w;x(this,Jw,!0,"f"),x(this,a,void 0,"f")}setRequestOptions(w){if(typeof w==="function")x(this,y,w(z(this,y,"f")),"f");else x(this,y,{...z(this,y,"f"),...w},"f")}async generateToolResponse(w=z(this,y,"f").signal){let N=await z(this,d,"f")??this.params.messages.at(-1);if(!N)return null;return z(this,w0,"m",FN).call(this,N,w)}done(){return z(this,Nw,"f").promise}async runUntilDone(){if(!z(this,Uw,"f"))for await(let w of this);return this.done()}get params(){return z(this,H,"f").params}pushMessages(...w){this.setMessagesParams((N)=>({...N,messages:[...N.messages,...w]}))}then(w,N){return this.runUntilDone().then(w,N)}}FN=async function(N,$=z(this,y,"f").signal){if(z(this,a,"f")!==void 0)return z(this,a,"f");return x(this,a,b1(z(this,H,"f").params,N,{...z(this,y,"f"),signal:$}),"f"),z(this,a,"f")};async function b1(w,N=w.messages.at(-1),$){if(!N||N.role!=="assistant"||!N.content||typeof N.content==="string")return null;let V=N.content.filter((W)=>W.type==="tool_use");if(V.length===0)return null;return{role:"user",content:await Promise.all(V.map(async(W)=>{let X=w.tools.find((Z)=>("name"in Z?Z.name:Z.mcp_server_name)===W.name);if(!X||!("run"in X))return{type:"tool_result",tool_use_id:W.id,content:`Error: Tool '${W.name}' not found`,is_error:!0};try{let Z=W.input;if("parse"in X&&X.parse)Z=X.parse(Z);let Y=await X.run(Z,{toolUseBlock:W,signal:$?.signal});return{type:"tool_result",tool_use_id:W.id,content:Y}}catch(Z){return{type:"tool_result",tool_use_id:W.id,content:Z instanceof Ow?Z.content:`Error: ${Z instanceof Error?Z.message:String(Z)}`,is_error:!0}}}))}}class fw{constructor(w,N){this.iterator=w,this.controller=N}async*decoder(){let w=new e;for await(let N of this.iterator)for(let $ of w.decode(N))yield JSON.parse($);for(let N of w.flush())yield JSON.parse(N)}[Symbol.asyncIterator](){return this.decoder()}static fromResponse(w,N){if(!w.body){if(N.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new Q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new Q("Attempted to iterate over a response with no body")}return new fw(Rw(w.body),N)}}class V0 extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/messages/batches?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"message-batches-2024-09-24"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/messages/batches/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/messages/batches?beta=true",i,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"message-batches-2024-09-24"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/messages/batches/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}cancel(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/messages/batches/${w}/cancel?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"message-batches-2024-09-24"].toString()},$?.headers])})}async results(w,N={},$){let V=await this.retrieve(w);if(!V.results_url)throw new Q(`No batch \`results_url\`; Has it finished processing? ${V.processing_status} - ${V.id}`);let{betas:K}=N??{};return this._client.get(V.results_url,{...$,headers:J([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString(),Accept:"application/binary"},$?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((W,X)=>fw.fromResponse(X.response,X.controller))}}var L$={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026"},c1=["claude-opus-4-6"];class $w extends O{constructor(){super(...arguments);this.batches=new V0(this._client)}create(w,N){let $=H$(w),{betas:V,...K}=$;if(K.model in L$)console.warn(`The model '${K.model}' is deprecated and will reach end-of-life on ${L$[K.model]}
+Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(K.model in c1&&K.thinking&&K.thinking.type==="enabled")console.warn(`Using Claude with ${K.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let W=this._client._options.timeout;if(!K.stream&&W==null){let Z=I0[K.model]??void 0;W=this._client.calculateNonstreamingTimeout(K.max_tokens,Z)}let X=B0(K.tools,K.messages);return this._client.post("/v1/messages?beta=true",{body:K,timeout:W??600000,...N,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},X,N?.headers]),stream:$.stream??!1})}parse(w,N){return N={...N,headers:J([{"anthropic-beta":[...w.betas??[],"structured-outputs-2025-12-15"].toString()},N?.headers])},this.create(w,N).then(($)=>TN($,w,{logger:this._client.logger??console}))}stream(w,N){return ew.createMessage(this,w,N)}countTokens(w,N){let $=H$(w),{betas:V,...K}=$;return this._client.post("/v1/messages/count_tokens?beta=true",{body:K,...N,headers:J([{"anthropic-beta":[...V??[],"token-counting-2024-11-01"].toString()},N?.headers])})}toolRunner(w,N){return new $0(this._client,w,N)}}function H$(w){if(!w.output_format)return w;if(w.output_config?.format)throw new Q("Both output_format and output_config.format were provided. Please use only output_config.format (output_format is deprecated).");let{output_format:N,...$}=w;return{...$,output_config:{...w.output_config,format:N}}}$w.Batches=V0;$w.BetaToolRunner=$0;$w.ToolError=Ow;class K0 extends O{list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/sessions/${w}/events?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}send(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}/events?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}stream(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/sessions/${w}/events/stream?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers]),stream:!0})}}class W0 extends O{retrieve(w,N,$){let{session_id:V,betas:K}=N;return this._client.get(G`/v1/sessions/${V}/resources/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{session_id:V,betas:K,...W}=N;return this._client.post(G`/v1/sessions/${V}/resources/${w}?beta=true`,{body:W,...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/sessions/${w}/resources?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,N,$){let{session_id:V,betas:K}=N;return this._client.delete(G`/v1/sessions/${V}/resources/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}add(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}/resources?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class Yw extends O{constructor(){super(...arguments);this.events=new K0(this._client),this.resources=new W0(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/sessions?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/sessions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/sessions/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/sessions?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/sessions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/sessions/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}Yw.Events=K0;Yw.Resources=W0;class X0 extends O{create(w,N={},$){let{betas:V,...K}=N??{};return this._client.post(G`/v1/skills/${w}/versions?beta=true`,Qw({body:K,...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])},this._client))}retrieve(w,N,$){let{skill_id:V,betas:K}=N;return this._client.get(G`/v1/skills/${V}/versions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/skills/${w}/versions?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}delete(w,N,$){let{skill_id:V,betas:K}=N;return this._client.delete(G`/v1/skills/${V}/versions/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},$?.headers])})}}class Fw extends O{constructor(){super(...arguments);this.versions=new X0(this._client)}create(w={},N){let{betas:$,...V}=w??{};return this._client.post("/v1/skills?beta=true",Qw({body:V,...N,headers:J([{"anthropic-beta":[...$??[],"skills-2025-10-02"].toString()},N?.headers])},this._client,!1))}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/skills/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/skills?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"skills-2025-10-02"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/skills/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"skills-2025-10-02"].toString()},$?.headers])})}}Fw.Versions=X0;class z0 extends O{create(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/vaults/${w}/credentials?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,N,$){let{vault_id:V,betas:K}=N;return this._client.get(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{vault_id:V,betas:K,...W}=N;return this._client.post(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{body:W,...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w,N={},$){let{betas:V,...K}=N??{};return this._client.getAPIList(G`/v1/vaults/${w}/credentials?beta=true`,q,{query:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,N,$){let{vault_id:V,betas:K}=N;return this._client.delete(G`/v1/vaults/${V}/credentials/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N,$){let{vault_id:V,betas:K}=N;return this._client.post(G`/v1/vaults/${V}/credentials/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}class Sw extends O{constructor(){super(...arguments);this.credentials=new z0(this._client)}create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/vaults?beta=true",{body:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/vaults/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}update(w,N,$){let{betas:V,...K}=N;return this._client.post(G`/v1/vaults/${w}?beta=true`,{body:K,...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/vaults?beta=true",q,{query:V,...N,headers:J([{"anthropic-beta":[...$??[],"managed-agents-2026-04-01"].toString()},N?.headers])})}delete(w,N={},$){let{betas:V}=N??{};return this._client.delete(G`/v1/vaults/${w}?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,N={},$){let{betas:V}=N??{};return this._client.post(G`/v1/vaults/${w}/archive?beta=true`,{...$,headers:J([{"anthropic-beta":[...V??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}}Sw.Credentials=z0;class j extends O{constructor(){super(...arguments);this.models=new pw(this._client),this.messages=new $w(this._client),this.agents=new Gw(this._client),this.environments=new mw(this._client),this.sessions=new Yw(this._client),this.vaults=new Sw(this._client),this.files=new iw(this._client),this.skills=new Fw(this._client)}}j.Models=pw;j.Messages=$w;j.Agents=Gw;j.Environments=mw;j.Sessions=Yw;j.Vaults=Sw;j.Files=iw;j.Skills=Fw;class Aw extends O{create(w,N){let{betas:$,...V}=w;return this._client.post("/v1/complete",{body:V,timeout:this._client._options.timeout??600000,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers]),stream:w.stream??!1})}}function P$(w){return w?.output_config?.format}function SN(w,N,$){let V=P$(N);if(!N||!("parse"in(V??{})))return{...w,content:w.content.map((K)=>{if(K.type==="text")return Object.defineProperty({...K},"parsed_output",{value:null,enumerable:!1});return K}),parsed_output:null};return AN(w,N,$)}function AN(w,N,$){let V=null,K=w.content.map((W)=>{if(W.type==="text"){let X=a1(N,W.text);if(V===null)V=X;return Object.defineProperty({...W},"parsed_output",{value:X,enumerable:!1})}return W});return{...w,content:K,parsed_output:V}}function a1(w,N){let $=P$(w);if($?.type!=="json_schema")return null;try{if("parse"in $)return $.parse(N);return JSON.parse(N)}catch(V){throw new Q(`Failed to parse structured output: ${V}`)}}var u,Vw,qw,Z0,c0,J0,Y0,v0,x0,o,M0,m0,u0,xw,i0,p0,Q0,qN,D$,EN,LN,HN,PN,g$,B$="__json_buf";function I$(w){return w.type==="tool_use"||w.type==="server_tool_use"}class G0{constructor(w,N){u.add(this),this.messages=[],this.receivedMessages=[],Vw.set(this,void 0),qw.set(this,null),this.controller=new AbortController,Z0.set(this,void 0),c0.set(this,()=>{}),J0.set(this,()=>{}),Y0.set(this,void 0),v0.set(this,()=>{}),x0.set(this,()=>{}),o.set(this,{}),M0.set(this,!1),m0.set(this,!1),u0.set(this,!1),xw.set(this,!1),i0.set(this,void 0),p0.set(this,void 0),Q0.set(this,void 0),EN.set(this,($)=>{if(x(this,m0,!0,"f"),l($))$=new B;if($ instanceof B)return x(this,u0,!0,"f"),this._emit("abort",$);if($ instanceof Q)return this._emit("error",$);if($ instanceof Error){let V=new Q($.message);return V.cause=$,this._emit("error",V)}return this._emit("error",new Q(String($)))}),x(this,Z0,new Promise(($,V)=>{x(this,c0,$,"f"),x(this,J0,V,"f")}),"f"),x(this,Y0,new Promise(($,V)=>{x(this,v0,$,"f"),x(this,x0,V,"f")}),"f"),z(this,Z0,"f").catch(()=>{}),z(this,Y0,"f").catch(()=>{}),x(this,qw,w,"f"),x(this,Q0,N?.logger??console,"f")}get response(){return z(this,i0,"f")}get request_id(){return z(this,p0,"f")}async withResponse(){x(this,xw,!0,"f");let w=await z(this,Z0,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let N=new G0(null);return N._run(()=>N._fromReadableStream(w)),N}static createMessage(w,N,$,{logger:V}={}){let K=new G0(N,{logger:V});for(let W of N.messages)K._addMessageParam(W);return x(K,qw,{...N,stream:!0},"f"),K._run(()=>K._createMessage(w,{...N,stream:!0},{...$,headers:{...$?.headers,"X-Stainless-Helper-Method":"stream"}})),K}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},z(this,EN,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,N=!0){if(this.receivedMessages.push(w),N)this._emit("message",w)}async _createMessage(w,N,$){let V=$?.signal,K;if(V){if(V.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),V.addEventListener("abort",K)}try{z(this,u,"m",LN).call(this);let{response:W,data:X}=await w.create({...N,stream:!0},{...$,signal:this.controller.signal}).withResponse();this._connected(W);for await(let Z of X)z(this,u,"m",HN).call(this,Z);if(X.controller.signal?.aborted)throw new B;z(this,u,"m",PN).call(this)}finally{if(V&&K)V.removeEventListener("abort",K)}}_connected(w){if(this.ended)return;x(this,i0,w,"f"),x(this,p0,w?.headers.get("request-id"),"f"),z(this,c0,"f").call(this,w),this._emit("connect")}get ended(){return z(this,M0,"f")}get errored(){return z(this,m0,"f")}get aborted(){return z(this,u0,"f")}abort(){this.controller.abort()}on(w,N){return(z(this,o,"f")[w]||(z(this,o,"f")[w]=[])).push({listener:N}),this}off(w,N){let $=z(this,o,"f")[w];if(!$)return this;let V=$.findIndex((K)=>K.listener===N);if(V>=0)$.splice(V,1);return this}once(w,N){return(z(this,o,"f")[w]||(z(this,o,"f")[w]=[])).push({listener:N,once:!0}),this}emitted(w){return new Promise((N,$)=>{if(x(this,xw,!0,"f"),w!=="error")this.once("error",$);this.once(w,N)})}async done(){x(this,xw,!0,"f"),await z(this,Y0,"f")}get currentMessage(){return z(this,Vw,"f")}async finalMessage(){return await this.done(),z(this,u,"m",qN).call(this)}async finalText(){return await this.done(),z(this,u,"m",D$).call(this)}_emit(w,...N){if(z(this,M0,"f"))return;if(w==="end")x(this,M0,!0,"f"),z(this,v0,"f").call(this);let $=z(this,o,"f")[w];if($)z(this,o,"f")[w]=$.filter((V)=>!V.once),$.forEach(({listener:V})=>V(...N));if(w==="abort"){let V=N[0];if(!z(this,xw,"f")&&!$?.length)Promise.reject(V);z(this,J0,"f").call(this,V),z(this,x0,"f").call(this,V),this._emit("end");return}if(w==="error"){let V=N[0];if(!z(this,xw,"f")&&!$?.length)Promise.reject(V);z(this,J0,"f").call(this,V),z(this,x0,"f").call(this,V),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",z(this,u,"m",qN).call(this))}async _fromReadableStream(w,N){let $=N?.signal,V;if($){if($.aborted)this.controller.abort();V=this.controller.abort.bind(this.controller),$.addEventListener("abort",V)}try{z(this,u,"m",LN).call(this),this._connected(null);let K=b.fromReadableStream(w,this.controller);for await(let W of K)z(this,u,"m",HN).call(this,W);if(K.controller.signal?.aborted)throw new B;z(this,u,"m",PN).call(this)}finally{if($&&V)$.removeEventListener("abort",V)}}[(Vw=new WeakMap,qw=new WeakMap,Z0=new WeakMap,c0=new WeakMap,J0=new WeakMap,Y0=new WeakMap,v0=new WeakMap,x0=new WeakMap,o=new WeakMap,M0=new WeakMap,m0=new WeakMap,u0=new WeakMap,xw=new WeakMap,i0=new WeakMap,p0=new WeakMap,Q0=new WeakMap,EN=new WeakMap,u=new WeakSet,qN=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},D$=function(){if(this.receivedMessages.length===0)throw new Q("stream ended without producing a Message with role=assistant");let N=this.receivedMessages.at(-1).content.filter(($)=>$.type==="text").map(($)=>$.text);if(N.length===0)throw new Q("stream ended without producing a content block with type=text");return N.join(" ")},LN=function(){if(this.ended)return;x(this,Vw,void 0,"f")},HN=function(N){if(this.ended)return;let $=z(this,u,"m",g$).call(this,N);switch(this._emit("streamEvent",N,$),N.type){case"content_block_delta":{let V=$.content.at(-1);switch(N.delta.type){case"text_delta":{if(V.type==="text")this._emit("text",N.delta.text,V.text||"");break}case"citations_delta":{if(V.type==="text")this._emit("citation",N.delta.citation,V.citations??[]);break}case"input_json_delta":{if(I$(V)&&V.input)this._emit("inputJson",N.delta.partial_json,V.input);break}case"thinking_delta":{if(V.type==="thinking")this._emit("thinking",N.delta.thinking,V.thinking);break}case"signature_delta":{if(V.type==="thinking")this._emit("signature",V.signature);break}default:y$(N.delta)}break}case"message_stop":{this._addMessageParam($),this._addMessage(SN($,z(this,qw,"f"),{logger:z(this,Q0,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",$.content.at(-1));break}case"message_start":{x(this,Vw,$,"f");break}case"content_block_start":case"message_delta":break}},PN=function(){if(this.ended)throw new Q("stream has ended, this shouldn't happen");let N=z(this,Vw,"f");if(!N)throw new Q("request ended without sending any chunks");return x(this,Vw,void 0,"f"),SN(N,z(this,qw,"f"),{logger:z(this,Q0,"f")})},g$=function(N){let $=z(this,Vw,"f");if(N.type==="message_start"){if($)throw new Q(`Unexpected event order, got ${N.type} before receiving "message_stop"`);return N.message}if(!$)throw new Q(`Unexpected event order, got ${N.type} before "message_start"`);switch(N.type){case"message_stop":return $;case"message_delta":if($.stop_reason=N.delta.stop_reason,$.stop_sequence=N.delta.stop_sequence,$.usage.output_tokens=N.usage.output_tokens,N.usage.input_tokens!=null)$.usage.input_tokens=N.usage.input_tokens;if(N.usage.cache_creation_input_tokens!=null)$.usage.cache_creation_input_tokens=N.usage.cache_creation_input_tokens;if(N.usage.cache_read_input_tokens!=null)$.usage.cache_read_input_tokens=N.usage.cache_read_input_tokens;if(N.usage.server_tool_use!=null)$.usage.server_tool_use=N.usage.server_tool_use;return $;case"content_block_start":return $.content.push({...N.content_block}),$;case"content_block_delta":{let V=$.content.at(N.index);switch(N.delta.type){case"text_delta":{if(V?.type==="text")$.content[N.index]={...V,text:(V.text||"")+N.delta.text};break}case"citations_delta":{if(V?.type==="text")$.content[N.index]={...V,citations:[...V.citations??[],N.delta.citation]};break}case"input_json_delta":{if(V&&I$(V)){let K=V[B$]||"";K+=N.delta.partial_json;let W={...V};if(Object.defineProperty(W,B$,{value:K,enumerable:!1,writable:!0}),K)W.input=y0(K);$.content[N.index]=W}break}case"thinking_delta":{if(V?.type==="thinking")$.content[N.index]={...V,thinking:V.thinking+N.delta.thinking};break}case"signature_delta":{if(V?.type==="thinking")$.content[N.index]={...V,signature:N.delta.signature};break}default:y$(N.delta)}return $}case"content_block_stop":return $}},Symbol.asyncIterator)](){let w=[],N=[],$=!1;return this.on("streamEvent",(V)=>{let K=N.shift();if(K)K.resolve(V);else w.push(V)}),this.on("end",()=>{$=!0;for(let V of N)V.resolve(void 0);N.length=0}),this.on("abort",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),this.on("error",(V)=>{$=!0;for(let K of N)K.reject(V);N.length=0}),{next:async()=>{if(!w.length){if($)return{value:void 0,done:!0};return new Promise((K,W)=>N.push({resolve:K,reject:W})).then((K)=>K?{value:K,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new b(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function y$(w){}class T0 extends O{create(w,N){return this._client.post("/v1/messages/batches",{body:w,...N})}retrieve(w,N){return this._client.get(G`/v1/messages/batches/${w}`,N)}list(w={},N){return this._client.getAPIList("/v1/messages/batches",i,{query:w,...N})}delete(w,N){return this._client.delete(G`/v1/messages/batches/${w}`,N)}cancel(w,N){return this._client.post(G`/v1/messages/batches/${w}/cancel`,N)}async results(w,N){let $=await this.retrieve(w);if(!$.results_url)throw new Q(`No batch \`results_url\`; Has it finished processing? ${$.processing_status} - ${$.id}`);return this._client.get($.results_url,{...N,headers:J([{Accept:"application/binary"},N?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((V,K)=>fw.fromResponse(K.response,K.controller))}}class Mw extends O{constructor(){super(...arguments);this.batches=new T0(this._client)}create(w,N){if(w.model in j$)console.warn(`The model '${w.model}' is deprecated and will reach end-of-life on ${j$[w.model]}
+Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(w.model in r1&&w.thinking&&w.thinking.type==="enabled")console.warn(`Using Claude with ${w.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let $=this._client._options.timeout;if(!w.stream&&$==null){let K=I0[w.model]??void 0;$=this._client.calculateNonstreamingTimeout(w.max_tokens,K)}let V=B0(w.tools,w.messages);return this._client.post("/v1/messages",{body:w,timeout:$??600000,...N,headers:J([V,N?.headers]),stream:w.stream??!1})}parse(w,N){return this.create(w,N).then(($)=>AN($,w,{logger:this._client.logger??console}))}stream(w,N){return G0.createMessage(this,w,N,{logger:this._client.logger??console})}countTokens(w,N){return this._client.post("/v1/messages/count_tokens",{body:w,...N})}}var j$={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026","claude-3-5-haiku-latest":"February 19th, 2026","claude-3-5-haiku-20241022":"February 19th, 2026","claude-opus-4-0":"June 15th, 2026","claude-opus-4-20250514":"June 15th, 2026","claude-sonnet-4-0":"June 15th, 2026","claude-sonnet-4-20250514":"June 15th, 2026"},r1=["claude-opus-4-6"];Mw.Batches=T0;class Ew extends O{retrieve(w,N={},$){let{betas:V}=N??{};return this._client.get(G`/v1/models/${w}`,{...$,headers:J([{...V?.toString()!=null?{"anthropic-beta":V?.toString()}:void 0},$?.headers])})}list(w={},N){let{betas:$,...V}=w??{};return this._client.getAPIList("/v1/models",i,{query:V,...N,headers:J([{...$?.toString()!=null?{"anthropic-beta":$?.toString()}:void 0},N?.headers])})}}var C0=(w)=>{if(typeof globalThis.process<"u")return globalThis.process.env?.[w]?.trim()||void 0;if(typeof globalThis.Deno<"u")return globalThis.Deno.env?.get?.(w)?.trim()||void 0;return};var DN,gN,l0,k$,h$="\\n\\nHuman:",R$="\\n\\nAssistant:";class A{constructor({baseURL:w=C0("ANTHROPIC_BASE_URL"),apiKey:N=C0("ANTHROPIC_API_KEY")??null,authToken:$=C0("ANTHROPIC_AUTH_TOKEN")??null,...V}={}){DN.add(this),l0.set(this,void 0);let K={apiKey:N,authToken:$,...V,baseURL:w||"https://api.anthropic.com"};if(!K.dangerouslyAllowBrowser&&rN())throw new Q(`It looks like you're running in a browser-like environment.
 
 This is disabled by default, as it risks exposing your secret API credentials to attackers.
 If you understand the risks and have appropriate mitigations in place,
 you can set the \`dangerouslyAllowBrowser\` option to \`true\`, e.g.,
 
 new Anthropic({ apiKey, dangerouslyAllowBrowser: true });
-`);this.baseURL=K.baseURL,this.timeout=K.timeout??gN.DEFAULT_TIMEOUT,this.logger=K.logger??console;let W="warn";this.logLevel=W,this.logLevel=XN(K.logLevel,"ClientOptions.logLevel",this)??XN(C0("ANTHROPIC_LOG"),"process.env['ANTHROPIC_LOG']",this)??W,this.fetchOptions=K.fetchOptions,this.maxRetries=K.maxRetries??2,this.fetch=K.fetch??eN(),x(this,l0,N1,"f"),this._options=K,this.apiKey=typeof N==="string"?N:null,this.authToken=$}withOptions(w){return new this.constructor({...this._options,baseURL:this.baseURL,maxRetries:this.maxRetries,timeout:this.timeout,logger:this.logger,logLevel:this.logLevel,fetch:this.fetch,fetchOptions:this.fetchOptions,apiKey:this.apiKey,authToken:this.authToken,...w})}defaultQuery(){return this._options.defaultQuery}validateHeaders({values:w,nulls:N}){if(w.get("x-api-key")||w.get("authorization"))return;if(this.apiKey&&w.get("x-api-key"))return;if(N.has("x-api-key"))return;if(this.authToken&&w.get("authorization"))return;if(N.has("authorization"))return;throw Error('Could not resolve authentication method. Expected either apiKey or authToken to be set. Or for one of the "X-Api-Key" or "Authorization" headers to be explicitly omitted')}async authHeaders(w){return J([await this.apiKeyAuth(w),await this.bearerAuth(w)])}async apiKeyAuth(w){if(this.apiKey==null)return;return J([{"X-Api-Key":this.apiKey}])}async bearerAuth(w){if(this.authToken==null)return;return J([{Authorization:`Bearer ${this.authToken}`}])}stringifyQuery(w){return $1(w)}getUserAgent(){return`${this.constructor.name}/JS ${s}`}defaultIdempotencyKey(){return`stainless-node-retry-${wN()}`}makeStatusError(w,N,$,V){return L.generate(w,N,$,V)}buildURL(w,N,$){let V=!z(this,DN,"m",k1).call(this)&&$||this.baseURL,K=iN(w)?new URL(w):new URL(V+(V.endsWith("/")&&w.startsWith("/")?w.slice(1):w)),W=this.defaultQuery(),X=Object.fromEntries(K.searchParams);if(!VN(W)||!VN(X))N={...X,...W,...N};if(typeof N==="object"&&N&&!Array.isArray(N))K.search=this.stringifyQuery(N);return K.toString()}_calculateNonstreamingTimeout(w){if(3600*w/128000>600)throw new Q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#streaming-responses for more details");return 600000}async prepareOptions(w){}async prepareRequest(w,{url:N,options:$}){}get(w,N){return this.methodRequest("get",w,N)}post(w,N){return this.methodRequest("post",w,N)}patch(w,N){return this.methodRequest("patch",w,N)}put(w,N){return this.methodRequest("put",w,N)}delete(w,N){return this.methodRequest("delete",w,N)}methodRequest(w,N,$){return this.request(Promise.resolve($).then((V)=>{return{method:w,path:N,...V}}))}request(w,N=null){return new Xw(this,this.makeRequest(w,N,void 0))}async makeRequest(w,N,$){let V=await w,K=V.maxRetries??this.maxRetries;if(N==null)N=K;await this.prepareOptions(V);let{req:W,url:X,timeout:Z}=await this.buildRequest(V,{retryCount:K-N});await this.prepareRequest(W,{url:X,options:V});let Y="log_"+(Math.random()*16777216|0).toString(16).padStart(6,"0"),M=$===void 0?"":`, retryOf: ${$}`,C=Date.now();if(D(this).debug(`[${Y}] sending request`,n({retryOfRequestLogID:$,method:V.method,url:X,options:V,headers:W.headers})),V.signal?.aborted)throw new B;let S=new AbortController,T=await this.fetchWithTimeout(X,W,Z,S).catch(Ew),U=Date.now();if(T instanceof globalThis.Error){let f=`retrying, ${N} attempts remaining`;if(V.signal?.aborted)throw new B;let g=l(T)||/timed? ?out/i.test(String(T)+("cause"in T?String(T.cause):""));if(N)return D(this).info(`[${Y}] connection ${g?"timed out":"failed"} - ${f}`),D(this).debug(`[${Y}] connection ${g?"timed out":"failed"} (${f})`,n({retryOfRequestLogID:$,url:X,durationMs:U-C,message:T.message})),this.retryRequest(V,N,$??Y);if(D(this).info(`[${Y}] connection ${g?"timed out":"failed"} - error; no more retries left`),D(this).debug(`[${Y}] connection ${g?"timed out":"failed"} (error; no more retries left)`,n({retryOfRequestLogID:$,url:X,durationMs:U-C,message:T.message})),g)throw new Pw;throw new Ww({cause:T})}let h=[...T.headers.entries()].filter(([f])=>f==="request-id").map(([f,g])=>", "+f+": "+JSON.stringify(g)).join(""),I=`[${Y}${M}${h}] ${W.method} ${X} ${T.ok?"succeeded":"failed"} with status ${T.status} in ${U-C}ms`;if(!T.ok){let f=await this.shouldRetry(T);if(N&&f){let F0=`retrying, ${N} attempts remaining`;return await w1(T.body),D(this).info(`${I} - ${F0}`),D(this).debug(`[${Y}] response error (${F0})`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,durationMs:U-C})),this.retryRequest(V,N,$??Y,T.headers)}let g=f?"error; no more retries left":"error; not retryable";D(this).info(`${I} - ${g}`);let Kw=await T.text().catch((F0)=>Ew(F0).message),vN=f0(Kw),mN=vN?void 0:Kw;throw D(this).debug(`[${Y}] response error (${g})`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,message:mN,durationMs:Date.now()-C})),this.makeStatusError(T.status,vN,mN,T.headers)}return D(this).info(I),D(this).debug(`[${Y}] response start`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,durationMs:U-C})),{response:T,options:V,controller:S,requestLogID:Y,retryOfRequestLogID:$,startTime:C}}getAPIList(w,N,$){return this.requestAPIList(N,$&&"then"in $?$.then((V)=>({method:"get",path:w,...V})):{method:"get",path:w,...$})}requestAPIList(w,N){let $=this.makeRequest(N,null,void 0);return new P0(this,$,w)}async fetchWithTimeout(w,N,$,V){let{signal:K,method:W,...X}=N||{},Z=this._makeAbort(V);if(K)K.addEventListener("abort",Z,{once:!0});let Y=setTimeout(Z,$),M=globalThis.ReadableStream&&X.body instanceof globalThis.ReadableStream||typeof X.body==="object"&&X.body!==null&&Symbol.asyncIterator in X.body,C={signal:V.signal,...M?{duplex:"half"}:{},method:"GET",...X};if(W)C.method=W.toUpperCase();try{return await this.fetch.call(void 0,w,C)}finally{clearTimeout(Y)}}async shouldRetry(w){let N=w.headers.get("x-should-retry");if(N==="true")return!0;if(N==="false")return!1;if(w.status===408)return!0;if(w.status===409)return!0;if(w.status===429)return!0;if(w.status>=500)return!0;return!1}async retryRequest(w,N,$,V){let K,W=V?.get("retry-after-ms");if(W){let Z=parseFloat(W);if(!Number.isNaN(Z))K=Z}let X=V?.get("retry-after");if(X&&!K){let Z=parseFloat(X);if(!Number.isNaN(Z))K=Z*1000;else K=Date.parse(X)-Date.now()}if(K===void 0){let Z=w.maxRetries??this.maxRetries;K=this.calculateDefaultRetryTimeoutMillis(N,Z)}return await nN(K),this.makeRequest(w,N-1,$)}calculateDefaultRetryTimeoutMillis(w,N){let K=N-w,W=Math.min(0.5*Math.pow(2,K),8),X=1-Math.random()*0.25;return W*X*1000}calculateNonstreamingTimeout(w,N){if(3600000*w/128000>600000||N!=null&&w>N)throw new Q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#long-requests for more details");return 600000}async buildRequest(w,{retryCount:N=0}={}){let $={...w},{method:V,path:K,query:W,defaultBaseURL:X}=$,Z=this.buildURL(K,W,X);if("timeout"in $)lN("timeout",$.timeout);$.timeout=$.timeout??this.timeout;let{bodyHeaders:Y,body:M}=this.buildBody({options:$}),C=await this.buildHeaders({options:w,method:V,bodyHeaders:Y,retryCount:N});return{req:{method:V,headers:C,...$.signal&&{signal:$.signal},...globalThis.ReadableStream&&M instanceof globalThis.ReadableStream&&{duplex:"half"},...M&&{body:M},...this.fetchOptions??{},...$.fetchOptions??{}},url:Z,timeout:$.timeout}}async buildHeaders({options:w,method:N,bodyHeaders:$,retryCount:V}){let K={};if(this.idempotencyHeader&&N!=="get"){if(!w.idempotencyKey)w.idempotencyKey=this.defaultIdempotencyKey();K[this.idempotencyHeader]=w.idempotencyKey}let W=J([K,{Accept:"application/json","User-Agent":this.getUserAgent(),"X-Stainless-Retry-Count":String(V),...w.timeout?{"X-Stainless-Timeout":String(Math.trunc(w.timeout/1000))}:{},...sN(),...this._options.dangerouslyAllowBrowser?{"anthropic-dangerous-direct-browser-access":"true"}:void 0,"anthropic-version":"2023-06-01"},await this.authHeaders(w),this._options.defaultHeaders,$,w.headers]);return this.validateHeaders(W),W.values}_makeAbort(w){return()=>w.abort()}buildBody({options:{body:w,headers:N}}){if(!w)return{bodyHeaders:void 0,body:void 0};let $=J([N]);if(ArrayBuffer.isView(w)||w instanceof ArrayBuffer||w instanceof DataView||typeof w==="string"&&$.values.has("content-type")||globalThis.Blob&&w instanceof globalThis.Blob||w instanceof FormData||w instanceof URLSearchParams||globalThis.ReadableStream&&w instanceof globalThis.ReadableStream)return{bodyHeaders:void 0,body:w};else if(typeof w==="object"&&((Symbol.asyncIterator in w)||(Symbol.iterator in w)&&("next"in w)&&typeof w.next==="function"))return{bodyHeaders:void 0,body:A0(w)};else if(typeof w==="object"&&$.values.get("content-type")==="application/x-www-form-urlencoded")return{bodyHeaders:{"content-type":"application/x-www-form-urlencoded"},body:this.stringifyQuery(w)};else return z(this,l0,"f").call(this,{body:w,headers:$})}}gN=A,l0=new WeakMap,DN=new WeakSet,k1=function(){return this.baseURL!=="https://api.anthropic.com"};A.Anthropic=gN;A.HUMAN_PROMPT=h1;A.AI_PROMPT=R1;A.DEFAULT_TIMEOUT=600000;A.AnthropicError=Q;A.APIError=L;A.APIConnectionError=Ww;A.APIConnectionTimeoutError=Pw;A.APIUserAbortError=B;A.NotFoundError=Iw;A.ConflictError=jw;A.RateLimitError=kw;A.BadRequestError=Dw;A.AuthenticationError=gw;A.InternalServerError=hw;A.PermissionDeniedError=Bw;A.UnprocessableEntityError=yw;A.toFile=D0;class c extends A{constructor(){super(...arguments);this.completions=new Aw(this),this.messages=new Mw(this),this.models=new Hw(this),this.beta=new y(this)}}c.Completions=Aw;c.Messages=Mw;c.Models=Hw;c.Beta=y;var o$=2048,r$=60000,BN="submit_metadata";class P extends Error{kind;retryAfterMs;constructor(w,N,$){super(N);this.kind=w,this.retryAfterMs=$,this.name="ClaudeApiError"}}function _1(w){let N=w?.get?.("retry-after");if(!N)return;let $=Number(N);if(Number.isFinite($)&&$>=0)return Math.round($*1000);return}function s$(w){let N={tags:{type:"string",description:"Comma-separated tags describing the article. Follow the user's tag instructions."},description:{type:"string",description:"Brief summary of the article. Follow the user's description instructions."}},$=["tags","description"];if(w)N.title={type:"string",description:"Concise title for the article. Follow the user's title instructions."},$.push("title");return{name:BN,description:"Submit the generated metadata for the article.",input_schema:{type:"object",properties:N,required:$}}}function e$(w){if(w instanceof c.AuthenticationError)return new P("auth",w.message);if(w instanceof c.RateLimitError)return new P("rate_limit",w.message,_1(w.headers));if(w instanceof c.InternalServerError)return new P("overloaded",w.message,_1(w.headers));if(w instanceof c.APIError)return new P("api",w.message);let N=w instanceof Error?w.message:String(w);return new P("unknown",N)}function w4(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}function N4(w,N){if(!w||typeof w!=="object"||Array.isArray(w))throw new P("api","Tool input was not an object");let $=w;if(typeof $.tags!=="string")throw new P("api","Tool input field 'tags' is missing or not a string");if(typeof $.description!=="string")throw new P("api","Tool input field 'description' is missing or not a string");let V={tags:$.tags,description:$.description};if(N){if(typeof $.title!=="string")throw new P("api","Tool input field 'title' is missing or not a string");V.title=$.title}else if($.title!==void 0){if(typeof $.title!=="string")throw new P("api","Tool input field 'title' is not a string");V.title=$.title}return V}async function b1(w,N,$,V={}){let K=new c({apiKey:$.anthropicApiKey,dangerouslyAllowBrowser:!0}),W=s$($.enableTitle),X;try{X=await K.messages.create({model:$.anthropicModel,max_tokens:o$,system:w,messages:[{role:"user",content:N}],tools:[W],tool_choice:{type:"tool",name:BN}},{timeout:r$,signal:V.signal})}catch(M){if(w4(M))throw M;throw e$(M)}if(!Array.isArray(X.content))throw new P("api","Response had no content blocks");let Z=X.content.filter((M)=>M.type==="tool_use"),Y=Z.find((M)=>M.type==="tool_use"&&M.name===BN);if(!Y||Y.type!=="tool_use"){if(Z.length>0){let M=Z.map((C)=>C.type==="tool_use"?C.name:"").filter((C)=>C!=="").join(", ");throw new P("api",`Model called unexpected tool(s): ${M}`)}throw new P("api","Model did not call the metadata tool")}return N4(Y.input,$.enableTitle)}var v=require("obsidian");async function n0(w,N,$,V,K){let W=!1;return await w.fileManager.processFrontMatter(N,(X)=>{if(K==="append"){let Z=V,Y=X[$],M=Array.isArray(Y)?Y:Y!=null?[String(Y)]:[],C=Array.from(new Set(M.concat(Z)));W=!Array.isArray(Y)||M.length!==C.length||M.some((S,T)=>S!==C[T]),X[$]=C}else if(K==="update"){if(X[$]!==V)W=!0;X[$]=V}else if(X[$]===void 0)X[$]=V,W=!0}),W}function $4(){try{return new RegExp(`[一-龥]|[぀-ヿ]|[가-힯]|[[\\p{Letter}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]][[\\p{Letter}\\p{Mark}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]]*|[.,!?;，。！？；#]|\\n`,"gv")}catch{return/[一-龥]|[぀-ヿ]|[가-힯]|[\p{Letter}\p{Number}][\p{Letter}\p{Mark}\p{Number}]*|[.,!?;，。！？；#]|\n/gu}}var V4=$4(),K4=/^[一-龥぀-ヿ가-힯.,!?;，。！？；#]$/;function O0(w){return w.match(V4)??[]}function r(w){let N="";for(let $=0;$<w.length;$++){let V=w[$];if(V===`
-`)N+=V;else if(K4.test(V))N+=V;else{let K=$>0?w[$-1]:void 0,W=$>0&&K!==`
-`;N+=(W?" ":"")+V}}return N.trim()}function d1(w,N){let $=w.slice(0,N),V=$.length<w.length?"...":"";return`${r($)}${V}`}function c1(w,N){if(N>=w.length)return r(w);let $=Math.max(1,Math.floor(N*0.8)),V=Math.max(0,N-$),K=w.slice(0,$);if(V<=0)return r(K);let W=w.slice(-V);return`${r(K)}
+`);this.baseURL=K.baseURL,this.timeout=K.timeout??gN.DEFAULT_TIMEOUT,this.logger=K.logger??console;let W="warn";this.logLevel=W,this.logLevel=XN(K.logLevel,"ClientOptions.logLevel",this)??XN(C0("ANTHROPIC_LOG"),"process.env['ANTHROPIC_LOG']",this)??W,this.fetchOptions=K.fetchOptions,this.maxRetries=K.maxRetries??2,this.fetch=K.fetch??eN(),x(this,l0,N$,"f"),this._options=K,this.apiKey=typeof N==="string"?N:null,this.authToken=$}withOptions(w){return new this.constructor({...this._options,baseURL:this.baseURL,maxRetries:this.maxRetries,timeout:this.timeout,logger:this.logger,logLevel:this.logLevel,fetch:this.fetch,fetchOptions:this.fetchOptions,apiKey:this.apiKey,authToken:this.authToken,...w})}defaultQuery(){return this._options.defaultQuery}validateHeaders({values:w,nulls:N}){if(w.get("x-api-key")||w.get("authorization"))return;if(this.apiKey&&w.get("x-api-key"))return;if(N.has("x-api-key"))return;if(this.authToken&&w.get("authorization"))return;if(N.has("authorization"))return;throw Error('Could not resolve authentication method. Expected either apiKey or authToken to be set. Or for one of the "X-Api-Key" or "Authorization" headers to be explicitly omitted')}async authHeaders(w){return J([await this.apiKeyAuth(w),await this.bearerAuth(w)])}async apiKeyAuth(w){if(this.apiKey==null)return;return J([{"X-Api-Key":this.apiKey}])}async bearerAuth(w){if(this.authToken==null)return;return J([{Authorization:`Bearer ${this.authToken}`}])}stringifyQuery(w){return $$(w)}getUserAgent(){return`${this.constructor.name}/JS ${s}`}defaultIdempotencyKey(){return`stainless-node-retry-${wN()}`}makeStatusError(w,N,$,V){return L.generate(w,N,$,V)}buildURL(w,N,$){let V=!z(this,DN,"m",k$).call(this)&&$||this.baseURL,K=iN(w)?new URL(w):new URL(V+(V.endsWith("/")&&w.startsWith("/")?w.slice(1):w)),W=this.defaultQuery(),X=Object.fromEntries(K.searchParams);if(!VN(W)||!VN(X))N={...X,...W,...N};if(typeof N==="object"&&N&&!Array.isArray(N))K.search=this.stringifyQuery(N);return K.toString()}_calculateNonstreamingTimeout(w){if(3600*w/128000>600)throw new Q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#streaming-responses for more details");return 600000}async prepareOptions(w){}async prepareRequest(w,{url:N,options:$}){}get(w,N){return this.methodRequest("get",w,N)}post(w,N){return this.methodRequest("post",w,N)}patch(w,N){return this.methodRequest("patch",w,N)}put(w,N){return this.methodRequest("put",w,N)}delete(w,N){return this.methodRequest("delete",w,N)}methodRequest(w,N,$){return this.request(Promise.resolve($).then((V)=>{return{method:w,path:N,...V}}))}request(w,N=null){return new Xw(this,this.makeRequest(w,N,void 0))}async makeRequest(w,N,$){let V=await w,K=V.maxRetries??this.maxRetries;if(N==null)N=K;await this.prepareOptions(V);let{req:W,url:X,timeout:Z}=await this.buildRequest(V,{retryCount:K-N});await this.prepareRequest(W,{url:X,options:V});let Y="log_"+(Math.random()*16777216|0).toString(16).padStart(6,"0"),M=$===void 0?"":`, retryOf: ${$}`,C=Date.now();if(D(this).debug(`[${Y}] sending request`,n({retryOfRequestLogID:$,method:V.method,url:X,options:V,headers:W.headers})),V.signal?.aborted)throw new B;let F=new AbortController,T=await this.fetchWithTimeout(X,W,Z,F).catch(Hw),U=Date.now();if(T instanceof globalThis.Error){let S=`retrying, ${N} attempts remaining`;if(V.signal?.aborted)throw new B;let g=l(T)||/timed? ?out/i.test(String(T)+("cause"in T?String(T.cause):""));if(N)return D(this).info(`[${Y}] connection ${g?"timed out":"failed"} - ${S}`),D(this).debug(`[${Y}] connection ${g?"timed out":"failed"} (${S})`,n({retryOfRequestLogID:$,url:X,durationMs:U-C,message:T.message})),this.retryRequest(V,N,$??Y);if(D(this).info(`[${Y}] connection ${g?"timed out":"failed"} - error; no more retries left`),D(this).debug(`[${Y}] connection ${g?"timed out":"failed"} (error; no more retries left)`,n({retryOfRequestLogID:$,url:X,durationMs:U-C,message:T.message})),g)throw new Pw;throw new Ww({cause:T})}let h=[...T.headers.entries()].filter(([S])=>S==="request-id").map(([S,g])=>", "+S+": "+JSON.stringify(g)).join(""),I=`[${Y}${M}${h}] ${W.method} ${X} ${T.ok?"succeeded":"failed"} with status ${T.status} in ${U-C}ms`;if(!T.ok){let S=await this.shouldRetry(T);if(N&&S){let f0=`retrying, ${N} attempts remaining`;return await w$(T.body),D(this).info(`${I} - ${f0}`),D(this).debug(`[${Y}] response error (${f0})`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,durationMs:U-C})),this.retryRequest(V,N,$??Y,T.headers)}let g=S?"error; no more retries left":"error; not retryable";D(this).info(`${I} - ${g}`);let Kw=await T.text().catch((f0)=>Hw(f0).message),cN=S0(Kw),vN=cN?void 0:Kw;throw D(this).debug(`[${Y}] response error (${g})`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,message:vN,durationMs:Date.now()-C})),this.makeStatusError(T.status,cN,vN,T.headers)}return D(this).info(I),D(this).debug(`[${Y}] response start`,n({retryOfRequestLogID:$,url:T.url,status:T.status,headers:T.headers,durationMs:U-C})),{response:T,options:V,controller:F,requestLogID:Y,retryOfRequestLogID:$,startTime:C}}getAPIList(w,N,$){return this.requestAPIList(N,$&&"then"in $?$.then((V)=>({method:"get",path:w,...V})):{method:"get",path:w,...$})}requestAPIList(w,N){let $=this.makeRequest(N,null,void 0);return new P0(this,$,w)}async fetchWithTimeout(w,N,$,V){let{signal:K,method:W,...X}=N||{},Z=this._makeAbort(V);if(K)K.addEventListener("abort",Z,{once:!0});let Y=setTimeout(Z,$),M=globalThis.ReadableStream&&X.body instanceof globalThis.ReadableStream||typeof X.body==="object"&&X.body!==null&&Symbol.asyncIterator in X.body,C={signal:V.signal,...M?{duplex:"half"}:{},method:"GET",...X};if(W)C.method=W.toUpperCase();try{return await this.fetch.call(void 0,w,C)}finally{clearTimeout(Y)}}async shouldRetry(w){let N=w.headers.get("x-should-retry");if(N==="true")return!0;if(N==="false")return!1;if(w.status===408)return!0;if(w.status===409)return!0;if(w.status===429)return!0;if(w.status>=500)return!0;return!1}async retryRequest(w,N,$,V){let K,W=V?.get("retry-after-ms");if(W){let Z=parseFloat(W);if(!Number.isNaN(Z))K=Z}let X=V?.get("retry-after");if(X&&!K){let Z=parseFloat(X);if(!Number.isNaN(Z))K=Z*1000;else K=Date.parse(X)-Date.now()}if(K===void 0){let Z=w.maxRetries??this.maxRetries;K=this.calculateDefaultRetryTimeoutMillis(N,Z)}return await nN(K),this.makeRequest(w,N-1,$)}calculateDefaultRetryTimeoutMillis(w,N){let K=N-w,W=Math.min(0.5*Math.pow(2,K),8),X=1-Math.random()*0.25;return W*X*1000}calculateNonstreamingTimeout(w,N){if(3600000*w/128000>600000||N!=null&&w>N)throw new Q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#long-requests for more details");return 600000}async buildRequest(w,{retryCount:N=0}={}){let $={...w},{method:V,path:K,query:W,defaultBaseURL:X}=$,Z=this.buildURL(K,W,X);if("timeout"in $)lN("timeout",$.timeout);$.timeout=$.timeout??this.timeout;let{bodyHeaders:Y,body:M}=this.buildBody({options:$}),C=await this.buildHeaders({options:w,method:V,bodyHeaders:Y,retryCount:N});return{req:{method:V,headers:C,...$.signal&&{signal:$.signal},...globalThis.ReadableStream&&M instanceof globalThis.ReadableStream&&{duplex:"half"},...M&&{body:M},...this.fetchOptions??{},...$.fetchOptions??{}},url:Z,timeout:$.timeout}}async buildHeaders({options:w,method:N,bodyHeaders:$,retryCount:V}){let K={};if(this.idempotencyHeader&&N!=="get"){if(!w.idempotencyKey)w.idempotencyKey=this.defaultIdempotencyKey();K[this.idempotencyHeader]=w.idempotencyKey}let W=J([K,{Accept:"application/json","User-Agent":this.getUserAgent(),"X-Stainless-Retry-Count":String(V),...w.timeout?{"X-Stainless-Timeout":String(Math.trunc(w.timeout/1000))}:{},...sN(),...this._options.dangerouslyAllowBrowser?{"anthropic-dangerous-direct-browser-access":"true"}:void 0,"anthropic-version":"2023-06-01"},await this.authHeaders(w),this._options.defaultHeaders,$,w.headers]);return this.validateHeaders(W),W.values}_makeAbort(w){return()=>w.abort()}buildBody({options:{body:w,headers:N}}){if(!w)return{bodyHeaders:void 0,body:void 0};let $=J([N]);if(ArrayBuffer.isView(w)||w instanceof ArrayBuffer||w instanceof DataView||typeof w==="string"&&$.values.has("content-type")||globalThis.Blob&&w instanceof globalThis.Blob||w instanceof FormData||w instanceof URLSearchParams||globalThis.ReadableStream&&w instanceof globalThis.ReadableStream)return{bodyHeaders:void 0,body:w};else if(typeof w==="object"&&((Symbol.asyncIterator in w)||(Symbol.iterator in w)&&("next"in w)&&typeof w.next==="function"))return{bodyHeaders:void 0,body:A0(w)};else if(typeof w==="object"&&$.values.get("content-type")==="application/x-www-form-urlencoded")return{bodyHeaders:{"content-type":"application/x-www-form-urlencoded"},body:this.stringifyQuery(w)};else return z(this,l0,"f").call(this,{body:w,headers:$})}}gN=A,l0=new WeakMap,DN=new WeakSet,k$=function(){return this.baseURL!=="https://api.anthropic.com"};A.Anthropic=gN;A.HUMAN_PROMPT=h$;A.AI_PROMPT=R$;A.DEFAULT_TIMEOUT=600000;A.AnthropicError=Q;A.APIError=L;A.APIConnectionError=Ww;A.APIConnectionTimeoutError=Pw;A.APIUserAbortError=B;A.NotFoundError=Iw;A.ConflictError=yw;A.RateLimitError=kw;A.BadRequestError=Dw;A.AuthenticationError=gw;A.InternalServerError=hw;A.PermissionDeniedError=Bw;A.UnprocessableEntityError=jw;A.toFile=D0;class c extends A{constructor(){super(...arguments);this.completions=new Aw(this),this.messages=new Mw(this),this.models=new Ew(this),this.beta=new j(this)}}c.Completions=Aw;c.Messages=Mw;c.Models=Ew;c.Beta=j;var e1=2048,w2=60000,BN="submit_metadata";class P extends Error{kind;retryAfterMs;constructor(w,N,$){super(N);this.kind=w,this.retryAfterMs=$,this.name="ClaudeApiError"}}function _$(w){let N=w?.get?.("retry-after");if(!N)return;let $=Number(N);if(Number.isFinite($)&&$>=0)return Math.round($*1000);return}function N2(w){let N={tags:{type:"string",description:"Comma-separated tags describing the article. Follow the user's tag instructions."},description:{type:"string",description:"Brief summary of the article. Follow the user's description instructions."}},$=["tags","description"];if(w)N.title={type:"string",description:"Concise title for the article. Follow the user's title instructions."},$.push("title");return{name:BN,description:"Submit the generated metadata for the article.",input_schema:{type:"object",properties:N,required:$}}}function $2(w){if(w instanceof c.AuthenticationError)return new P("auth",w.message);if(w instanceof c.RateLimitError)return new P("rate_limit",w.message,_$(w.headers));if(w instanceof c.InternalServerError)return new P("overloaded",w.message,_$(w.headers));if(w instanceof c.APIError)return new P("api",w.message);let N=w instanceof Error?w.message:String(w);return new P("unknown",N)}function V2(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}function K2(w,N){if(!w||typeof w!=="object"||Array.isArray(w))throw new P("api","Tool input was not an object");let $=w;if(typeof $.tags!=="string")throw new P("api","Tool input field 'tags' is missing or not a string");if(typeof $.description!=="string")throw new P("api","Tool input field 'description' is missing or not a string");let V={tags:$.tags,description:$.description};if(N){if(typeof $.title!=="string")throw new P("api","Tool input field 'title' is missing or not a string");V.title=$.title}else if($.title!==void 0){if(typeof $.title!=="string")throw new P("api","Tool input field 'title' is not a string");V.title=$.title}return V}async function b$(w,N,$,V={}){let K=new c({apiKey:$.anthropicApiKey,dangerouslyAllowBrowser:!0}),W=N2($.enableTitle),X;try{X=await K.messages.create({model:$.anthropicModel,max_tokens:e1,system:w,messages:[{role:"user",content:N}],tools:[W],tool_choice:{type:"tool",name:BN}},{timeout:w2,signal:V.signal})}catch(M){if(V2(M))throw M;throw $2(M)}if(!Array.isArray(X.content))throw new P("api","Response had no content blocks");let Z=X.content.filter((M)=>M.type==="tool_use"),Y=Z.find((M)=>M.type==="tool_use"&&M.name===BN);if(!Y||Y.type!=="tool_use"){if(Z.length>0){let M=Z.map((C)=>C.type==="tool_use"?C.name:"").filter((C)=>C!=="").join(", ");throw new P("api",`Model called unexpected tool(s): ${M}`)}throw new P("api","Model did not call the metadata tool")}return K2(Y.input,$.enableTitle)}var v=require("obsidian");async function n0(w,N,$,V,K){let W=!1;return await w.fileManager.processFrontMatter(N,(X)=>{if(K==="append"){let Z=V,Y=X[$],M=Array.isArray(Y)?Y:Y!=null?[String(Y)]:[],C=Array.from(new Set(M.concat(Z)));W=!Array.isArray(Y)||M.length!==C.length||M.some((F,T)=>F!==C[T]),X[$]=C}else if(K==="update"){if(X[$]!==V)W=!0;X[$]=V}else if(X[$]===void 0)X[$]=V,W=!0}),W}function W2(){try{return new RegExp(`[一-龥]|[぀-ヿ]|[가-힯]|[[\\p{Letter}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]][[\\p{Letter}\\p{Mark}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]]*|[.,!?;，。！？；#]|\\n`,"gv")}catch{return/[一-龥]|[぀-ヿ]|[가-힯]|[\p{Letter}\p{Number}][\p{Letter}\p{Mark}\p{Number}]*|[.,!?;，。！？；#]|\n/gu}}var X2=W2(),z2=/^[一-龥぀-ヿ가-힯.,!?;，。！？；#]$/;function O0(w){return w.match(X2)??[]}function r(w){let N="";for(let $=0;$<w.length;$++){let V=w[$];if(V===`
+`)N+=V;else if(z2.test(V))N+=V;else{let K=$>0?w[$-1]:void 0,W=$>0&&K!==`
+`;N+=(W?" ":"")+V}}return N.trim()}function d$(w,N){let $=w.slice(0,N),V=$.length<w.length?"...":"";return`${r($)}${V}`}function c$(w,N){if(N>=w.length)return r(w);let $=Math.max(1,Math.floor(N*0.8)),V=Math.max(0,N-$),K=w.slice(0,$);if(V<=0)return r(K);let W=w.slice(-V);return`${r(K)}
 ...
-${r(W)}`}function v1(w,N,$){let V=w.split(`
-`),K=[],W=!1,X=0,Z=0;for(let U of V){let h=O0(U),I=X+h.length+1;if(U.startsWith("#"))K.push(U),W=!0,Z=I;else if(W&&U.trim()!==""){let f=h.slice(0,30),g=f.length<h.length?"...":"";K.push(`${r(f)}${g}`),W=!1,Z=I}X=I}let Y=K.join(`
-`),M=O0(Y);if(M.length>$)return r(M.slice(0,$));let C=$-M.length,S=N.slice(Z,Z+C),T=r(S);if(T!==""){let U=Z+C<N.length?"...":"";return`Outline: 
+${r(W)}`}function v$(w,N,$){let V=w.split(`
+`),K=[],W=!1,X=0,Z=0;for(let U of V){let h=O0(U),I=X+h.length+1;if(U.startsWith("#"))K.push(U),W=!0,Z=I;else if(W&&U.trim()!==""){let S=h.slice(0,30),g=S.length<h.length?"...":"";K.push(`${r(S)}${g}`),W=!1,Z=I}X=I}let Y=K.join(`
+`),M=O0(Y);if(M.length>$)return r(M.slice(0,$));let C=$-M.length,F=N.slice(Z,Z+C),T=r(F);if(T!==""){let U=Z+C<N.length?"...":"";return`Outline: 
 ${Y}
 
 Body: ${T}${U}`}return`Outline: 
-${Y}`}async function IN(w,N,$=1000,V="head_only"){let K=await w.vault.read(N);if(K.length===0)return"";if($<=0)return K;let W=O0(K);if(W.length>$){if(V==="head_tail")K=c1(W,$);else if(V==="head_only")K=d1(W,$);else if(V==="heading")K=v1(K,W,$)}return K}function W4(w){if(w instanceof P)switch(w.kind){case"auth":new v.Notice("Authentication failed. Please check your API key in Settings → Metadator",8000);return;case"rate_limit":new v.Notice("Rate limit exceeded. Please wait a moment and try again.",8000);return;case"overloaded":new v.Notice("API is currently overloaded. Please try again in a moment.",8000);return;case"api":new v.Notice(`API error: ${w.message}`,8000);return;case"unknown":new v.Notice(`Unexpected error: ${w.message}`,8000);return}new v.Notice(`Unexpected error: ${w instanceof Error?w.message:String(w)}`,8000)}function X4(w,N){let $=["Generate metadata for the provided article and submit it via the submit_metadata tool. Field requirements:","",`1. Tags: ${N.tagsPrompt}`,"",`2. Description: ${N.descriptionPrompt}`];if(N.enableTitle)$.push("",`3. Title: ${N.titlePrompt}`);let V=`<article>
+${Y}`}async function IN(w,N,$=1000,V="head_only"){let K=await w.vault.read(N);if(K.length===0)return"";if($<=0)return K;let W=O0(K);if(W.length>$){if(V==="head_tail")K=c$(W,$);else if(V==="head_only")K=d$(W,$);else if(V==="heading")K=v$(K,W,$)}return K}function Z2(w){if(w instanceof P)switch(w.kind){case"auth":new v.Notice("Authentication failed. Please check your API key in Settings → Metadator",8000);return;case"rate_limit":new v.Notice("Rate limit exceeded. Please wait a moment and try again.",8000);return;case"overloaded":new v.Notice("API is currently overloaded. Please try again in a moment.",8000);return;case"api":new v.Notice(`API error: ${w.message}`,8000);return;case"unknown":new v.Notice(`Unexpected error: ${w.message}`,8000);return}new v.Notice(`Unexpected error: ${w instanceof Error?w.message:String(w)}`,8000)}function J2(w,N){let $=["Generate metadata for the provided article and submit it via the submit_metadata tool. Field requirements:","",`1. Tags: ${N.tagsPrompt}`,"",`2. Description: ${N.descriptionPrompt}`];if(N.enableTitle)$.push("",`3. Title: ${N.titlePrompt}`);let V=`<article>
 ${w}
 </article>`;return{system:$.join(`
-`),userMessage:V}}function z4(w){return w.split(",").map((N)=>N.trim()).filter((N)=>N!=="")}function Z4(w){let N=w.trim();if(N.startsWith('"')&&N.endsWith('"')||N.startsWith("'")&&N.endsWith("'"))return N.substring(1,N.length-1);return N}function t0(w){if(!w)return!0;if(typeof w==="string")return w.trim()==="";if(Array.isArray(w))return w.length===0||w.every((N)=>String(N).trim()==="");return!1}function J4(w){return w.updateMethod==="always_regenerate"?"update_all":"only_empty"}function Y4(w,N){if(w==="update_all")return"update";return t0(N)?"update":"keep"}function jN(w,N){if(N.updateMethod==="always_regenerate")return!0;return t0(w[N.tagsFieldName])||t0(w[N.descriptionFieldName])||N.enableTitle&&t0(w[N.titleFieldName])}function x4(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}async function yN(w,N,$,V={}){if(N.extension!=="md")return{kind:"skipped",file:N,reason:"not a markdown file"};if(!$.anthropicApiKey)return{kind:"skipped",file:N,reason:"missing API key"};let W=w.metadataCache.getFileCache(N)?.frontmatter||{};if(!jN(W,$))return{kind:"skipped",file:N,reason:"all fields already populated"};if(V.signal?.aborted)return{kind:"skipped",file:N,reason:"cancelled before request"};try{return await M4(w,N,$,W,J4($),V.presentation??"interactive",V.signal)?{kind:"changed",file:N}:{kind:"skipped",file:N,reason:"no changes"}}catch(X){if(V.signal?.aborted||x4(X))return{kind:"skipped",file:N,reason:"cancelled"};return{kind:"error",file:N,reason:X instanceof Error?X.message:String(X),error:X}}}async function m1(w,N,$={}){let V=w.workspace.getActiveFile();if(!V){new v.Notice("Please open a file first");return}if(V.extension!=="md"){new v.Notice("Current file is not a markdown file");return}if(!N.anthropicApiKey){new v.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let K=await yN(w,V,N,{signal:$.signal});if(K.kind==="changed")new v.Notice("Metadata updated successfully");else if(K.kind==="error")W4(K.error),console.error("generateMetadata error:",K.error)}async function M4(w,N,$,V,K,W,X){let Z=W==="bulk",Y=$.truncateContent?await IN(w,N,$.contentTokenLimit,$.truncateMethod):await IN(w,N,-1,"head_only"),{system:M,userMessage:C}=X4(Y,$);if($.debugLogging&&!Z)console.log("[Metadator] System:",M),console.log("[Metadator] User message:",C);let S=Z?void 0:new v.Notice("Generating metadata...",0),T;try{T=await b1(M,C,$,{signal:X})}finally{S?.hide()}if($.debugLogging)if(Z)console.log(`[Metadator] [bulk] ${N.path} — fields received`);else console.log("[Metadator] Fields:",T);if(X?.aborted)return!1;let U=!1;async function h(f,g){try{if(g==="keep")return await n0(w,N,f.fieldName,f.value,"keep");if(f.updateMethod==="append")return await n0(w,N,f.fieldName,f.value,"append");return await n0(w,N,f.fieldName,f.value,"update")}catch(Kw){if(!Z)new v.Notice(`Failed to write ${f.fieldName}: ${Kw instanceof Error?Kw.message:String(Kw)}`);return console.error(`updateFrontMatter error (${f.fieldName}):`,Kw),!1}}let I=[];if(T.tags)I.push({fieldName:$.tagsFieldName,value:z4(T.tags),updateMethod:"append"});if(T.description)I.push({fieldName:$.descriptionFieldName,value:T.description,updateMethod:"update"});if($.enableTitle&&T.title)I.push({fieldName:$.titleFieldName,value:Z4(T.title),updateMethod:"update"});for(let f of I){if(X?.aborted)return U;let g=Y4(K,V[f.fieldName]);if(await h(f,g))U=!0}return U}var Q4=[2000,8000,30000];function kN(w){let N=[];for(let $ of w.children)if("children"in $)N.push(...kN($));else{let V=$;if(V.extension==="md")N.push(V)}return N}function u1(w,N,$){let V=[],K=[];for(let W of N){let X=w.metadataCache.getFileCache(W)?.frontmatter??{};if(jN(X,$))V.push(W);else K.push(W)}return{willChange:V,willSkip:K}}function G4(w){return w instanceof P&&(w.kind==="rate_limit"||w.kind==="overloaded")}var T4=2;function C4(w,N,$=Math.random){if(N instanceof P&&N.retryAfterMs!==void 0&&Number.isFinite(N.retryAfterMs))return Math.min(N.retryAfterMs,w*T4);return Math.round(w*(0.5+$()))}var O4=100;async function U4(w,N){if(w<=0)return N?.()??!1;let $=Date.now()+w;while(Date.now()<$){if(N?.())return!0;let V=$-Date.now();await new Promise((K)=>setTimeout(K,Math.min(O4,V)))}return N?.()??!1}async function F4(w,N,$,V,K,W,X=Math.random){let Z=()=>(K?.()??!1)||(W?.aborted??!1);for(let Y=0;Y<=V.length;Y++){if(Z())return{kind:"skipped",file:N,reason:"cancelled before attempt"};let M=await yN(w,N,$,{presentation:"bulk",signal:W});if(M.kind!=="error"||!G4(M.error))return M;if(Y===V.length)return M;let C=C4(V[Y],M.error,X);if(await U4(C,Z))return{kind:"skipped",file:N,reason:"cancelled during retry backoff"}}return{kind:"skipped",file:N,reason:"retry loop exited unexpectedly"}}async function i1(w,N,$,{onProgress:V,shouldAbort:K,retryDelaysMs:W,signal:X,random:Z}={}){let Y=W??Q4,M=[],C=0;for(let S=0;S<N.length;S++){if(K?.()||X?.aborted)break;let T=N[S];V?.({current:S+1,total:N.length,file:T,errors:C});let U=await F4(w,T,$,Y,K,X,Z);if(M.push(U),U.kind==="error")C++}return M}var a0=require("obsidian"),S4=100;class hN extends a0.Modal{resolver;resolved=!1;info;constructor(w,N){super(w);this.info=N}openAndAwait(){return new Promise((w)=>{this.resolver=w,this.open()})}onOpen(){let{contentEl:w}=this,{folderPath:N,total:$,willChange:V,willSkip:K,settings:W}=this.info;w.createEl("h2",{text:"Generate metadata for folder"}),w.createEl("p",{text:`Folder: ${N}`}),w.createEl("p",{text:`${$} notes — ${V} will change, ${K} will skip`});let X=W.truncateContent?W.truncateMethod:"disabled";if(w.createEl("p",{text:`Model: ${W.anthropicModel} · Update: ${W.updateMethod} · Truncation: ${X}`}),$>S4){let C=w.createEl("p",{text:`⚠ Large batch — this will make up to ${V} API calls`});C.style.color="var(--text-warning)",C.style.fontWeight="bold"}let Z=w.createDiv({cls:"modal-button-container"});Z.createEl("button",{text:"Cancel"}).addEventListener("click",()=>{this.resolve(!1),this.close()}),Z.createEl("button",{text:`Generate (${V})`,cls:"mod-cta"}).addEventListener("click",()=>{this.resolve(!0),this.close()})}onClose(){this.resolve(!1),this.contentEl.empty()}resolve(w){if(this.resolved)return;this.resolved=!0,this.resolver?.(w)}}class RN extends a0.Modal{aborted=!1;finishing=!1;statusEl;cancelBtn;onAbort;onOpen(){let{contentEl:w}=this;w.createEl("h2",{text:"Generating metadata"}),this.statusEl=w.createEl("p",{text:"Starting…"});let N=w.createDiv({cls:"modal-button-container"});this.cancelBtn=N.createEl("button",{text:"Cancel"}),this.cancelBtn.addEventListener("click",()=>{if(this.aborted=!0,this.onAbort?.(),this.cancelBtn)this.cancelBtn.disabled=!0,this.cancelBtn.textContent="Cancelling…"})}setAbortHandler(w){this.onAbort=w}setProgress(w){if(!this.statusEl)return;this.statusEl.textContent=`${w.current} / ${w.total} · ${w.file.path} · errors: ${w.errors}`}isAborted(){return this.aborted}finish(){this.finishing=!0,this.close()}onClose(){if(!this.finishing)this.aborted=!0,this.onAbort?.();this.contentEl.empty()}}class _N extends a0.Modal{results;aborted;totalPlanned;constructor(w,N,$,V){super(w);this.results=N,this.aborted=$,this.totalPlanned=V}onOpen(){let{contentEl:w}=this,N=this.results.filter((Y)=>Y.kind==="changed").length,$=this.results.filter((Y)=>Y.kind==="skipped").length,V=this.results.filter((Y)=>Y.kind==="error"),K=this.totalPlanned-this.results.length;w.createEl("h2",{text:this.aborted?"Cancelled":"Finished"});let W=w.createEl("ul");if(W.createEl("li",{text:`${N} changed`}),W.createEl("li",{text:`${$} skipped`}),W.createEl("li",{text:`${V.length} errored`}),K>0)W.createEl("li",{text:`${K} not processed (cancelled)`});if(V.length>0){w.createEl("h3",{text:"Errors"});let Y=w.createEl("ul");for(let M of V)if(M.kind==="error")Y.createEl("li",{text:`${M.file.path}: ${M.reason}`})}w.createDiv({cls:"modal-button-container"}).createEl("button",{text:"Close",cls:"mod-cta"}).addEventListener("click",()=>this.close())}onClose(){this.contentEl.empty()}}async function p1(w,N,$,V={}){if(!$.anthropicApiKey){new o0.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let K=kN(N);if(K.length===0){new o0.Notice("No markdown files found in folder");return}let{willChange:W,willSkip:X}=u1(w,K,$);if(W.length===0){new o0.Notice(`All ${K.length} notes already have metadata; nothing to do`);return}if(!await new hN(w,{folderPath:N.path,total:K.length,willChange:W.length,willSkip:X.length,settings:$}).openAndAwait())return;let Y=new RN(w),M=new AbortController;if(V.signal)if(V.signal.aborted)M.abort(V.signal.reason);else V.signal.addEventListener("abort",()=>M.abort(V.signal?.reason),{once:!0});Y.setAbortHandler(()=>M.abort("cancelled_by_user")),Y.open();let C=await i1(w,W,$,{onProgress:(T)=>Y.setProgress(T),shouldAbort:()=>(V.shouldAbort?.()??!1)||Y.isAborted(),signal:M.signal}),S=Y.isAborted()||(V.shouldAbort?.()??!1)||M.signal.aborted;Y.finish(),new _N(w,C,S,W.length).open()}var k=1000,r0=["claude-sonnet-4-6","claude-opus-4-6","claude-haiku-4-5-20251001"],l1={"claude-sonnet-4-6":"Claude Sonnet 4.6","claude-opus-4-6":"Claude Opus 4.6","claude-haiku-4-5-20251001":"Claude Haiku 4.5"},Lw=["head_only","head_tail","heading"],n1={head_only:"Beginning Only",head_tail:"Beginning + End",heading:"Headings + Summaries"},U0=["always_regenerate","preserve_existing"],t1={always_regenerate:"Always Regenerate",preserve_existing:"Preserve Existing"},H={anthropicApiKey:"",anthropicModel:"claude-sonnet-4-6",tagsFieldName:"tags",descriptionFieldName:"description",titleFieldName:"title",enableTitle:!0,debugLogging:!1,truncateContent:!0,contentTokenLimit:1000,truncateMethod:"head_only",updateMethod:"preserve_existing",tagsPrompt:"Select 3-5 relevant tags in lowercase with hyphens instead of spaces (e.g., 'knowledge-management', 'note-taking')",descriptionPrompt:"Write a concise but useful summary in 1-2 sentences that captures the main purpose and key points",titlePrompt:"Create a simple, concise title with minimal adjectives that clearly states the topic"};var F=require("obsidian");class bN extends F.PluginSettingTab{plugin;constructor(w,N){super(w,N);this.plugin=N}display(){let{containerEl:w}=this;w.empty(),new F.Setting(w).setName("Anthropic API Settings").setHeading(),w.createEl("p",{text:"Note: When you run the metadata command, your note content is sent to the Anthropic API for processing. No data is stored by Anthropic beyond the API request.",cls:"setting-item-description"}),new F.Setting(w).setName("API Key").setDesc("Your Anthropic API key. Get one at console.anthropic.com (requires an account with billing enabled)").addText((W)=>{W.setPlaceholder("sk-ant-...").setValue(this.plugin.settings.anthropicApiKey).onChange(async(X)=>{this.plugin.settings.anthropicApiKey=X,await this.plugin.saveSettings()}),W.inputEl.type="password"}),new F.Setting(w).setName("Model").setDesc("Model to use for metadata generation").addDropdown((W)=>{for(let X of r0)W.addOption(X,l1[X]);W.setValue(this.plugin.settings.anthropicModel).onChange(async(X)=>{this.plugin.settings.anthropicModel=X,await this.plugin.saveSettings()})}),new F.Setting(w).setName("Debug Logging").setDesc("Log prompts and responses to the developer console (View → Toggle Developer Tools)").addToggle((W)=>W.setValue(this.plugin.settings.debugLogging).onChange(async(X)=>{this.plugin.settings.debugLogging=X,await this.plugin.saveSettings()})),new F.Setting(w).setName("Update Settings").setHeading(),new F.Setting(w).setName("Update Method").setDesc("Always Regenerate: regenerate on every command; Preserve Existing: only generate empty fields").addDropdown((W)=>{for(let X of U0)W.addOption(X,t1[X]);W.setValue(this.plugin.settings.updateMethod).onChange(async(X)=>{this.plugin.settings.updateMethod=X,await this.plugin.saveSettings()})}),new F.Setting(w).setName("Truncate Content").setDesc("Limit content sent to API to reduce costs").addToggle((W)=>W.setValue(this.plugin.settings.truncateContent).onChange(async(X)=>{this.plugin.settings.truncateContent=X,await this.plugin.saveSettings(),N.setDisabled(!X),$.setDisabled(!X)}));let N=new F.Setting(w).setName("Content Token Limit").setDesc("Maximum number of tokens of note content sent to the API").addText((W)=>W.setValue(this.plugin.settings.contentTokenLimit.toString()).onChange(async(X)=>{let Z=Number.parseInt(X,10);if(Number.isNaN(Z)||Z<1){new F.Notice("Content token limit must be a positive integer"),W.setValue(this.plugin.settings.contentTokenLimit.toString());return}this.plugin.settings.contentTokenLimit=Z,await this.plugin.saveSettings()})),$=new F.Setting(w).setName("Truncate Method").setDesc("How to truncate long content").addDropdown((W)=>{for(let X of Lw)W.addOption(X,n1[X]);W.setValue(this.plugin.settings.truncateMethod).onChange(async(X)=>{this.plugin.settings.truncateMethod=X,await this.plugin.saveSettings()})});N.setDisabled(!this.plugin.settings.truncateContent),$.setDisabled(!this.plugin.settings.truncateContent),new F.Setting(w).setName("Tags Settings").setHeading(),new F.Setting(w).setName("Tags Field Name").setDesc("Frontmatter field name for tags").addText((W)=>W.setValue(this.plugin.settings.tagsFieldName).onChange(async(X)=>{this.plugin.settings.tagsFieldName=X||"tags",await this.plugin.saveSettings()})),new F.Setting(w).setName("Tags Prompt").setDesc(`Instructions for tag generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.tagsPrompt).onChange(async(X)=>{if(X.length>k){new F.Notice(`Tags prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.tagsPrompt);return}this.plugin.settings.tagsPrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")}),new F.Setting(w).setName("Description Settings").setHeading(),new F.Setting(w).setName("Description Field Name").setDesc("Frontmatter field name for description").addText((W)=>W.setValue(this.plugin.settings.descriptionFieldName).onChange(async(X)=>{this.plugin.settings.descriptionFieldName=X||"description",await this.plugin.saveSettings()})),new F.Setting(w).setName("Description Prompt").setDesc(`Instructions for description generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.descriptionPrompt).onChange(async(X)=>{if(X.length>k){new F.Notice(`Description prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.descriptionPrompt);return}this.plugin.settings.descriptionPrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")}),new F.Setting(w).setName("Title Settings").setHeading(),new F.Setting(w).setName("Enable Title").setDesc("Generate title metadata").addToggle((W)=>W.setValue(this.plugin.settings.enableTitle).onChange(async(X)=>{this.plugin.settings.enableTitle=X,await this.plugin.saveSettings(),V.setDisabled(!X),K.setDisabled(!X)}));let V=new F.Setting(w).setName("Title Field Name").setDesc("Frontmatter field name for title").addText((W)=>W.setValue(this.plugin.settings.titleFieldName).onChange(async(X)=>{this.plugin.settings.titleFieldName=X||"title",await this.plugin.saveSettings()})),K=new F.Setting(w).setName("Title Prompt").setDesc(`Instructions for title generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.titlePrompt).onChange(async(X)=>{if(X.length>k){new F.Notice(`Title prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.titlePrompt);return}this.plugin.settings.titlePrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")});V.setDisabled(!this.plugin.settings.enableTitle),K.setDisabled(!this.plugin.settings.enableTitle)}}var f4=new Set(r0);function p(w,N,{nonEmpty:$=!1,maxLength:V}={}){if(typeof w!=="string")return N;if($&&w.trim()==="")return N;if(V!==void 0&&w.length>V)return N;return w}function dN(w,N){return typeof w==="boolean"?w:N}function A4(w,N){return typeof w==="number"&&Number.isInteger(w)&&w>0?w:N}function q4(w){return w===Lw[0]||w===Lw[1]||w===Lw[2]}function H4(w){return w===U0[0]||w===U0[1]}function a1(w){if(!w||typeof w!=="object"||Array.isArray(w))return null;let $={...w};if($.anthropicModel==="claude-sonnet-4-5-20250929")$.anthropicModel="claude-sonnet-4-6";if($.anthropicModel==="claude-opus-4-5-20251101")$.anthropicModel="claude-opus-4-6";let V=p($.anthropicModel,H.anthropicModel),K=p($.truncateMethod,H.truncateMethod),W=p($.updateMethod,H.updateMethod);return{anthropicApiKey:p($.anthropicApiKey,H.anthropicApiKey),anthropicModel:f4.has(V)?V:H.anthropicModel,tagsFieldName:p($.tagsFieldName,H.tagsFieldName,{nonEmpty:!0}),descriptionFieldName:p($.descriptionFieldName,H.descriptionFieldName,{nonEmpty:!0}),titleFieldName:p($.titleFieldName,H.titleFieldName,{nonEmpty:!0}),enableTitle:dN($.enableTitle,H.enableTitle),debugLogging:dN($.debugLogging,H.debugLogging),truncateContent:dN($.truncateContent,H.truncateContent),contentTokenLimit:A4($.contentTokenLimit,H.contentTokenLimit),truncateMethod:q4(K)?K:H.truncateMethod,updateMethod:H4(W)?W:H.updateMethod,tagsPrompt:p($.tagsPrompt,H.tagsPrompt,{nonEmpty:!0,maxLength:k}),descriptionPrompt:p($.descriptionPrompt,H.descriptionPrompt,{nonEmpty:!0,maxLength:k}),titlePrompt:p($.titlePrompt,H.titlePrompt,{nonEmpty:!0,maxLength:k})}}class cN extends s0.Plugin{settings=H;runController=new AbortController;async onload(){this.runController=new AbortController,await this.loadSettings(),this.addCommand({id:"generate-metadata",name:"Generate metadata for current note",callback:async()=>{await m1(this.app,this.settings,{signal:this.runController.signal})}}),this.registerEvent(this.app.workspace.on("file-menu",(w,N)=>{if(!(N instanceof s0.TFolder))return;w.addItem(($)=>$.setTitle("Generate metadata (recursive)").setIcon("tags").onClick(async()=>{await p1(this.app,N,{...this.settings},{signal:this.runController.signal})}))})),this.addSettingTab(new bN(this.app,this))}onunload(){this.runController.abort("plugin_unloaded")}async loadSettings(){let w=a1(await this.loadData());this.settings={...w??H}}async saveSettings(){await this.saveData(this.settings)}}
+`),userMessage:V}}function Y2(w){return w.split(",").map((N)=>N.trim()).filter((N)=>N!=="")}function x2(w){let N=w.trim();if(N.startsWith('"')&&N.endsWith('"')||N.startsWith("'")&&N.endsWith("'"))return N.substring(1,N.length-1);return N}function t0(w){if(!w)return!0;if(typeof w==="string")return w.trim()==="";if(Array.isArray(w))return w.length===0||w.every((N)=>String(N).trim()==="");return!1}function M2(w){return w.updateMethod==="always_regenerate"?"update_all":"only_empty"}function Q2(w,N){if(w==="update_all")return"update";return t0(N)?"update":"keep"}function yN(w,N){if(N.updateMethod==="always_regenerate")return!0;return t0(w[N.tagsFieldName])||t0(w[N.descriptionFieldName])||N.enableTitle&&t0(w[N.titleFieldName])}function G2(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}async function jN(w,N,$,V={}){if(N.extension!=="md")return{kind:"skipped",file:N,reason:"not a markdown file"};if(!$.anthropicApiKey)return{kind:"skipped",file:N,reason:"missing API key"};let W=w.metadataCache.getFileCache(N)?.frontmatter||{};if(!yN(W,$))return{kind:"skipped",file:N,reason:"all fields already populated"};if(V.signal?.aborted)return{kind:"skipped",file:N,reason:"cancelled before request"};try{return await T2(w,N,$,W,M2($),V.presentation??"interactive",V.signal)?{kind:"changed",file:N}:{kind:"skipped",file:N,reason:"no changes"}}catch(X){if(V.signal?.aborted||G2(X))return{kind:"skipped",file:N,reason:"cancelled"};return{kind:"error",file:N,reason:X instanceof Error?X.message:String(X),error:X}}}async function m$(w,N,$={}){let V=w.workspace.getActiveFile();if(!V){new v.Notice("Please open a file first");return}if(V.extension!=="md"){new v.Notice("Current file is not a markdown file");return}if(!N.anthropicApiKey){new v.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let K=await jN(w,V,N,{signal:$.signal});if(K.kind==="changed")new v.Notice("Metadata updated successfully");else if(K.kind==="error")Z2(K.error),console.error("generateMetadata error:",K.error)}async function T2(w,N,$,V,K,W,X){let Z=W==="bulk",Y=$.truncateContent?await IN(w,N,$.contentTokenLimit,$.truncateMethod):await IN(w,N,-1,"head_only"),{system:M,userMessage:C}=J2(Y,$);if($.debugLogging&&!Z)console.log("[Metadator] System:",M),console.log("[Metadator] User message:",C);let F=Z?void 0:new v.Notice("Generating metadata...",0),T;try{T=await b$(M,C,$,{signal:X})}finally{F?.hide()}if($.debugLogging)if(Z)console.log(`[Metadator] [bulk] ${N.path} — fields received`);else console.log("[Metadator] Fields:",T);if(X?.aborted)return!1;let U=!1;async function h(S,g){try{if(g==="keep")return await n0(w,N,S.fieldName,S.value,"keep");if(S.updateMethod==="append")return await n0(w,N,S.fieldName,S.value,"append");return await n0(w,N,S.fieldName,S.value,"update")}catch(Kw){if(!Z)new v.Notice(`Failed to write ${S.fieldName}: ${Kw instanceof Error?Kw.message:String(Kw)}`);return console.error(`updateFrontMatter error (${S.fieldName}):`,Kw),!1}}let I=[];if(T.tags)I.push({fieldName:$.tagsFieldName,value:Y2(T.tags),updateMethod:"append"});if(T.description)I.push({fieldName:$.descriptionFieldName,value:T.description,updateMethod:"update"});if($.enableTitle&&T.title)I.push({fieldName:$.titleFieldName,value:x2(T.title),updateMethod:"update"});for(let S of I){if(X?.aborted)return U;let g=Q2(K,V[S.fieldName]);if(await h(S,g))U=!0}return U}var C2=[2000,8000,30000];function kN(w){let N=[];for(let $ of w.children)if("children"in $)N.push(...kN($));else{let V=$;if(V.extension==="md")N.push(V)}return N}function u$(w,N,$){let V=[],K=[];for(let W of N){let X=w.metadataCache.getFileCache(W)?.frontmatter??{};if(yN(X,$))V.push(W);else K.push(W)}return{willChange:V,willSkip:K}}function O2(w){return w instanceof P&&(w.kind==="rate_limit"||w.kind==="overloaded")}var U2=2;function f2(w,N,$=Math.random){if(N instanceof P&&N.retryAfterMs!==void 0&&Number.isFinite(N.retryAfterMs))return Math.min(N.retryAfterMs,w*U2);return Math.round(w*(0.5+$()))}var F2=100;async function S2(w,N){if(w<=0)return N?.()??!1;let $=Date.now()+w;while(Date.now()<$){if(N?.())return!0;let V=$-Date.now();await new Promise((K)=>setTimeout(K,Math.min(F2,V)))}return N?.()??!1}async function A2(w,N,$,V,K,W,X=Math.random){let Z=()=>(K?.()??!1)||(W?.aborted??!1);for(let Y=0;Y<=V.length;Y++){if(Z())return{kind:"skipped",file:N,reason:"cancelled before attempt"};let M=await jN(w,N,$,{presentation:"bulk",signal:W});if(M.kind!=="error"||!O2(M.error))return M;if(Y===V.length)return M;let C=f2(V[Y],M.error,X);if(await S2(C,Z))return{kind:"skipped",file:N,reason:"cancelled during retry backoff"}}return{kind:"skipped",file:N,reason:"retry loop exited unexpectedly"}}async function i$(w,N,$,{onProgress:V,shouldAbort:K,retryDelaysMs:W,signal:X,random:Z}={}){let Y=W??C2,M=[],C=0;for(let F=0;F<N.length;F++){if(K?.()||X?.aborted)break;let T=N[F];V?.({current:F+1,total:N.length,file:T,errors:C});let U=await A2(w,T,$,Y,K,X,Z);if(M.push(U),U.kind==="error")C++}return M}var p$=require("obsidian");class hN extends p$.Modal{aborted=!1;finishing=!1;statusEl;cancelBtn;onAbort;onOpen(){let{contentEl:w}=this;w.createEl("h2",{text:"Generating metadata"}),this.statusEl=w.createEl("p",{text:"Starting…"});let N=w.createDiv({cls:"modal-button-container"});this.cancelBtn=N.createEl("button",{text:"Cancel"}),this.cancelBtn.addEventListener("click",()=>{if(this.aborted=!0,this.onAbort?.(),this.cancelBtn)this.cancelBtn.disabled=!0,this.cancelBtn.textContent="Cancelling…"})}setAbortHandler(w){this.onAbort=w}setProgress(w){if(!this.statusEl)return;this.statusEl.textContent=`${w.current} / ${w.total} · ${w.file.path} · errors: ${w.errors}`}isAborted(){return this.aborted}finish(){this.finishing=!0,this.close()}onClose(){if(!this.finishing)this.aborted=!0,this.onAbort?.();this.contentEl.empty()}}var l$=require("obsidian");class RN extends l$.Modal{results;aborted;totalPlanned;constructor(w,N,$,V){super(w);this.results=N,this.aborted=$,this.totalPlanned=V}onOpen(){let{contentEl:w}=this,N=this.results.filter((Y)=>Y.kind==="changed").length,$=this.results.filter((Y)=>Y.kind==="skipped").length,V=this.results.filter((Y)=>Y.kind==="error"),K=this.totalPlanned-this.results.length;w.createEl("h2",{text:this.aborted?"Cancelled":"Finished"});let W=w.createEl("ul");if(W.createEl("li",{text:`${N} changed`}),W.createEl("li",{text:`${$} skipped`}),W.createEl("li",{text:`${V.length} errored`}),K>0)W.createEl("li",{text:`${K} not processed (cancelled)`});if(V.length>0){w.createEl("h3",{text:"Errors"});let Y=w.createEl("ul");for(let M of V)if(M.kind==="error")Y.createEl("li",{text:`${M.file.path}: ${M.reason}`})}w.createDiv({cls:"modal-button-container"}).createEl("button",{text:"Close",cls:"mod-cta"}).addEventListener("click",()=>this.close())}onClose(){this.contentEl.empty()}}async function n$(w,N,$,V={}){if(!$.anthropicApiKey){new a0.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let K=kN(N);if(K.length===0){new a0.Notice("No markdown files found in folder");return}let{willChange:W,willSkip:X}=u$(w,K,$);if(W.length===0){new a0.Notice(`All ${K.length} notes already have metadata; nothing to do`);return}if(!await new e0(w,{folderPath:N.path,total:K.length,willChange:W.length,willSkip:X.length,settings:$}).openAndAwait())return;let Y=new hN(w),M=new AbortController;if(V.signal)if(V.signal.aborted)M.abort(V.signal.reason);else V.signal.addEventListener("abort",()=>M.abort(V.signal?.reason),{once:!0});Y.setAbortHandler(()=>M.abort("cancelled_by_user")),Y.open();let C=await i$(w,W,$,{onProgress:(T)=>Y.setProgress(T),shouldAbort:()=>(V.shouldAbort?.()??!1)||Y.isAborted(),signal:M.signal}),F=Y.isAborted()||(V.shouldAbort?.()??!1)||M.signal.aborted;Y.finish(),new RN(w,C,F,W.length).open()}var k=1000,o0=["claude-sonnet-4-6","claude-opus-4-6","claude-haiku-4-5-20251001"],t$={"claude-sonnet-4-6":"Claude Sonnet 4.6","claude-opus-4-6":"Claude Opus 4.6","claude-haiku-4-5-20251001":"Claude Haiku 4.5"},Lw=["head_only","head_tail","heading"],a$={head_only:"Beginning Only",head_tail:"Beginning + End",heading:"Headings + Summaries"},U0=["always_regenerate","preserve_existing"],o$={always_regenerate:"Always Regenerate",preserve_existing:"Preserve Existing"},E={anthropicApiKey:"",anthropicModel:"claude-sonnet-4-6",tagsFieldName:"tags",descriptionFieldName:"description",titleFieldName:"title",enableTitle:!0,debugLogging:!1,truncateContent:!0,contentTokenLimit:1000,truncateMethod:"head_only",updateMethod:"preserve_existing",tagsPrompt:"Select 3-5 relevant tags in lowercase with hyphens instead of spaces (e.g., 'knowledge-management', 'note-taking')",descriptionPrompt:"Write a concise but useful summary in 1-2 sentences that captures the main purpose and key points",titlePrompt:"Create a simple, concise title with minimal adjectives that clearly states the topic"};var f=require("obsidian");class _N extends f.PluginSettingTab{plugin;constructor(w,N){super(w,N);this.plugin=N}display(){let{containerEl:w}=this;w.empty(),new f.Setting(w).setName("Anthropic API Settings").setHeading(),w.createEl("p",{text:"Note: When you run the metadata command, your note content is sent to the Anthropic API for processing. No data is stored by Anthropic beyond the API request.",cls:"setting-item-description"}),new f.Setting(w).setName("API Key").setDesc("Your Anthropic API key. Get one at console.anthropic.com (requires an account with billing enabled)").addText((W)=>{W.setPlaceholder("sk-ant-...").setValue(this.plugin.settings.anthropicApiKey).onChange(async(X)=>{this.plugin.settings.anthropicApiKey=X,await this.plugin.saveSettings()}),W.inputEl.type="password"}),new f.Setting(w).setName("Model").setDesc("Model to use for metadata generation").addDropdown((W)=>{for(let X of o0)W.addOption(X,t$[X]);W.setValue(this.plugin.settings.anthropicModel).onChange(async(X)=>{this.plugin.settings.anthropicModel=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Debug Logging").setDesc("Log prompts and responses to the developer console (View → Toggle Developer Tools)").addToggle((W)=>W.setValue(this.plugin.settings.debugLogging).onChange(async(X)=>{this.plugin.settings.debugLogging=X,await this.plugin.saveSettings()})),new f.Setting(w).setName("Update Settings").setHeading(),new f.Setting(w).setName("Update Method").setDesc("Always Regenerate: regenerate on every command; Preserve Existing: only generate empty fields").addDropdown((W)=>{for(let X of U0)W.addOption(X,o$[X]);W.setValue(this.plugin.settings.updateMethod).onChange(async(X)=>{this.plugin.settings.updateMethod=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Truncate Content").setDesc("Limit content sent to API to reduce costs").addToggle((W)=>W.setValue(this.plugin.settings.truncateContent).onChange(async(X)=>{this.plugin.settings.truncateContent=X,await this.plugin.saveSettings(),N.setDisabled(!X),$.setDisabled(!X)}));let N=new f.Setting(w).setName("Content Token Limit").setDesc("Maximum number of tokens of note content sent to the API").addText((W)=>W.setValue(this.plugin.settings.contentTokenLimit.toString()).onChange(async(X)=>{let Z=Number.parseInt(X,10);if(Number.isNaN(Z)||Z<1){new f.Notice("Content token limit must be a positive integer"),W.setValue(this.plugin.settings.contentTokenLimit.toString());return}this.plugin.settings.contentTokenLimit=Z,await this.plugin.saveSettings()})),$=new f.Setting(w).setName("Truncate Method").setDesc("How to truncate long content").addDropdown((W)=>{for(let X of Lw)W.addOption(X,a$[X]);W.setValue(this.plugin.settings.truncateMethod).onChange(async(X)=>{this.plugin.settings.truncateMethod=X,await this.plugin.saveSettings()})});N.setDisabled(!this.plugin.settings.truncateContent),$.setDisabled(!this.plugin.settings.truncateContent),new f.Setting(w).setName("Tags Settings").setHeading(),new f.Setting(w).setName("Tags Field Name").setDesc("Frontmatter field name for tags").addText((W)=>W.setValue(this.plugin.settings.tagsFieldName).onChange(async(X)=>{this.plugin.settings.tagsFieldName=X||"tags",await this.plugin.saveSettings()})),new f.Setting(w).setName("Tags Prompt").setDesc(`Instructions for tag generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.tagsPrompt).onChange(async(X)=>{if(X.length>k){new f.Notice(`Tags prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.tagsPrompt);return}this.plugin.settings.tagsPrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Description Settings").setHeading(),new f.Setting(w).setName("Description Field Name").setDesc("Frontmatter field name for description").addText((W)=>W.setValue(this.plugin.settings.descriptionFieldName).onChange(async(X)=>{this.plugin.settings.descriptionFieldName=X||"description",await this.plugin.saveSettings()})),new f.Setting(w).setName("Description Prompt").setDesc(`Instructions for description generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.descriptionPrompt).onChange(async(X)=>{if(X.length>k){new f.Notice(`Description prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.descriptionPrompt);return}this.plugin.settings.descriptionPrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Title Settings").setHeading(),new f.Setting(w).setName("Enable Title").setDesc("Generate title metadata").addToggle((W)=>W.setValue(this.plugin.settings.enableTitle).onChange(async(X)=>{this.plugin.settings.enableTitle=X,await this.plugin.saveSettings(),V.setDisabled(!X),K.setDisabled(!X)}));let V=new f.Setting(w).setName("Title Field Name").setDesc("Frontmatter field name for title").addText((W)=>W.setValue(this.plugin.settings.titleFieldName).onChange(async(X)=>{this.plugin.settings.titleFieldName=X||"title",await this.plugin.saveSettings()})),K=new f.Setting(w).setName("Title Prompt").setDesc(`Instructions for title generation (max ${k} chars)`).addTextArea((W)=>{W.setValue(this.plugin.settings.titlePrompt).onChange(async(X)=>{if(X.length>k){new f.Notice(`Title prompt cannot exceed ${k} characters`),W.setValue(this.plugin.settings.titlePrompt);return}this.plugin.settings.titlePrompt=X,await this.plugin.saveSettings()}),W.inputEl.setAttr("rows","3")});V.setDisabled(!this.plugin.settings.enableTitle),K.setDisabled(!this.plugin.settings.enableTitle)}}var q2=new Set(o0);function p(w,N,{nonEmpty:$=!1,maxLength:V}={}){if(typeof w!=="string")return N;if($&&w.trim()==="")return N;if(V!==void 0&&w.length>V)return N;return w}function bN(w,N){return typeof w==="boolean"?w:N}function E2(w,N){return typeof w==="number"&&Number.isInteger(w)&&w>0?w:N}function L2(w){return w===Lw[0]||w===Lw[1]||w===Lw[2]}function H2(w){return w===U0[0]||w===U0[1]}function r$(w){if(!w||typeof w!=="object"||Array.isArray(w))return null;let $={...w};if($.anthropicModel==="claude-sonnet-4-5-20250929")$.anthropicModel="claude-sonnet-4-6";if($.anthropicModel==="claude-opus-4-5-20251101")$.anthropicModel="claude-opus-4-6";let V=p($.anthropicModel,E.anthropicModel),K=p($.truncateMethod,E.truncateMethod),W=p($.updateMethod,E.updateMethod);return{anthropicApiKey:p($.anthropicApiKey,E.anthropicApiKey),anthropicModel:q2.has(V)?V:E.anthropicModel,tagsFieldName:p($.tagsFieldName,E.tagsFieldName,{nonEmpty:!0}),descriptionFieldName:p($.descriptionFieldName,E.descriptionFieldName,{nonEmpty:!0}),titleFieldName:p($.titleFieldName,E.titleFieldName,{nonEmpty:!0}),enableTitle:bN($.enableTitle,E.enableTitle),debugLogging:bN($.debugLogging,E.debugLogging),truncateContent:bN($.truncateContent,E.truncateContent),contentTokenLimit:E2($.contentTokenLimit,E.contentTokenLimit),truncateMethod:L2(K)?K:E.truncateMethod,updateMethod:H2(W)?W:E.updateMethod,tagsPrompt:p($.tagsPrompt,E.tagsPrompt,{nonEmpty:!0,maxLength:k}),descriptionPrompt:p($.descriptionPrompt,E.descriptionPrompt,{nonEmpty:!0,maxLength:k}),titlePrompt:p($.titlePrompt,E.titlePrompt,{nonEmpty:!0,maxLength:k})}}class dN extends r0.Plugin{settings=E;runController=new AbortController;async onload(){this.runController=new AbortController,await this.loadSettings(),this.addCommand({id:"generate-metadata",name:"Generate metadata for current note",callback:async()=>{await m$(this.app,this.settings,{signal:this.runController.signal})}}),this.registerEvent(this.app.workspace.on("file-menu",(w,N)=>{if(!(N instanceof r0.TFolder))return;w.addItem(($)=>$.setTitle("Generate metadata (recursive)").setIcon("tags").onClick(async()=>{await n$(this.app,N,{...this.settings},{signal:this.runController.signal})}))})),this.addSettingTab(new _N(this.app,this))}onunload(){this.runController.abort("plugin_unloaded")}async loadSettings(){let w=r$(await this.loadData());this.settings={...w??E}}async saveSettings(){await this.saveData(this.settings)}}
